% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.rba_api_check}
\alias{.rba_api_check}
\title{Test connection with a rest server}
\usage{
.rba_api_check(url, diagnostics = FALSE)
}
\arguments{
\item{url}{A URL to to resource being examined.}

\item{diagnostics}{logical: Generate diagnostics and detailed messages with
internal information.}
}
\value{
An informative message with the result of HEAD request's success or
  failure.
}
\description{
Internal helper function for rba_connection_test(). It will  make HTTP HEAD
  request to the given resource.
}
\seealso{
Other internal_inernet_connectivity: 
\code{\link{.rba_http_status}()},
\code{\link{.rba_net_handle}()}
}
\concept{internal_inernet_connectivity}
