% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.rba_args_cons_chk}
\alias{.rba_args_cons_chk}
\title{Check If A cons Element Follows A Constrain Type}
\usage{
.rba_args_cons_chk(cons_i, what)
}
\arguments{
\item{cons_i}{element i from .rba_args()'s cons argument.}

\item{what}{what constrain to check? it should be one of the possible cons
types defined in .rba_args()'s documentations.}
}
\value{
Logical. TRUE if element i is correct with regard to the constrain
  "what"; FALSE otherwise.
}
\description{
This function will take a single element from the .rba_args()'s
   cons argument and a single constrain type and checks if it is TRUE.
}
\seealso{
Other internal_arguments_check: 
\code{\link{.rba_args_cons_msg}()},
\code{\link{.rba_args_cons_wrp}()},
\code{\link{.rba_args_opts}()},
\code{\link{.rba_args}()}
}
\concept{internal_arguments_check}
