% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.rba_args_cons_msg}
\alias{.rba_args_cons_msg}
\title{Produce Error Message If an Element doesn't Follow a constrain}
\usage{
.rba_args_cons_msg(cons_i, what)
}
\arguments{
\item{cons_i}{element i from .rba_args()'s cons argument.}

\item{what}{what constrain produced the error? it should be one of the
possible cons types defined in .rba_args()'s documentations.}
}
\value{
A character string.
}
\description{
In case of Constrain Error (i.e. a FALSE returned by
  .rba_args_cons_chk()), this function will produce a related error
  message.
}
\seealso{
Other internal_arguments_check: 
\code{\link{.rba_args_cond}()},
\code{\link{.rba_args_cons_chk}()},
\code{\link{.rba_args_cons_wrp}()},
\code{\link{.rba_args_opts}()},
\code{\link{.rba_args}()}
}
\concept{internal_arguments_check}
