% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.rba_args_opts}
\alias{.rba_args_opts}
\title{Add rbioapi options to user's Arguments Check}
\usage{
.rba_args_opts(cons = NULL, cond = NULL, what)
}
\arguments{
\item{cond}{Constrains to be evaluated.}

\item{what}{what to build? cond or cons?}
}
\value{
NULL. If The arguments check failed, the code execution will be
halted or a warning will be issued.
}
\description{
This function is an internal component of .rba_args(). It will
  add user-defiended rbioapi options variables (provided by the "..."
  arguments in the exported function call) to .rba_args's cond and cons.
}
\details{
The aim of this function is to eliminate the need
  to write explicit options arguments checking when writing the exported
  functions. Without this, the developer was forced to repeatably include
  every rbioapi options arguments in argument checking segment of each
  exported function.
}
\seealso{
Other internal_arguments_check: 
\code{\link{.rba_args_cons_chk}()},
\code{\link{.rba_args_cons_msg}()},
\code{\link{.rba_args_cons_wrp}()},
\code{\link{.rba_args}()}
}
\concept{internal_arguments_check}
