% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{.rba_enrichr_enrich_internal}
\alias{.rba_enrichr_enrich_internal}
\title{Internal function for rba_enrichr_enrich}
\usage{
.rba_enrichr_enrich_internal(user_list_id, gene_set_library, save_name, ...)
}
\arguments{
\item{user_list_id}{An ID returned to you after uploading a gene
list using \code{\link{rba_enrichr_add_list}}}

\item{gene_set_library}{a valid gene-set library name which exists
in the results retrieved via \code{\link{rba_enrichr_libs}}.}

\item{save_name}{default raw file name}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A data frame with the enrichment results of the provided user_list_id
  against the gene_set_library
}
\description{
This is an internal helper function which will retrieve the enrichment
  results of one_user_list id against one library name
}
\details{
The function will be called within \code{\link{rba_enrichr_enrich}} and will
handle API requests to the server.
}
\section{Corresponding API Resources}{

 "GET https://maayanlab.cloud/Enrichr/enrich"
}

\references{
\itemize{
  \item Chen EY, Tan CM, Kou Y, Duan Q, Wang Z, Meirelles GV, Clark NR,
  Ma'ayan A. Enrichr: interactive and collaborative HTML5 gene list
  enrichment analysis tool. BMC Bioinformatics. 2013;128(14).
  \item Kuleshov MV, Jones MR, Rouillard AD, Fernandez NF, Duan Q, Wang Z,
  Koplev S, Jenkins SL, Jagodnik KM, Lachmann A, McDermott MG, Monteiro CD,
  Gundersen GW, Ma'ayan A. Enrichr: a comprehensive gene set enrichment
  analysis web server 2016 update. Nucleic Acids Research. 2016; gkw377.
  \item \href{https://maayanlab.cloud/Enrichr/help#api}{Enrichr API
  Documentation}
  }
}
