% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.rba_ext_args}
\alias{.rba_ext_args}
\title{Temporary Change rbioapi Options During a Function Call}
\usage{
.rba_ext_args(...)
}
\arguments{
\item{...}{Extra arguments that were provided in the endpoints functions.}
}
\value{
NULL, if arguments check failed, code execution will be stopped.
  otherwise, nothing will be returned nor displayed.
}
\description{
The '...' argument of any exported function will be passed to this function.
  It will temporary alter the standard rbioapi options during the caller
  function execution.
}
\details{
The availabe rbioapi options will be retrieved from
  getOption("rba_user_options"). If the name of parameter in '...' is a
  standrad rbioapi option, the content of that option will be checked and
  in case that the content is valid, the caller function's environment will
  be altered in response to the change.\cr
  Also the function will ignore any arguments which is not standard and
  issues an informative warning for the user.
}
\concept{internal_options}
