% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{.rba_mieaa_species}
\alias{.rba_mieaa_species}
\title{Handle Species argument input for miEAA endpoints}
\usage{
.rba_mieaa_species(sp, to_name = FALSE)
}
\arguments{
\item{sp}{specie input.}

\item{to_name}{(logical) (default = FALSE) to convert a provided species
abbrevation to specie's scientific name.}
}
\value{
If to_name = FALSE, a three-lettered character string of
  a supported miEAA species. otherwise, if to_name = TRUE, a character
  string with the scientific name of that specie.
}
\description{
This internal function will make it possible for the users to provide
  variety of specie's name type or if they have enterned a wrongly-formatted
  or not supported species, to produce an informative error message
}
\examples{
.rba_mieaa_species(9606)
.rba_mieaa_species("hsa", to_name = TRUE)

}
\seealso{
Other "miEAA API": 
\code{\link{rba_mieaa_cats}()},
\code{\link{rba_mieaa_convert_type}()},
\code{\link{rba_mieaa_convert_version}()},
\code{\link{rba_mieaa_enrich_results}()},
\code{\link{rba_mieaa_enrich_status}()},
\code{\link{rba_mieaa_enrich_submit}()},
\code{\link{rba_mieaa_enrich}()}
}
\concept{"miEAA API"}
