% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.rba_pages_do}
\alias{.rba_pages_do}
\title{Iterate over function calls}
\usage{
.rba_pages_do(input_call, pb_switch)
}
\arguments{
\item{input_call}{a list that contains the calls.}

\item{pb_switch}{Display a progress bar?}
}
\value{
The evaluation results of each input call.
}
\description{
This function accepts a list where each of its elements is a character
  vector which can be parserd and evaluated. currently, this is only used
  in rba_pages
}
