% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{.rba_uniprot_search_namer}
\alias{.rba_uniprot_search_namer}
\title{Name UniProt search hits elements}
\usage{
.rba_uniprot_search_namer(x)
}
\arguments{
\item{x}{object to be parserd}
}
\value{
a list with the same structure of the input, only named.
}
\description{
Every search hit in uniprot has a character element whitin it, named
  "accession". this function should be used as the second response
  parser in the *_search functions to set the name each search hit to
  it's accession
}
