% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{paste_2}
\alias{paste_2}
\title{Grammatically Correct Pasting}
\usage{
paste_2(..., last = " and ", sep = ", ", quote = NA, quote_all = NA)
}
\arguments{
\item{...}{words to be appended together.}

\item{last}{(default: "AND") The separator between the last two words.}

\item{sep}{The separator between every words except the last two.}

\item{quote}{Should every word be quoted between a character?}

\item{quote_all}{Should the final result be quoted between a character?}
}
\value{
A character string of appended words, in a natural english way.
}
\description{
This function will append every element by comma and the last element by
  'and'/'or', just like natural and correct english sentence.
}
\seealso{
Other internal_misc: 
\code{\link{rba_ba_file}()},
\code{\link{v_msg}()}
}
\concept{internal_misc}
