% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_api_call}
\alias{rba_ba_api_call}
\title{Internal function to make http request}
\usage{
rba_ba_api_call(
  input_call,
  skip_error = FALSE,
  no_interet_retry_max = 1,
  no_internet_wait_time = 10,
  verbose = TRUE,
  diagnostics = FALSE
)
}
\arguments{
\item{input_call}{A httr function call made  by rba_ba_httr().}

\item{skip_error}{logical: If TRUE, in case of an error HTTP status other
than 200, instead of halting the code execution, the error message will be
returned as the function's output.}

\item{no_interet_retry_max}{numeric: A value to be passed to
rba_ba_net_handle() retry_max argument.}

\item{no_internet_wait_time}{numeric: A value to be passed to
rba_ba_net_handle() wait_time argument.}

\item{verbose}{should the function generate informative messages?}

\item{diagnostics}{logical: Generate diagnostics and detailed messages with
internal information.}
}
\value{
A raw server response in the format of httr's class "response". in
  the case of status code other than 200 and skip_error = TRUE, a character
  string with the pertinent error message.
}
\description{
This function will be called by rba_ba_skeleton() and is the internal
  function which resides between making an httr function call using
  rba_ba_httr and evaluating that call to retrieve a response from the API
  server.
}
\details{
In case of an error (anything other than status code 200), the function will
  perform extra steps according to the context:\cr
  1- If it was not possible to establish a connection with the server,
  rba_ba_net_handle() will be called to handle the situation.\cr
  2- If the server returned a status code 5xx, calling the server will be
  retried accordingly.\cr
  3- if the server returned status code other than 200 or 5xx, the response
  and status code will be handed to rba_error_parser() to handle the
  situation.
}
\seealso{
Other internal_api_calls: 
\code{\link{rba_ba_httr}()},
\code{\link{rba_ba_query}()},
\code{\link{rba_ba_skeleton}()}
}
\concept{internal_api_calls}
