% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_api_check}
\alias{rba_ba_api_check}
\title{Test connection with a rest server}
\usage{
rba_ba_api_check(url, diagnostics = FALSE)
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Test connection with a rest server
}
