% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_args}
\alias{rba_ba_args}
\title{Check provided Arguments}
\usage{
rba_ba_args(
  cons = NULL,
  cond = NULL,
  cond_warning = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Check provided Arguments
}
