% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_args_cond}
\alias{rba_ba_args_cond}
\title{Produce Error Message If an Element Doesn't Follow a Constrain}
\usage{
rba_ba_args_cond(cond_i)
}
\arguments{
\item{cond_i}{element i from \code{\link{rba_ba_args}}'s cond argument.}
}
\value{
A list containing the messages and warn element to
  determine the behaviour of \code{\link{rba_ba_args}}.
}
\description{
In case of Condition Error (i.e. a TRUE returned by evaluating the
 defined conditions in cond), this function will produce  a list with:
 1- messages that could be used as error or warning, 2- an element named
 "warn" that if FALSE, \code{\link{rba_ba_args}} will stop the code
 execution with message as error, or if TRUE, issues a warning with that
 message.
}
\concept{A list containing the messages and warn element to
  determine the behaviour of \code{\link{rba_ba_args}}.}
