% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_args_cons_wrp}
\alias{rba_ba_args_cons_wrp}
\title{A wrapper to Iterate Constrain Types on a cons' Element}
\usage{
rba_ba_args_cons_wrp(cons_i)
}
\arguments{
\item{cons_i}{element i from rba_ba_args()'s cons argument.}
}
\value{
A character vector with containing the error message for failed
  constrains, NA otherwise.
}
\description{
Iterates rba_ba_args_cons_chk() on every defined constrain
  for element i of a cons element. and produce an error message if necessary.
}
\seealso{
Other internal_arguments_check: 
\code{\link{rba_ba_args_cons_chk}()},
\code{\link{rba_ba_args_cons_msg}()},
\code{\link{rba_ba_args_opts}()},
\code{\link{rba_ba_args}()}
}
\concept{internal_arguments_check}
