% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_arguments_check}
\alias{rba_ba_arguments_check}
\title{Check provided Arguments}
\usage{
rba_ba_arguments_check(cons = NULL, cond = NULL, diagnostics = FALSE)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Check provided Arguments
}
