% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_body_add_pars}
\alias{rba_ba_body_add_pars}
\title{Add additional parameters to API call's body}
\usage{
rba_ba_body_add_pars(call_body, additional_pars)
}
\arguments{
\item{call_body}{}
}
\value{

}
\description{
Add additional parameters to API call's body
}
