% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_call_add_pars}
\alias{rba_ba_call_add_pars}
\title{add addition parameters to function call}
\usage{
rba_ba_call_add_pars(call_func_input, diagnostics, progress_bar)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
add addition parameters to function call
}
