% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_error_parser}
\alias{rba_ba_error_parser}
\title{Try to Pars an appropriate error response}
\usage{
rba_ba_error_parser(response, verbose = verbose)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Try to Pars an appropriate error response
}
