% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_ext_args}
\alias{rba_ba_ext_args}
\title{Temporary change a rbioapi option during a function call}
\usage{
rba_ba_ext_args(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Temporary change a rbioapi option during a function call
}
