% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_file}
\alias{rba_ba_file}
\title{check provided file address or create one}
\usage{
rba_ba_file(file, save_to = NA, dir_name = NA)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
check provided file address or create one
}
