% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_file_path}
\alias{rba_ba_file_path}
\title{Create/check provided file address}
\usage{
rba_ba_file_path(
  file_ext,
  file_name = NA,
  dir_name = NA,
  save_to = NA,
  verbose = TRUE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Create/check provided file address
}
