% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_httr}
\alias{rba_ba_httr}
\title{Build httr HTTP queries}
\usage{
rba_ba_httr(httr, url = NULL, path = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Build httr HTTP queries
}
