% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_internet_handler}
\alias{rba_ba_internet_handler}
\title{Handle situations when internet connection is disturbed}
\usage{
rba_ba_internet_handler(
  retry_max = 1,
  wait_time = 1,
  verbose = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Handle situations when internet connection is disturbed
}
