% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_multi_batch}
\alias{rba_ba_multi_batch}
\title{A wrapper for General skeleton for long inputs}
\usage{
rba_ba_multi_batch()
}
\value{

}
\description{
A wrapper for General skeleton for long inputs
}
