% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_net_handle}
\alias{rba_ba_net_handle}
\title{Handle situations when internet connection is disturbed or http statis 5xx}
\usage{
rba_ba_net_handle(
  retry_max = 1,
  wait_time = 10,
  verbose = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
changed name from rba_ba_internet_handler
}
