% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_query}
\alias{rba_ba_query}
\title{Add additional parameters to API call's body}
\usage{
rba_ba_query(init, ...)
}
\arguments{
\item{call_body}{}
}
\value{

}
\description{
changed the name from rba_ba_body_add_pars
the format shoud be 1- init which is a named list, followed by optional
lists with the following order: argument name, a condition which may be TRUE,
a value that should be appended to the list
}
