% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_response_parser}
\alias{rba_ba_response_parser}
\title{Parse API Response}
\usage{
rba_ba_response_parser(type = NA, parser = NULL)
}
\arguments{
\item{parser}{}
}
\value{

}
\description{
Parse API Response
}
