% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_skeletion}
\alias{rba_ba_skeletion}
\title{General skeleton for all functions in the package}
\usage{
rba_ba_skeletion(
  call_function,
  response_parser,
  no_interet_retry_max = 1,
  no_internet_wait_time = 10,
  verbose = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
General skeleton for all functions in the package
}
