% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_skeleton}
\alias{rba_ba_skeleton}
\title{General skeleton for all functions in the package}
\usage{
rba_ba_skeleton(input_call, response_parser = NULL)
}
\arguments{
\item{input_call}{}
}
\value{

}
\description{
General skeleton for all functions in the package
}
