% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_translate}
\alias{rba_ba_translate}
\title{Translate HTTP status code to human readable explanation}
\usage{
rba_ba_translate(http_status)
}
\arguments{
\item{http_status}{}
}
\value{

}
\description{
Translate HTTP status code to human readable explanation
}
