% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_ba_uniprot_call}
\alias{rba_ba_uniprot_call}
\title{Internal function to alternate uniprot calls between saving files/R objects}
\usage{
rba_ba_uniprot_call(
  call_func_input,
  save_switch,
  file_accept,
  file_parser,
  file_ext,
  file_name,
  obj_accept,
  obj_parser,
  verbose = TRUE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Internal function to alternate uniprot calls between saving files/R objects
}
