% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_alignment_region}
\alias{rba_ensembl_alignment_region}
\title{Retrieves genomic alignments as separate blocks based on a region and species}
\usage{
rba_ensembl_alignment_region(
  region,
  species,
  aligned = TRUE,
  compact = TRUE,
  compara = "vertebrates",
  display_species_set = NA,
  mask = NA,
  method = "EPO",
  species_set = NA,
  species_set_group = "mammals",
  ...
)
}
\arguments{
\item{species_set_group}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Retrieves genomic alignments as separate blocks based on a region and species
}
