% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_Archive}
\alias{rba_ensembl_Archive}
\title{Retrieve the latest version for a set of identifiers}
\usage{
rba_ensembl_Archive(
  id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieve the latest version for a set of identifiers
}
