% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_cafe_genetree}
\alias{rba_ensembl_cafe_genetree}
\title{Retrieves a cafe tree of the gene tree using the gene tree stable identifier}
\usage{
rba_ensembl_cafe_genetree(genetree_id, compara = "vertebrates", ...)
}
\arguments{
\item{compara}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Retrieves a cafe tree of the gene tree using the gene tree stable identifier
}
