% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_cafe_genetree_id}
\alias{rba_ensembl_cafe_genetree_id}
\title{Retrieves a cafe tree of the gene tree using the gene tree stable identifier}
\usage{
rba_ensembl_cafe_genetree_id(genetree_id, compara = "vertebrates", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Retrieves a cafe tree of the gene tree using the gene tree stable identifier
}
