% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_cafe_genetree_member_id}
\alias{rba_ensembl_cafe_genetree_member_id}
\title{Retrieves the cafe tree of the gene tree that contains the gene /
transcript / translation stable identifier}
\usage{
rba_ensembl_cafe_genetree_member_id(
  ensembl_id,
  compara = "vertebrates",
  db_type = NA,
  object_type = NA,
  species = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves the cafe tree of the gene tree that contains the gene /
transcript / translation stable identifier
}
