% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_cafe_genetree_member_symbol}
\alias{rba_ensembl_cafe_genetree_member_symbol}
\title{Retrieves the cafe tree of the gene tree that contains the gene
identified by a symbol}
\usage{
rba_ensembl_cafe_genetree_member_symbol(
  gene_symbol,
  species,
  compara = "vertebrates",
  db_type = "core",
  external_db = NA,
  object_type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves the cafe tree of the gene tree that contains the gene
identified by a symbol
}
