% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_cafe_genetree_search}
\alias{rba_ensembl_cafe_genetree_search}
\title{Retrieves the cafe tree of the gene tree that contains the gene /
transcript / translation stable identifier or identified by symbol}
\usage{
rba_ensembl_cafe_genetree_search(
  ensembl_id = NA,
  gene_symbol = NA,
  species = NA,
  external_db = NA,
  compara = "vertebrates",
  db_type = NA,
  object_type = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Retrieves the cafe tree of the gene tree that contains the gene /
transcript / translation stable identifier or identified by symbol
}
