% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_eqtl_gene}
\alias{rba_ensembl_eqtl_gene}
\title{Returns the p-value for each SNP in a given gene}
\usage{
rba_ensembl_eqtl_gene(
  gene_id,
  species,
  statistic = NA,
  tissue = NA,
  variant_name = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Returns the p-value for each SNP in a given gene
}
