% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_eqtl_variant}
\alias{rba_ensembl_eqtl_variant}
\title{Returns the p-values for a SNP (e.g. rs123)}
\usage{
rba_ensembl_eqtl_variant(
  variant_name,
  species,
  stable_id = NA,
  statistic = NA,
  tissue = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Returns the p-values for a SNP (e.g. rs123)
}
