% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_familiy_member_symbol}
\alias{rba_ensembl_familiy_member_symbol}
\title{Retrieves the information for all the families that contains the gene
identified by a symbol}
\usage{
rba_ensembl_familiy_member_symbol(
  gene_symbol,
  species,
  aligned = TRUE,
  compara = "vertebrates",
  db_type = "core",
  external_db = NA,
  member_source = "all",
  object_type = NA,
  sequence = "protein",
  ...
)
}
\arguments{
\item{sequence}{}
}
\value{

}
\description{
Retrieves the information for all the families that contains the gene
identified by a symbol
}
