% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_familiy_search}
\alias{rba_ensembl_familiy_search}
\title{Retrieves the information for all the families that contains the gene
identified by a symbol}
\usage{
rba_ensembl_familiy_search(
  ensembl_id = NA,
  gene_symbol = NA,
  species = NA,
  aligned = TRUE,
  compara = "vertebrates",
  db_type = NA,
  external_db = NA,
  member_source = "all",
  object_type = NA,
  sequence = "protein",
  ...
)
}
\arguments{
\item{ensembl_id}{}
}
\value{

}
\description{
Retrieves the information for all the families that contains the gene
identified by a symbol
}
