% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_genetree_id}
\alias{rba_ensembl_genetree_id}
\title{Retrieves a gene tree for a gene tree stable identifier}
\usage{
rba_ensembl_genetree_id(
  genetree_id,
  aligned = FALSE,
  cigar_line = FALSE,
  cluterset_id = NA,
  compara = "vertebrates",
  prune_species = NA,
  prune_taxon = NA,
  sequence = "protein",
  ...
)
}
\arguments{
\item{sequence}{}
}
\value{

}
\description{
Retrieves a gene tree for a gene tree stable identifier
}
