% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_genetree_member_symbol}
\alias{rba_ensembl_genetree_member_symbol}
\title{Retrieves the gene tree that contains the gene identified by a symbol}
\usage{
rba_ensembl_genetree_member_symbol(
  gene_symbol,
  species,
  aligned = FALSE,
  cigar_line = FALSE,
  cluterset_id = NA,
  compara = "vertebrates",
  db_type = "core",
  external_db = NA,
  object_type = NA,
  prune_species = NA,
  prune_taxon = NA,
  sequence = "protein",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves the gene tree that contains the gene identified by a symbol
}
