% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_genetree_search}
\alias{rba_ensembl_genetree_search}
\title{Retrieves the gene tree that contains the gene identified by a symbol}
\usage{
rba_ensembl_genetree_search(
  ensembl_id = NA,
  gene_symbol = NA,
  species = NA,
  aligned = FALSE,
  cigar_line = FALSE,
  cluterset_id = NA,
  compara = "vertebrates",
  db_type = "core",
  external_db = NA,
  object_type = NA,
  nh_format = NA,
  prune_species = NA,
  prune_taxon = NA,
  sequence = "protein",
  ...
)
}
\arguments{
\item{nh_format}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Retrieves the gene tree that contains the gene identified by a symbol
}
