% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_homology}
\alias{rba_ensembl_homology}
\title{Retrieves homology information (orthologs) by symbol or ensembl id}
\usage{
rba_ensembl_homology(
  ensembl_id = NA,
  gene_symbol = NA,
  species = NA,
  aligned = TRUE,
  cigar_line = TRUE,
  compara = "vertebrates",
  external_db = NA,
  format = "full",
  sequence = "protein",
  target_species = NA,
  target_taxon = NA,
  type = "all",
  ...
)
}
\arguments{
\item{ensembl_id}{}
}
\value{

}
\description{
Retrieves homology information (orthologs) by symbol or ensembl id
}
