% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_homology_id}
\alias{rba_ensembl_homology_id}
\title{Retrieves homology information (orthologs) by Ensembl gene id}
\usage{
rba_ensembl_homology_id(
  ensemble_id,
  aligned = TRUE,
  cigar_line = TRUE,
  compara = "vertebrates",
  format = "full",
  sequence = "protein",
  target_species = NA,
  target_taxon = NA,
  type = "all",
  ...
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Retrieves homology information (orthologs) by Ensembl gene id
}
