% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_assembly_region_name}
\alias{rba_ensembl_info_assembly_region_name}
\title{Returns information about the specified toplevel sequence region for the
given species.}
\usage{
rba_ensembl_info_assembly_region_name(
  species,
  region_name,
  bands = FALSE,
  synonyms = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  store_in_options = TRUE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Returns information about the specified toplevel sequence region for the
given species.
}
