% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_infomation.R
\name{rba_ensembl_info_biotypes}
\alias{rba_ensembl_info_biotypes}
\title{List the functional classifications of gene models that Ensembl associates
 with a particular species. Useful for restricting the type of
 genes/transcripts retrieved by other endpoints.}
\usage{
rba_ensembl_info_biotypes(
  species = NA,
  name = NA,
  group = NA,
  object_type = NA,
  ...
)
}
\arguments{
\item{object_type}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
List the functional classifications of gene models that Ensembl associates
 with a particular species. Useful for restricting the type of
 genes/transcripts retrieved by other endpoints.
}
