% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_biotypes_names}
\alias{rba_ensembl_info_biotypes_names}
\title{List the properties of biotypes with a given name. Object type
(gene or transcript) can be provided for filtering.}
\usage{
rba_ensembl_info_biotypes_names(
  name,
  object_type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)

rba_ensembl_info_biotypes_names(
  name,
  object_type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{



}
\description{
List the properties of biotypes with a given name. Object type
(gene or transcript) can be provided for filtering.

List the properties of biotypes with a given name. Object type
(gene or transcript) can be provided for filtering.
}
