% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_compara_methods}
\alias{rba_ensembl_info_compara_methods}
\title{List all compara analyses available (an analysis defines the type of
comparative data).}
\usage{
rba_ensembl_info_compara_methods(class = NA, compara = "vertebrates", ...)
}
\arguments{
\item{compara}{}
}
\value{

}
\description{
List all compara analyses available (an analysis defines the type of
comparative data).
}
