% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_compara_species}
\alias{rba_ensembl_info_compara_species}
\title{List all collections of species analysed with the specified compara method.}
\usage{
rba_ensembl_info_compara_species(method, compara = "vertebrates", ...)
}
\arguments{
\item{compara}{}
}
\value{

}
\description{
List all collections of species analysed with the specified compara method.
}
