% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_data}
\alias{rba_ensembl_info_data}
\title{Shows the data releases available on this REST server. May return more than
one release (unfrequent non-standard Ensembl configuration).}
\usage{
rba_ensembl_info_data(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Shows the data releases available on this REST server. May return more than
one release (unfrequent non-standard Ensembl configuration).
}
