% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_infomation.R
\name{rba_ensembl_info_divisions}
\alias{rba_ensembl_info_divisions}
\title{Get list of all Ensembl divisions for which information is available}
\usage{
rba_ensembl_info_divisions(...)
}
\arguments{
\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get list of all Ensembl divisions for which information is available
}
