% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes}
\alias{rba_ensembl_info_genomes}
\title{Find information about a given genome}
\usage{
rba_ensembl_info_genomes(
  genome_name = NA,
  INSDC_sequence_accession = NA,
  INSDC_assembly_id = NA,
  division_name = NA,
  taxon_name = NA,
  expand = FALSE,
  ...
)
}
\arguments{
\item{taxon_name}{}
}
\value{

}
\description{
Find information about a given genome
}
