% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes_accession}
\alias{rba_ensembl_info_genomes_accession}
\title{Find information about genomes containing a specified INSDC accession}
\usage{
rba_ensembl_info_genomes_accession(accession, expand = FALSE, ...)
}
\arguments{
\item{expand}{}
}
\value{

}
\description{
Find information about genomes containing a specified INSDC accession
}
