% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes_assembly}
\alias{rba_ensembl_info_genomes_assembly}
\title{Find information about a genome with a specified assembly}
\usage{
rba_ensembl_info_genomes_assembly(assembly_id, expand = FALSE, ...)

rba_ensembl_info_genomes_assembly(assembly_id, expand = FALSE, ...)
}
\arguments{
\item{expand}{}
}
\value{



}
\description{
Find information about a genome with a specified assembly

Find information about a genome with a specified assembly
}
