% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes_division}
\alias{rba_ensembl_info_genomes_division}
\title{Find information about all genomes in a given division.}
\usage{
rba_ensembl_info_genomes_division(
  division_name,
  expand = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Find information about all genomes in a given division.
}
