% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes_taxonomy}
\alias{rba_ensembl_info_genomes_taxonomy}
\title{Find information about all genomes beneath a given node of the taxonomy}
\usage{
rba_ensembl_info_genomes_taxonomy(taxon_name, expand = FALSE, ...)

rba_ensembl_info_genomes_taxonomy(taxon_name, expand = FALSE, ...)
}
\arguments{
\item{expand}{}
}
\value{



}
\description{
Find information about all genomes beneath a given node of the taxonomy

Find information about all genomes beneath a given node of the taxonomy
}
