% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_rest}
\alias{rba_ensembl_info_rest}
\title{Shows the current version of the Ensembl REST API.}
\usage{
rba_ensembl_info_rest(
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Shows the current version of the Ensembl REST API.
}
