% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_software}
\alias{rba_ensembl_info_software}
\title{Shows the current version of the Ensembl API used by the REST server.}
\usage{
rba_ensembl_info_software(
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Shows the current version of the Ensembl API used by the REST server.
}
