% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_infomation.R
\name{rba_ensembl_info_species}
\alias{rba_ensembl_info_species}
\title{Lists all available species, their aliases, available adaptor groups and data release.}
\usage{
rba_ensembl_info_species(
  division = "EnsemblVertebrates",
  hide_strain_info = FALSE,
  strain_collection = NA,
  expand = FALSE,
  ...
)
}
\arguments{
\item{expand}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Lists all available species, their aliases, available adaptor groups and data release.
}
