% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_variation_populations_species}
\alias{rba_ensembl_info_variation_populations_species}
\title{List all individuals for a population from a species}
\usage{
rba_ensembl_info_variation_populations_species(species, population_name, ...)
}
\arguments{
\item{population_name}{}
}
\value{

}
\description{
List all individuals for a population from a species
}
