% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_link_diseq.R
\name{rba_ensembl_ld_pairwise}
\alias{rba_ensembl_ld_pairwise}
\title{Computes and returns LD values between the given variants}
\usage{
rba_ensembl_ld_pairwise(
  variant_id_1,
  variant_id_2,
  species,
  population_name = NA,
  d_prime = NA,
  r2 = NA,
  ...
)
}
\arguments{
\item{r2}{}
}
\value{

}
\description{
Computes and returns LD values between the given variants
}
