% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_ld_region}
\alias{rba_ensembl_ld_region}
\title{Computes and returns LD values between all pairs of variants in the defined
region}
\usage{
rba_ensembl_ld_region(
  region,
  population_name,
  species,
  d_prime = NA,
  r2 = NA,
  ...
)
}
\arguments{
\item{r2}{}
}
\value{

}
\description{
Computes and returns LD values between all pairs of variants in the defined
region
}
