% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_link_diseq.R
\name{rba_ensembl_ld_variants}
\alias{rba_ensembl_ld_variants}
\title{Computes and returns LD values between the given variant and all other
variants in a window centered around the given variant. The window size
is set to 500 kb.}
\usage{
rba_ensembl_ld_variants(
  variant_id,
  species,
  population_name,
  attribs = FALSE,
  d_prime = NA,
  r2 = NA,
  window_size = 500,
  ...
)
}
\arguments{
\item{window_size}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Computes and returns LD values between the given variant and all other
variants in a window centered around the given variant. The window size
is set to 500 kb.
}
