% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_map_assembly}
\alias{rba_ensembl_map_assembly}
\title{Convert the co-ordinates of one assembly to another}
\usage{
rba_ensembl_map_assembly(
  asm_one,
  asm_two,
  region,
  species,
  coord_system = NA,
  target_coord_system = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Convert the co-ordinates of one assembly to another
}
