% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_mapping.R
\name{rba_ensembl_map_cds}
\alias{rba_ensembl_map_cds}
\title{Convert from CDS coordinates to genomic coordinates. Output reflects
forward orientation coordinates as returned from the Ensembl API.}
\usage{
rba_ensembl_map_cds(
  ensembl_id,
  region,
  species = NA,
  include_original_region = FALSE,
  ...
)
}
\arguments{
\item{include_original_region}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Convert from CDS coordinates to genomic coordinates. Output reflects
forward orientation coordinates as returned from the Ensembl API.
}
