% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_map_translation}
\alias{rba_ensembl_map_translation}
\title{Convert from protein (translation) coordinates to genomic coordinates.
Output reflects forward orientation coordinates as returned from the Ensembl API}
\usage{
rba_ensembl_map_translation(
  ensembl_id,
  region,
  species = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Convert from protein (translation) coordinates to genomic coordinates.
Output reflects forward orientation coordinates as returned from the Ensembl API
}
