% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_member_id}
\alias{rba_ensembl_member_id}
\title{Retrieves the information for all the families that contains the gene /
transcript / translation stable identifier}
\usage{
rba_ensembl_member_id(
  ensembl_id,
  aligned = TRUE,
  compara = "vertebrates",
  member_source = "all",
  sequence = "protein",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves the information for all the families that contains the gene /
transcript / translation stable identifier
}
