% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_microarray_binding_matrix}
\alias{rba_ensembl_microarray_binding_matrix}
\title{Return the specified binding matrix}
\usage{
rba_ensembl_microarray_binding_matrix(
  binding_matrix_id,
  species,
  unit = "frequencies",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Return the specified binding matrix
}
