% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_regulation.R
\name{rba_ensembl_microarray_epigenome}
\alias{rba_ensembl_microarray_epigenome}
\title{Returns information about all epigenomes available for the given species}
\usage{
rba_ensembl_microarray_epigenome(species, ...)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Returns information about all epigenomes available for the given species
}
