% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_regulation.R
\name{rba_ensembl_microarray_probe}
\alias{rba_ensembl_microarray_probe}
\title{Returns information about a specific probe from a microarray}
\usage{
rba_ensembl_microarray_probe(microarray, probe, species, ...)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Returns information about a specific probe from a microarray
}
