% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_microarray_probe_set}
\alias{rba_ensembl_microarray_probe_set}
\title{Returns information about a specific probe_set from a microarray}
\usage{
rba_ensembl_microarray_probe_set(microarray, probe_set, species, ...)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Returns information about a specific probe_set from a microarray
}
