% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_regulation.R
\name{rba_ensembl_microarray_regulatory_feature}
\alias{rba_ensembl_microarray_regulatory_feature}
\title{Returns a RegulatoryFeature given its stable ID}
\usage{
rba_ensembl_microarray_regulatory_feature(
  regulatory_feature_id,
  species,
  activity = FALSE,
  ...
)
}
\arguments{
\item{activity}{}
}
\value{

}
\description{
Returns a RegulatoryFeature given its stable ID
}
