% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_microarray_vendor}
\alias{rba_ensembl_microarray_vendor}
\title{Returns information about a specific microarray}
\usage{
rba_ensembl_microarray_vendor(
  microarray,
  vendor,
  species,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Returns information about a specific microarray
}
