% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_ontologies.R
\name{rba_ensembl_ontology_ancestors}
\alias{rba_ensembl_ontology_ancestors}
\title{Reconstruct the entire ancestry of a term from is_a and part_of relationships}
\usage{
rba_ensembl_ontology_ancestors(term_id, chart = FALSE, ontology = NA, ...)
}
\arguments{
\item{ontology}{}
}
\value{

}
\description{
Reconstruct the entire ancestry of a term from is_a and part_of relationships
}
