% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_ontologies.R
\name{rba_ensembl_ontology_descendants}
\alias{rba_ensembl_ontology_descendants}
\title{Find all the terms descended from a given term. By default searches are
conducted within the namespace of the given identifier}
\usage{
rba_ensembl_ontology_descendants(
  term_id,
  closest_term = FALSE,
  ontology = NA,
  subset = NA,
  zero_distance = NA,
  ...
)
}
\arguments{
\item{zero_distance}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Find all the terms descended from a given term. By default searches are
conducted within the namespace of the given identifier
}
