% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_ontologies.R
\name{rba_ensembl_ontology_id}
\alias{rba_ensembl_ontology_id}
\title{Search for an ontological term by its namespaced identifier}
\usage{
rba_ensembl_ontology_id(term_id, relation = NA, simple = FALSE, ...)
}
\arguments{
\item{simple}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search for an ontological term by its namespaced identifier
}
