% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_ontologies.R
\name{rba_ensembl_ontology_name}
\alias{rba_ensembl_ontology_name}
\title{Search for a list of ontological terms by their name}
\usage{
rba_ensembl_ontology_name(
  name,
  ontology = NA,
  relation = NA,
  simple = FALSE,
  ...
)
}
\arguments{
\item{simple}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search for a list of ontological terms by their name
}
