% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_overlap.R
\name{rba_ensembl_overlap_id}
\alias{rba_ensembl_overlap_id}
\title{Retrieves features (e.g. genes, transcripts, variants and more) that overlap
a region defined by the given identifier.}
\usage{
rba_ensembl_overlap_id(
  ensembl_id,
  feature,
  biotype = NA,
  db_type = NA,
  logic_name = NA,
  misc_set = NA,
  object_type = NA,
  so_term = NA,
  species = NA,
  species_set = "mammals",
  variant_set = NA,
  ...
)
}
\arguments{
\item{variant_set}{}
}
\value{

}
\description{
Retrieves features (e.g. genes, transcripts, variants and more) that overlap
a region defined by the given identifier.
}
