% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_overlap_region}
\alias{rba_ensembl_overlap_region}
\title{Retrieves features (e.g. genes, transcripts, variants and more) that
overlap a given region}
\usage{
rba_ensembl_overlap_region(
  region,
  feature,
  species,
  biotype = NA,
  db_type = NA,
  logic_name = NA,
  misc_set = NA,
  so_term = NA,
  species_set = "mammals",
  variant_set = NA,
  trim_downstream = FALSE,
  trim_upstream = FALSE,
  ...
)
}
\arguments{
\item{trim_upstream}{}
}
\value{

}
\description{
Retrieves features (e.g. genes, transcripts, variants and more) that
overlap a given region
}
