% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_overlap.R
\name{rba_ensembl_overlap_translation}
\alias{rba_ensembl_overlap_translation}
\title{Retrieve features related to a specific Translation as described by
its stable ID (e.g. domains, variants)}
\usage{
rba_ensembl_overlap_translation(
  ensembl_id,
  db_type = NA,
  feature = "protein_feature",
  so_term = NA,
  species = NA,
  type = NA,
  ...
)
}
\arguments{
\item{type}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Retrieve features related to a specific Translation as described by
its stable ID (e.g. domains, variants)
}
