% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_phenotype_region}
\alias{rba_ensembl_phenotype_region}
\title{Return phenotype annotations that overlap a given genomic region}
\usage{
rba_ensembl_phenotype_region(
  region,
  species,
  feature_type = NA,
  include_pubmed_id = FALSE,
  include_review_status = FALSE,
  include_submitter = FALSE,
  only_phenotypes = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Return phenotype annotations that overlap a given genomic region
}
