% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_pheno_annot.R
\name{rba_ensembl_phenotype_term}
\alias{rba_ensembl_phenotype_term}
\title{Return phenotype annotations for genomic features given a phenotype
ontology term}
\usage{
rba_ensembl_phenotype_term(
  term,
  species,
  include_children = FALSE,
  include_pubmed_id = FALSE,
  include_review_status = FALSE,
  source = "undef",
  ...
)
}
\arguments{
\item{source}{}
}
\value{

}
\description{
Return phenotype annotations for genomic features given a phenotype
ontology term
}
