% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_sequence_id}
\alias{rba_ensembl_sequence_id}
\title{Request multiple types of sequence by a stable identifier list}
\usage{
rba_ensembl_sequence_id(
  ids,
  db_type = NA,
  start = NA,
  end = NA,
  expand_3prime = NA,
  expand_5prime = NA,
  format = NA,
  mask = NA,
  mask_feature = FALSE,
  object_type = NA,
  type = "genomic",
  species = NA,
  ...
)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Request multiple types of sequence by a stable identifier list
}
