% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_taxonomy_classification}
\alias{rba_ensembl_taxonomy_classification}
\title{Return the taxonomic classification of a taxon node}
\usage{
rba_ensembl_taxonomy_classification(
  taxon_id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Return the taxonomic classification of a taxon node
}
