% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_taxonomy_id}
\alias{rba_ensembl_taxonomy_id}
\title{Search for a taxonomic term by its identifier or name}
\usage{
rba_ensembl_taxonomy_id(
  taxon_id,
  simple = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Search for a taxonomic term by its identifier or name
}
