% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_taxonomy_name}
\alias{rba_ensembl_taxonomy_name}
\title{Search for a taxonomic id by a non-scientific name}
\usage{
rba_ensembl_taxonomy_name(
  name,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Search for a taxonomic id by a non-scientific name
}
