% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_taxonomy.R
\name{rba_ensembl_taxonomy_search}
\alias{rba_ensembl_taxonomy_search}
\title{Search for a taxonomic term by its identifier, name or non-scientific name}
\usage{
rba_ensembl_taxonomy_search(
  taxon = NA,
  non_scientific_name = NA,
  simple = FALSE,
  ...
)
}
\arguments{
\item{non_scientific_name}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search for a taxonomic term by its identifier, name or non-scientific name
}
