% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_variant_recoder}
\alias{rba_ensembl_variant_recoder}
\title{Translate a list of variant identifiers, HGVS notations or genomic SPDI
notations to all possible variant IDs, HGVS and genomic SPDI}
\usage{
rba_ensembl_variant_recoder(ids, species, fields = NA, ...)
}
\arguments{
\item{fields}{}
}
\value{

}
\description{
Translate a list of variant identifiers, HGVS notations or genomic SPDI
notations to all possible variant IDs, HGVS and genomic SPDI
}
