% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_variation_id}
\alias{rba_ensembl_variation_id}
\title{Uses a list of variant identifiers (e.g. rsID) to return the variation
features including optional genotype, phenotype and population data}
\usage{
rba_ensembl_variation_id(
  ids,
  species,
  genotypes = FALSE,
  phenotypes = FALSE,
  pops = FALSE,
  population_genotypes = FALSE,
  ...
)
}
\arguments{
\item{population_genotypes}{}
}
\value{

}
\description{
Uses a list of variant identifiers (e.g. rsID) to return the variation
features including optional genotype, phenotype and population data
}
