% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_vep_ids}
\alias{rba_ensembl_vep_ids}
\title{Fetch variant consequences for multiple ids}
\usage{
rba_ensembl_vep_ids(
  ids,
  species,
  Blosum62 = FALSE,
  CADD = FALSE,
  GeneSplicer = FALSE,
  LoF = FALSE,
  MaxEntScan = FALSE,
  Phenotypes = FALSE,
  SpliceAI = FALSE,
  SpliceRegion = FALSE,
  appris = FALSE,
  canonical = FALSE,
  ccds = FALSE,
  dbNSFP = NA,
  dbscSNV = FALSE,
  distance = 5000,
  domains = FALSE,
  failed = FALSE,
  hgvs = FALSE,
  mane = FALSE,
  merged = FALSE,
  miRNA = FALSE,
  minimal = FALSE,
  numbers = FALSE,
  protein = FALSE,
  refseq = FALSE,
  shift_3prime = FALSE,
  shift_genomic = FALSE,
  transcript_id = NA,
  transcript_version = FALSE,
  tsl = FALSE,
  uniprot = FALSE,
  variant_class = FALSE,
  vcf_string = FALSE,
  xref_refseq = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Fetch variant consequences for multiple ids
}
