% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_vep.R
\name{rba_ensembl_vep_variant}
\alias{rba_ensembl_vep_variant}
\title{Fetch variant consequences for multiple regions}
\usage{
rba_ensembl_vep_variant(
  variants,
  species,
  Blosum62 = FALSE,
  CADD = FALSE,
  GeneSplicer = FALSE,
  LoF = FALSE,
  MaxEntScan = FALSE,
  Phenotypes = FALSE,
  SpliceAI = FALSE,
  SpliceRegion = FALSE,
  appris = FALSE,
  canonical = FALSE,
  ccds = FALSE,
  dbNSFP = NA,
  dbscSNV = FALSE,
  distance = 5000,
  domains = FALSE,
  failed = FALSE,
  hgvs = FALSE,
  mane = FALSE,
  merged = FALSE,
  miRNA = FALSE,
  minimal = FALSE,
  numbers = FALSE,
  protein = FALSE,
  refseq = FALSE,
  shift_3prime = FALSE,
  shift_genomic = FALSE,
  transcript_id = NA,
  transcript_version = FALSE,
  tsl = FALSE,
  uniprot = FALSE,
  variant_class = FALSE,
  vcf_string = FALSE,
  xref_refseq = FALSE,
  ...
)
}
\arguments{
\item{xref_refseq}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Fetch variant consequences for multiple regions
}
