% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_xrefs_symbol}
\alias{rba_ensembl_xrefs_symbol}
\title{Looks up an external symbol and returns all Ensembl objects linked to it.
This can be a display name for a gene/transcript/translation, a synonym or
an externally linked reference. If a gene's transcript is linked to the
supplied symbol the service will return both gene and transcript
(it supports transient links).}
\usage{
rba_ensembl_xrefs_symbol(
  external_symbol,
  species,
  db_type = NA,
  external_db = NA,
  object_type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{external_symbol}{}
}
\value{

}
\description{
Looks up an external symbol and returns all Ensembl objects linked to it.
This can be a display name for a gene/transcript/translation, a synonym or
an externally linked reference. If a gene's transcript is linked to the
supplied symbol the service will return both gene and transcript
(it supports transient links).
}
