% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar.R
\name{rba_jaspar_matrix}
\alias{rba_jaspar_matrix}
\title{Get a position frequency matrix (PFM) with annotations}
\usage{
rba_jaspar_matrix(matrix_id, file_format = NULL, save_to = NULL, ...)
}
\arguments{
\item{matrix_id}{Character: A matrix profile
identifier. It follows a "base_id.version" naming scheme.}

\item{file_format}{Character: Instead of returning an R object, you
can directly download the profile matrix as a file in this format.
Supported formats are "yaml", "jaspar", "transfac", "meme", and "pfm".}

\item{save_to}{NULL or Character:\itemize{
\item NULL: (only if file_format was supplied) Save the file to an
  automatically-generated path.
\item Character string: A valid file or directory path to save the file to.}}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the PFM, details, and annotations. If
  \code{file_format} is supplied, the function returns the unparsed file
  content as a character string.
}
\description{
Using this function, you can retrieve the position frequency matrix (PFM)
  associated with a matrix profile identifier, along with its details and
  annotations. If a base ID (i.e., without a version suffix) is supplied,
 the latest version will be returned.
}
\section{Corresponding API Resources}{

 "GET https://jaspar.elixir.no/api/v1/matrix/\{matrix_id\}/"
}

\examples{
\donttest{
rba_jaspar_matrix("MA0600.2")
}
\dontrun{
rba_jaspar_matrix(matrix_id = "MA0600.2",
                  file_format = "meme",
                  save_to = "my_matrix.meme")
}

}
\references{
\itemize{
  \item Baydar Ovek D, et al. JASPAR 2026: expansion of transcription
  factor binding profiles and integration of deep learning models.
  Nucleic Acids Res. 2026;54(D1):D184-D193;
  doi: 10.1093/nar/gkaf1209
  \item Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data
  from any programming language. Bioinformatics, 2017,
  doi: 10.1093/bioinformatics/btx804
  \item
  \href{https://jaspar.elixir.no/api/v1/docs/}{JASPAR API Documentation}
  \item \href{https://jaspar.elixir.no/faq/}{Citations note
  on JASPAR website}
  }
}
\seealso{
Other "JASPAR":
\code{\link[=rba_jaspar_collections]{rba_jaspar_collections()}},
\code{\link[=rba_jaspar_collections_matrices]{rba_jaspar_collections_matrices()}},
\code{\link[=rba_jaspar_matrix_search]{rba_jaspar_matrix_search()}},
\code{\link[=rba_jaspar_matrix_versions]{rba_jaspar_matrix_versions()}},
\code{\link[=rba_jaspar_releases]{rba_jaspar_releases()}},
\code{\link[=rba_jaspar_sites]{rba_jaspar_sites()}},
\code{\link[=rba_jaspar_species]{rba_jaspar_species()}},
\code{\link[=rba_jaspar_species_matrices]{rba_jaspar_species_matrices()}},
\code{\link[=rba_jaspar_taxons]{rba_jaspar_taxons()}},
\code{\link[=rba_jaspar_taxons_matrices]{rba_jaspar_taxons_matrices()}},
\code{\link[=rba_jaspar_tffm]{rba_jaspar_tffm()}},
\code{\link[=rba_jaspar_tffm_search]{rba_jaspar_tffm_search()}}
}
\concept{"JASPAR"}
