% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{rba_mieaa_convert_type}
\alias{rba_mieaa_convert_type}
\title{Convert Between Mature and Precursor miRNA Identifiers}
\usage{
rba_mieaa_convert_type(
  mirna,
  input_type,
  only_unique = FALSE,
  simple_output = FALSE,
  ...
)
}
\arguments{
\item{mirna}{Character vector: miRNA identifiers to convert.}

\item{input_type}{Character: Type of the supplied miRNA identifiers; either
"mature" or "precursor".}

\item{only_unique}{Logical: (default = \code{FALSE}) Mature and precursor
miRNA identifiers do not always map uniquely. If \code{TRUE}, do not
return mappings for inputs with multiple matches. In tabular output, these
inputs remain as rows with \code{"-"} in the output column.}

\item{simple_output}{Logical: (default = \code{FALSE}) If \code{FALSE},
return a two-column data frame containing the input and output identifier
mappings; multiple output identifiers are separated by semicolons. If
\code{TRUE}, expand one-to-many mappings into a flat character vector of
converted identifiers without their association with the supplied
identifiers.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Depending on \code{simple_output}, a data frame or character vector
  containing the mappings returned by miEAA. Unrecognized or unmapped
  supplied identifiers can be omitted from the output.
}
\description{
miRBase identifiers can refer to either mature or precursor miRNAs.
  (see: \href{https://rnajournal.cshlp.org/content/9/3/277}{A uniform system
  for microRNA annotation}). Use this function to convert mature miRNA
  identifiers to precursor identifiers or vice versa.
}
\section{Corresponding API Resources}{

 "POST https://ccb-compute2.cs.uni-saarland.de/mieaa/api/v1/mirna_precursor_converter/"
}

\examples{
\donttest{
Sys.sleep(1) # to prevent 429 error during R CMD check
rba_mieaa_convert_type(mirna = c("hsa-miR-20b-5p", "hsa-miR-144-5p"),
    input_type = "mature")
}

}
\references{
\itemize{
  \item Ernesto Aparicio-Puerta, Pascal Hirsch, Georges P. Schmartz,
  Fabian Kern, Tobias Fehlmann, Andreas Keller, miEAA 2023: updates,
  new functional microRNA sets and improved enrichment visualizations,
  Nucleic Acids Research, Volume 51, Issue W1, 5 July 2023,
  Pages W319–W325, https://doi.org/10.1093/nar/gkad392
  \item
  \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/tutorial/api/}{miEAA
  browsable API tutorial}
  \item \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/}{Citation note
  on miEAA website}
  }
}
\seealso{
Other "miEAA":
\code{\link[=rba_mieaa_cats]{rba_mieaa_cats()}},
\code{\link[=rba_mieaa_convert_version]{rba_mieaa_convert_version()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_mieaa_enrich_results]{rba_mieaa_enrich_results()}},
\code{\link[=rba_mieaa_enrich_status]{rba_mieaa_enrich_status()}},
\code{\link[=rba_mieaa_enrich_submit]{rba_mieaa_enrich_submit()}}
}
\concept{"miEAA"}
