% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{rba_mieaa_enrich_submit}
\alias{rba_mieaa_enrich_submit}
\title{Submit miEAA miRNA Enrichment Analysis Request}
\usage{
rba_mieaa_enrich_submit(
  test_set,
  mirna_type,
  test_type,
  species,
  categories = NULL,
  p_adj_method = "fdr",
  independent_p_adj = TRUE,
  sig_level = 0.05,
  min_hits = 2,
  ref_set = NULL,
  ...
)
}
\arguments{
\item{test_set}{Character: Mature or precursor miRBase miRNA
identifiers. Note that \enumerate{
\item Only miRBase v22 identifiers are accepted. You can use
\code{\link{rba_mieaa_convert_version}} to convert older identifiers to
miRBase v22.
\item The list must contain either mature or precursor miRNA identifiers,
not a mixture of both.
}}

\item{mirna_type}{Character: Type of the supplied miRNA identifiers; either
"mature" or "precursor".}

\item{test_type}{Character: Analysis to perform; either "ORA" for
over-representation analysis or "GSEA" for miRNA gene set enrichment
analysis. For GSEA, the input list must already be ranked by an appropriate
criterion.}

\item{species}{Character or Numeric: Scientific name, abbreviation, or NCBI
 taxon ID of one of the following species:
 \enumerate{
\item "Homo sapiens", "hsa" or 9606
\item "Mus musculus", "mmu" or 10090
\item "Rattus norvegicus", "rno" or 10116
\item "Arabidopsis thaliana", "ath" or 3702
\item "Bos taurus", "bta" or 9913
\item "Caenorhabditis elegans", "cel" or 6239
\item "Drosophila melanogaster", "dme" or 7227
\item "Danio rerio", "dre" or 7955
\item "Gallus gallus", "gga" or 9031
\item "Sus scrofa", "ssc" or  9823
}}

\item{categories}{Character: (optional) One or more
category identifiers to use for miRNA set enrichment analysis. Note that
\itemize{
\item Available categories vary with the selected species and whether
 the supplied miRNAs are mature or precursor. Use
 \code{\link{rba_mieaa_cats}} to retrieve a list of available category
 identifiers for a given species and miRNA type.
\item If \code{NULL}, the analysis is performed using all available
 categories.}}

\item{p_adj_method}{Character: (default = \code{"fdr"}) P-value adjustment
method to use. One of: "none", "fdr", "bonferroni", "BY", "hochberg",
"holm", or "hommel".}

\item{independent_p_adj}{Logical: (default = \code{TRUE}) The scope of
p-value adjustment. If \code{TRUE}, p-values are adjusted separately
within each category. If \code{FALSE}, p-values are adjusted collectively
over all categories.}

\item{sig_level}{Numeric: (default = \code{0.05}) Significance threshold for
adjusted p-values. Values equal to or greater than this threshold are
omitted from the results. Must be greater than 0 and at most 1.}

\item{min_hits}{Numeric: (default = \code{2}) Minimum number of miRNAs from
the test set that a subcategory must contain to be included in the
results. Must be a positive integer.}

\item{ref_set}{Character: (optional) Only applicable
when \code{test_type = "ORA"}. Used as the reference (background or
universe) set for p-value calculations.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list that contains your submitted job's ID and a URL to
  manually check for your job status.
}
\description{
Submit a request to the miEAA server to perform over-representation analysis
  or gene set enrichment analysis for a set of miRNA identifiers.
}
\details{
Note that using \code{\link{rba_mieaa_enrich}} is a more convenient way to
  automatically perform this and other required function calls to
  perform enrichment analysis on your input miRNA-set using miEAA.
}
\section{Corresponding API Resources}{

 "POST https://ccb-compute2.cs.uni-saarland.de/mieaa/api/v1/enrichment_analysis/\{species\}/\{type\}/\{test\}/"
}

\examples{
\donttest{
Sys.sleep(1) # to prevent 429 error during R CMD check
rba_mieaa_enrich_submit(test_set = c("hsa-miR-20b-5p", "hsa-miR-144-5p"),
    mirna_type = "mature",
    test_type = "GSEA",
    species = 9606,
    categories = NULL)
}

}
\references{
\itemize{
  \item Ernesto Aparicio-Puerta, Pascal Hirsch, Georges P. Schmartz,
  Fabian Kern, Tobias Fehlmann, Andreas Keller, miEAA 2023: updates,
  new functional microRNA sets and improved enrichment visualizations,
  Nucleic Acids Research, Volume 51, Issue W1, 5 July 2023,
  Pages W319–W325, https://doi.org/10.1093/nar/gkad392
  \item
  \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/tutorial/api/}{miEAA
  browsable API tutorial}
  \item \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/}{Citation note
  on miEAA website}
  }
}
\seealso{
Other "miEAA":
\code{\link[=rba_mieaa_cats]{rba_mieaa_cats()}},
\code{\link[=rba_mieaa_convert_type]{rba_mieaa_convert_type()}},
\code{\link[=rba_mieaa_convert_version]{rba_mieaa_convert_version()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_mieaa_enrich_results]{rba_mieaa_enrich_results()}},
\code{\link[=rba_mieaa_enrich_status]{rba_mieaa_enrich_status()}}
}
\concept{"miEAA"}
