% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_homolog}
\alias{rba_panther_homolog}
\title{Search PANTHER for Homologs of Gene(s)}
\usage{
rba_panther_homolog(genes, organism, type = "P", target_organisms = NULL, ...)
}
\arguments{
\item{genes}{Character or Numeric: A vector of gene identifiers with maximum
length of 10. Can be any of: Ensembl gene ID, Ensembl protein ID, Ensembl
transcript ID, Entrez gene ID, gene symbol, NCBI GI, HGNC ID,
International protein index ID, NCBI UniGene ID, UniProt accession and/or
UniProt ID.}

\item{organism}{Numeric: NCBI taxon ID of the organism of your supplied
genes. run \code{\link{rba_panther_info}} with argument
'what = "organisms"' to get a list of PANTHER's supported organisms.}

\item{type}{Character: (default = \code{"P"}) Homolog types to return. either "P" (default) for paralogs,
"X" for horizontal gene transfer and "LDX" for least diverged horizontal
gene transfer.}

\item{target_organisms}{Numeric: (optional) NCBI taxon ID(s) to filter the results.
run \code{\link{rba_panther_info}} with argument 'what = "organisms"' to
get a list of PANTHER's supported organisms.
This argument is ignored for paralogs, which are searched within the input
organism. For horizontal gene transfers, target organisms should differ
from the input organism.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame with homolog information.
}
\description{
Using this function you can search and retrieve homolog of given gene(s).
}
\section{Corresponding API Resources}{

 "POST https://www.pantherdb.org/services/oai/pantherdb/ortholog/homologOther"
}

\examples{
\donttest{
rba_panther_homolog("OR4F5", organism = 9606, type = "P")
}

}
\references{
\itemize{
  \item Thomas PD, Ebert D, Muruganujan A, Mushayahama T, Albou L-P,
  Mi H. (2022) PANTHER: Making genome-scale phylogenetics accessible to all.
  Protein Science, 31(1), 8–22.
  https://doi.org/10.1002/pro.4218
  \item \href{https://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  \item
  \href{https://www.pantherdb.org/publications.jsp#HowToCitePANTHER}{Citations
  note on PANTHER website}
  }
}
\seealso{
Other "PANTHER":
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_panther_family]{rba_panther_family()}},
\code{\link[=rba_panther_genome]{rba_panther_genome()}},
\code{\link[=rba_panther_info]{rba_panther_info()}},
\code{\link[=rba_panther_mapping]{rba_panther_mapping()}},
\code{\link[=rba_panther_ortholog]{rba_panther_ortholog()}},
\code{\link[=rba_panther_tree_grafter]{rba_panther_tree_grafter()}}
}
\concept{"PANTHER"}
