% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_ortholog}
\alias{rba_panther_ortholog}
\title{Search PANTHER for Orthologs of Gene(s)}
\usage{
rba_panther_ortholog(
  genes,
  organism,
  type = "all",
  target_organisms = NULL,
  seq_pos = NULL,
  include_msa = NULL,
  ...
)
}
\arguments{
\item{genes}{Character or Numeric: A vector of gene identifiers with maximum
length of 10, or only one if \code{seq_pos} is supplied. Can be any of: Ensembl
gene ID, Ensembl protein ID, Ensembl transcript ID, Entrez gene ID, gene
symbol, NCBI GI, HGNC ID, International protein index ID, NCBI UniGene ID,
UniProt accession and/or UniProt ID.}

\item{organism}{Numeric: NCBI taxon ID of the organism of your supplied
genes. run \code{\link{rba_panther_info}} with argument
'what = "organisms"' to get a list of PANTHER's supported organisms.}

\item{type}{Character: (default = \code{"all"}) Ortholog types to return. either "all" (default) or "LDO" to
only return least diverged orthologs.}

\item{target_organisms}{Numeric: (optional) NCBI taxon ID(s) to filter the results.
run \code{\link{rba_panther_info}} with argument 'what = "organisms"' to
get a list of PANTHER's supported organisms.}

\item{seq_pos}{Numeric: (optional) A position in the protein's sequence of the
supplied gene. should be in the range of the protein's length.}

\item{include_msa}{Logical: (optional) Only if a sequence position is supplied,
should MSA (Multiple Sequence Alignment) information be included in the
results?}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame with Orthologs information.
}
\description{
Using this function you can search and retrieve orthologs of given gene(s),
  and optionally return the corresponding position in the target organisms'
  protein sequences.
}
\section{Corresponding API Resources}{

 "POST https://www.pantherdb.org/services/oai/pantherdb/ortholog/matchortho"
 \cr "POST https://www.pantherdb.org/services/oai/pantherdb/ortholog/homologpos"
}

\examples{
\donttest{
rba_panther_ortholog("CD40", organism = 9606, type = "LDO")
}

}
\references{
\itemize{
  \item Thomas PD, Ebert D, Muruganujan A, Mushayahama T, Albou L-P,
  Mi H. (2022) PANTHER: Making genome-scale phylogenetics accessible to all.
  Protein Science, 31(1), 8–22.
  https://doi.org/10.1002/pro.4218
  \item \href{https://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  \item
  \href{https://www.pantherdb.org/publications.jsp#HowToCitePANTHER}{Citations
  note on PANTHER website}
  }
}
\seealso{
Other "PANTHER":
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_panther_family]{rba_panther_family()}},
\code{\link[=rba_panther_genome]{rba_panther_genome()}},
\code{\link[=rba_panther_homolog]{rba_panther_homolog()}},
\code{\link[=rba_panther_info]{rba_panther_info()}},
\code{\link[=rba_panther_mapping]{rba_panther_mapping()}},
\code{\link[=rba_panther_tree_grafter]{rba_panther_tree_grafter()}}
}
\concept{"PANTHER"}
