% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_import}
\alias{rba_reactome_analysis_import}
\title{Import Saved Analysis JSON to Reactome}
\usage{
rba_reactome_analysis_import(input, input_format = NULL, ...)
}
\arguments{
\item{input}{Character: A local file path or HTTP or HTTPS URL that points
to plain or gzipped saved analysis results.}

\item{input_format}{Character: (optional) This function will automatically
identify your supplied input's format. To be explicit, set this argument to
one of:\itemize{
 \item "file": If you supplied a local file path pointing to the saved
 results file.
 \item "url": If you supplied an HTTP or HTTPS URL pointing to the saved
 results file.}
 An explicit value takes precedence. Otherwise, HTTP and HTTPS addresses
 are identified before existing local files. Other inputs are rejected.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the new token and other information of your
  imported results.
}
\description{
If you have a JSON file of analysis results (only obtained via
  \code{\link{rba_reactome_analysis_download}} with the result argument
  set to "results", or "results_gz"), you can import the results back to
  Reactome and retrieve a token.
  \cr This is useful when you want to use other Reactome services which
  require a token but you do not have a token or your token has been
  expired (i.e. more than 7 days passed from your analysis).
}
\section{Corresponding API Resources}{

"POST https://reactome.org/AnalysisService/import/form"
\cr "POST https://reactome.org/AnalysisService/import/url"
}

\examples{
\dontrun{
rba_reactome_analysis_import("c:/rbioapi/res.json")
}
\dontrun{
rba_reactome_analysis_import("https://qaz.com/res.json.gz")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/AnalysisService/}{Reactome Analysis
  Services API Documentation}
  \item \href{https://reactome.org/cite}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Analysis Service":
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_reactome_analysis_download]{rba_reactome_analysis_download()}},
\code{\link[=rba_reactome_analysis_mapping]{rba_reactome_analysis_mapping()}},
\code{\link[=rba_reactome_analysis_pdf]{rba_reactome_analysis_pdf()}},
\code{\link[=rba_reactome_analysis_species]{rba_reactome_analysis_species()}},
\code{\link[=rba_reactome_analysis_token]{rba_reactome_analysis_token()}}
}
\concept{"Reactome Analysis Service"}
