% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_token}
\alias{rba_reactome_analysis_token}
\title{Return the Results Associated with a Token}
\usage{
rba_reactome_analysis_token(
  token,
  species = NULL,
  sort_by = "ENTITIES_PVALUE",
  order = "ASC",
  resource = "TOTAL",
  p_value = 1,
  include_disease = TRUE,
  min = NULL,
  max = NULL,
  ...
)
}
\arguments{
\item{token}{Character: A token associated to your previous Reactome analysis.}

\item{species}{Character or Numeric: (optional) NCBI Taxonomy identifier
(Human is 9606), species name (e.g. "Homo sapiens") or Reactome DbId
(e.g. Homo sapiens is 48887). See
 \code{\link{rba_reactome_species}} or
 \href{https://reactome.org/content/schema/objects/Species/}{Reactome
 Data Schema: Entries: Species}.}

\item{sort_by}{Character: (default = \code{"ENTITIES_PVALUE"}) Sort the
result based on what column? Available choices are: "NAME",
"TOTAL_ENTITIES", "TOTAL_INTERACTORS", "TOTAL_REACTIONS", "FOUND_ENTITIES",
"FOUND_INTERACTORS", "FOUND_REACTIONS", "ENTITIES_RATIO",
"ENTITIES_PVALUE", "ENTITIES_FDR" or "REACTIONS_RATIO".}

\item{order}{Character: (default = \code{"ASC"}) Sort Order. Can be either
"ASC" or "DESC".}

\item{resource}{Character: (default = \code{"TOTAL"}) Filter results based on
the resource. Available choices are: "TOTAL", "UNIPROT", "ENSEMBL",
"CHEBI", "IUPHAR", "MIRBASE", "NCBI_PROTEIN", "EMBL", "COMPOUND" or
"PUBCHEM_COMPOUND".}

\item{p_value}{Numeric: (default = \code{1}) Set a P value threshold. Only
results with P value equal to or less than your supplied threshold will be
returned (1 means no P value filtering).}

\item{include_disease}{Logical: (default = \code{TRUE}) Should the disease
pathways be included in the results?}

\item{min}{Numeric: (optional) Minimum number of entities that a pathways
should have to be included in the results.}

\item{max}{Numeric: (optional) Maximum number of entities that a pathways
should have to be included in the results.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the results and information about the analysis.
  The \code{pathways} element is a data frame with information about each
  pathway expanded into columns; it is an empty data frame when no pathways
  match. Its structure is the same as the output from
  \code{\link{rba_reactome_analysis}}.
}
\description{
Use a token generated After a Reactome analysis
  (via \code{\link{rba_reactome_analysis}}) to Retrieve the analysis results.
  The output format is identical to the returned object of
  \code{\link{rba_reactome_analysis}}.
}
\details{
After any analysis, Reactome will associate a token with your analysis. It
  can later be used in functions that require the token (e.g. to retrieve
  the analysis results, download pdf).
  \cr Note that Reactome will store your token for only 7 days. You can
  download your full results with
  \code{\link{rba_reactome_analysis_download}}, and re-import it anytime to
  reactome (using \code{\link{rba_reactome_analysis_import}}) to generate
  a new token.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/AnalysisService/token/\{token\}"
}

\examples{
\dontrun{
rba_reactome_analysis_token(token = "MjAyMDEwMTYwMTI3MTNfMjY1MjM",
    species = 9606)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/AnalysisService/}{Reactome Analysis
  Services API Documentation}
  \item \href{https://reactome.org/cite}{Citations note on Reactome website}
  }
}
\seealso{
\code{\link{rba_reactome_analysis}}

Other "Reactome Analysis Service":
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_reactome_analysis_download]{rba_reactome_analysis_download()}},
\code{\link[=rba_reactome_analysis_import]{rba_reactome_analysis_import()}},
\code{\link[=rba_reactome_analysis_mapping]{rba_reactome_analysis_mapping()}},
\code{\link[=rba_reactome_analysis_pdf]{rba_reactome_analysis_pdf()}},
\code{\link[=rba_reactome_analysis_species]{rba_reactome_analysis_species()}}
}
\concept{"Reactome Analysis Service"}
