% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_entity_component_of}
\alias{rba_reactome_entity_component_of}
\title{Retrieves the list of structures (Complexes and Sets) that include the
given entity as their component. It should be mentioned that the list
includes only simplified entries (type, names, ids) and not full
information about each item.}
\usage{
rba_reactome_entity_component_of(
  entity_id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves the list of structures (Complexes and Sets) that include the
given entity as their component. It should be mentioned that the list
includes only simplified entries (type, names, ids) and not full
information about each item.
}
