% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_entity_other_forms}
\alias{rba_reactome_entity_other_forms}
\title{Get Other forms of a Reactome Entity}
\usage{
rba_reactome_entity_other_forms(entity_id, ...)
}
\arguments{
\item{entity_id}{Character: Reactome's entity ID.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Data frame where each row is other forms of your supplied Entity ID
  and columns are pertinent information.
}
\description{
This function retrieve a list containing all other forms of your supplied
  Physical Entity ID.
}
\details{
According to Reactome API documentation, "These other forms are Physical
  Entities that share the same Reference Entity identifier, e.g. PTEN
  H93R R-HSA-2318524 and PTEN C124R R-HSA-2317439 are two forms of PTEN."
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/entity/\{id\}/otherForms"
}

\examples{
\donttest{
rba_reactome_entity_other_forms("R-HSA-199420")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Physical Entity Queries":
\code{\link[=rba_reactome_complex_list]{rba_reactome_complex_list()}},
\code{\link[=rba_reactome_complex_subunits]{rba_reactome_complex_subunits()}},
\code{\link[=rba_reactome_participant_of]{rba_reactome_participant_of()}}
}
\concept{"Reactome Content Service - Physical Entity Queries"}
