% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_document}
\alias{rba_reactome_exporter_document}
\title{This method accepts identifiers for Event class instances.
The generated document contains the details for the given event and,
optionally, its children (see level parameter). These details include:
- A diagram image
- Summation
- Literature references
- Edit history
- Other details: type, location, compartments, diseases}
\usage{
rba_reactome_exporter_document(
  event_id,
  save_to = NA,
  level = 1,
  diagram_profile = "Modern",
  token = NA,
  resource = "TOTAL",
  analysis_profile = "Standard",
  exp_column = NA,
  ...
)
}
\arguments{
\item{exp_column}{}
}
\value{

}
\description{
This method accepts identifiers for Event class instances.
The generated document contains the details for the given event and,
optionally, its children (see level parameter). These details include:
- A diagram image
- Summation
- Literature references
- Edit history
- Other details: type, location, compartments, diseases
}
