% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_event}
\alias{rba_reactome_exporter_event}
\title{Exports A Reactome Event to SBGN or SBML}
\usage{
rba_reactome_exporter_event(event_id, output_format, save_to = NULL, ...)
}
\arguments{
\item{event_id}{Character: Reactome event's database IDs (DbId) or Stable IDs
(StId).}

\item{output_format}{Character: Either "sbgn" or "sbml".}

\item{save_to}{Character: (optional) \itemize{
\item NULL: Save the file to an automatically-generated path.
\item Character string: A valid file path to save the file to.}}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
NULL, According to the inputs, a SBGN or SBML file will be saved to
  disk.
}
\description{
This function will export a supplied Reactome Event (Pathway or Reaction)
  to a SBGN (Systems Biology Graphical Notation) or SBML (Systems Biology
  Markup Language)
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/exporter/event/
 \{identifier\}.sbgn"
 \cr "GET https://reactome.org/ContentService/exporter/event/
 \{identifier\}.sbml"
}

\examples{
\dontrun{
rba_reactome_exporter_event(event_id = "R-HSA-177929",
    output_format = "sbgn",
    save_to = "R-HSA-177929.sbgn")
}
\dontrun{
rba_reactome_exporter_event(event_id = "R-HSA-177929",
    output_format = "sbgn")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Format Exporter":
\code{\link[=rba_reactome_exporter_diagram]{rba_reactome_exporter_diagram()}},
\code{\link[=rba_reactome_exporter_overview]{rba_reactome_exporter_overview()}},
\code{\link[=rba_reactome_exporter_reaction]{rba_reactome_exporter_reaction()}}
}
\concept{"Reactome Content Service - Format Exporter"}
