% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_pathways_participants}
\alias{rba_reactome_pathways_participants}
\title{Events are the building blocks used in Reactome to represent all biological
processes, and they include pathways and reactions. Typically, an event can
contain other events. For example, a pathway can contain smaller pathways
(subpathways) and reactions. This method recursively retrieves a single
attribute for each of the events contained in the given event.}
\usage{
rba_reactome_pathways_participants(id, attribute_name = NA, ...)
}
\arguments{
\item{attribute_name}{}
}
\value{

}
\description{
Events are the building blocks used in Reactome to represent all biological
processes, and they include pathways and reactions. Typically, an event can
contain other events. For example, a pathway can contain smaller pathways
(subpathways) and reactions. This method recursively retrieves a single
attribute for each of the events contained in the given event.
}
