% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_people_id}
\alias{rba_reactome_people_id}
\title{Get a Person by Identifier}
\usage{
rba_reactome_people_id(
  person_id,
  authored_pathways = FALSE,
  publications = FALSE,
  attribute_name = NULL,
  ...
)
}
\arguments{
\item{person_id}{Character: Reactome database ID (DbId) or ORCID identifier.}

\item{authored_pathways}{Logical: (default = \code{FALSE}) Only return
Pathway list authored by the person?}

\item{publications}{Logical: (default = \code{FALSE}) Only return
publications list authored by the person?}

\item{attribute_name}{Character: (optional) A Reactome person attribute to
return only.
see \href{https://reactome.org/content/schema/Person/}{Reactome Data
Schema: person} for available options.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
List containing the requested informations of your supplied
  person.
}
\description{
Get a Person by Identifier
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/person/\{id\}"
 \cr "GET https://reactome.org/ContentService/data/person/\{id\}/
 authoredPathways"
 \cr "GET https://reactome.org/ContentService/data/person/\{id\}/publications"
 \cr "GET https://reactome.org/ContentService/data/person/\{id\}/
 \{attributeName\}"
}

\examples{
\donttest{
rba_reactome_people_id("391309")
}
\donttest{
rba_reactome_people_id(person_id = "391309", authored_pathways = TRUE)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Person Queries":
\code{\link[=rba_reactome_people_name]{rba_reactome_people_name()}}
}
\concept{"Reactome Content Service - Person Queries"}
