% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_functional_annotation}
\alias{rba_string_functional_annotation}
\title{Retrieving functional annotation}
\usage{
rba_string_functional_annotation(
  input,
  species = NA,
  allow_pubmed = NA,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
#' @family STRING
}
