% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_functional_terms}
\alias{rba_string_functional_terms}
\title{Search STRING Functional Terms}
\usage{
rba_string_functional_terms(term_text, species, ...)
}
\arguments{
\item{term_text}{Character: A functional term identifier or descriptive text
used to match one or more functional terms.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame in which every row is a matching functional term and
  the columns contain the term category, identifier, description, number of
  annotated proteins, preferred protein names, and STRING protein IDs.
  \code{preferredNames} and \code{stringIds} are returned as list-columns.
}
\description{
STRING maps several functional annotation resources onto its proteins. This
  function searches for functional terms using an identifier or descriptive
  text and retrieves the matching terms and their annotated proteins.
}
\details{
This endpoint supports only one species per query. If multiple functional
  terms match \code{term_text}, STRING returns them in order of relevance,
  with the best match first.
  \cr The complete number of annotated proteins is reported in
  \code{proteinCount}.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/functional_terms?term_text=
 \{your_term\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_functional_terms(
    term_text = "T cell receptor signaling pathway",
    species = 9606
)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{
  \link{rba_string_annotations},
  \link{rba_string_enrichment},
  \link{rba_string_enrichment_image}
  }

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}
}
\concept{"STRING"}
