% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_homology}
\alias{rba_string_homology}
\title{Retrieving similarity scores of the protein set}
\usage{
rba_string_homology(
  input,
  species = NA,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
#' @family STRING
}
