% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_homology_best}
\alias{rba_string_homology_best}
\title{Retrieving best similarity hits between species}
\usage{
rba_string_homology_best(
  input,
  species = NA,
  species_b = NA,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
#' @family STRING
}
