% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_map_ids}
\alias{rba_string_map_ids}
\title{Map a Set of Identifiers to STRING Identifiers}
\usage{
rba_string_map_ids(ids, species = NULL, echo_query = TRUE, limit = NULL, ...)
}
\arguments{
\item{ids}{Your common gene/protein identifier(s) to be mapped.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.
(Recommended, but optional.)}

\item{echo_query}{(default = \code{TRUE}) Include your input IDs as a column of the
results.}

\item{limit}{Deprecated: Retained temporarily for backward compatibility.
STRING v12 returns only the single best match per input ID, so this
argument has no effect.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame with at most one mapped STRING ID per input ID and
  other pertinent information. The \code{queryIndex} column contains the
  zero-based position of each resolved ID in the input vector. Unresolved
  inputs are omitted; if none can be resolved, a zero-row data frame
  retaining the response columns is returned.
}
\description{
This function calls STRING's API to map a set of common gene or protein
  identifiers to STRING identifiers. Although STRING services accept a
  variety of identifiers, the STRING API documentation recommends mapping
  them to STRING identifiers before using other STRING functions.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/tsv/get_string_ids?identifiers=
 \{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_map_ids(ids = c("TP53", "TNF", "EGFR"), species = 9606)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}
}
\concept{"STRING"}
