% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_ppi_enrichment}
\alias{rba_string_ppi_enrichment}
\title{Getting protein-protein interaction enrichment}
\usage{
rba_string_ppi_enrichment(
  input,
  species = NA,
  required_score = NA,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
#' @family STRING
}
