% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_antigen}
\alias{rba_uniprot_antigen}
\title{Get antigen by UniProt accession}
\usage{
rba_uniprot_antigen(
  accession,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get antigen by UniProt accession
}
