% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates_location}
\alias{rba_uniprot_coordinates_location}
\title{Search UniProt entries by taxonomy and genomic coordinates}
\usage{
rba_uniprot_coordinates_location(
  taxid,
  locations,
  in_range = TRUE,
  feature = FALSE,
  ...
)
}
\arguments{
\item{taxid}{Numeric: NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID}.}

\item{locations}{Character: Genomic location formatted as
 chromosome:start-end.
(e.g. "Y:17100001-19600000"). If you omit chromosome, it will be interpreted
as any chromosome (e.g. "1-10000").}

\item{in_range}{Logical: (default = \code{TRUE}) If \code{TRUE}, return only
proteins that are fully contained in the supplied range.}

\item{feature}{Logical: (default = \code{FALSE}) If \code{TRUE}, return
mapped protein features rather than protein coordinate records.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing UniProt proteins that match the supplied genomic
  location and taxonomy ID.
}
\description{
Retrieve UniProt entries or mapped protein features for a taxon and
  supplied \href{https://www.uniprot.org/help/genomic_coordinates}{genomic
  coordinates}.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/coordinates/\{taxonomy\}/\{locations\}/feature"
 \cr "GET https://www.ebi.ac.uk/proteins/api/coordinates/\{taxonomy\}/\{locations\}"
}

\examples{
\donttest{
rba_uniprot_coordinates_location(taxid = 9606,
    locations = "Y:17100001-19600000", in_range = TRUE)
}
\donttest{
rba_uniprot_coordinates_location(taxid = 9606,
    locations = "20:39000001", in_range = FALSE)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item McGarvey, P. B., Nightingale, A., Luo, J., Huang, H., Martin,
  M. J., Wu, C., & The UniProt Consortium. (2019). UniProt genomic
  mapping for deciphering functional effects of missense variants.
  Human Mutation, 40(6), 694–705.
  https://doi.org/10.1002/humu.23738
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Coordinates":
\code{\link[=rba_uniprot_coordinates]{rba_uniprot_coordinates()}},
\code{\link[=rba_uniprot_coordinates_location_genome]{rba_uniprot_coordinates_location_genome()}},
\code{\link[=rba_uniprot_coordinates_location_protein]{rba_uniprot_coordinates_location_protein()}},
\code{\link[=rba_uniprot_coordinates_search]{rba_uniprot_coordinates_search()}}
}
\concept{"UniProt - Coordinates"}
