% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates_location_protein}
\alias{rba_uniprot_coordinates_location_protein}
\title{Map Protein Sequence Positions to Genomic Coordinates}
\usage{
rba_uniprot_coordinates_location_protein(
  accession,
  p_position = NULL,
  p_start = NULL,
  p_end = NULL,
  ...
)
}
\arguments{
\item{accession}{Character:
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}.}

\item{p_position}{Numeric: (optional) Protein sequence position. Supply this
alone, or supply both \code{p_start} and \code{p_end}.}

\item{p_start}{Numeric: (optional) Protein sequence range start.}

\item{p_end}{Numeric: (optional) Protein sequence range end.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list with a \code{locations} element containing the mapped
  protein and genomic boundaries. Records can include chromosome, strand,
  genome assembly, nucleotide and Ensembl identifiers, amino acids, and
  mapped sequence features.
}
\description{
Map an amino-acid position or range in a UniProt protein sequence to its
  corresponding
  \href{https://www.uniprot.org/help/genomic_coordinates}{genomic
  coordinates}. A protein sequence location may have more than one genomic
  mapping. Supply \code{p_position} alone, or supply \code{p_start} and
  \code{p_end} together.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/coordinates/location/\{accession\}:\{pPosition\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/coordinates/location/\{accession\}:\{pStart\}-\{pEnd\}"
}

\examples{
\donttest{
rba_uniprot_coordinates_location_protein(accession = "P25942", p_position = 1)
}
\donttest{
rba_uniprot_coordinates_location_protein(accession = "P25942",
    p_start = 1, p_end = 277)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item McGarvey, P. B., Nightingale, A., Luo, J., Huang, H., Martin,
  M. J., Wu, C., & The UniProt Consortium. (2019). UniProt genomic
  mapping for deciphering functional effects of missense variants.
  Human Mutation, 40(6), 694–705.
  https://doi.org/10.1002/humu.23738
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Coordinates":
\code{\link[=rba_uniprot_coordinates]{rba_uniprot_coordinates()}},
\code{\link[=rba_uniprot_coordinates_location]{rba_uniprot_coordinates_location()}},
\code{\link[=rba_uniprot_coordinates_location_genome]{rba_uniprot_coordinates_location_genome()}},
\code{\link[=rba_uniprot_coordinates_search]{rba_uniprot_coordinates_search()}}
}
\concept{"UniProt - Coordinates"}
