% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates_taxonomy}
\alias{rba_uniprot_coordinates_taxonomy}
\title{Search UniProt entries by taxonomy and genomic coordinates}
\usage{
rba_uniprot_coordinates_taxonomy(
  taxid,
  locations,
  in_range = NA,
  feature = FALSE,
  ...
)
}
\arguments{
\item{feature}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search UniProt entries by taxonomy and genomic coordinates
}
