% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_genecentric_search}
\alias{rba_uniprot_genecentric_search}
\title{Search Gene-Centric Proteins}
\usage{
rba_uniprot_genecentric_search(upid = NULL, accession = NULL, gene = NULL, ...)
}
\arguments{
\item{upid}{Character: (optional)
\href{https://www.uniprot.org/help/proteome_id}{UniProt Proteome
identifier (UPID)}. You can supply up to 100 UPIDs.}

\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{gene}{Character: (optional) Unique gene identifier(s) found in MOD,
\href{https://www.ensembl.org/info/genome/genebuild/gene_names.html}{Ensembl},
Ensembl Genomes, \href{https://www.uniprot.org/help/gene_name}{OLN},
\href{https://www.uniprot.org/help/gene_name}{ORF} or
\href{https://www.uniprot.org/help/gene_name}{UniProt Gene Name}.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing matching gene-centric protein groups.
}
\description{
UniProt gene-centric protein groups organize related protein entries from a
  proteome by gene. Search these groups by proteome, accession, or gene
  identifier. For more information, see
  \href{https://www.uniprot.org/help/proteome}{What are proteomes?} and
  \href{https://www.uniprot.org/help/gene_centric_isoform_mapping}{Automatic
  gene-centric isoform mapping for eukaryotic reference proteome entries.}
}
\details{
At least one search criterion is required.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/genecentric"
}

\examples{
\donttest{
rba_uniprot_genecentric_search(accession = "P59594")
}
\donttest{
rba_uniprot_genecentric_search(gene = "Spike")
}
\donttest{
rba_uniprot_genecentric_search(upid = "UP000000354")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Proteomes":
\code{\link[=rba_uniprot_genecentric]{rba_uniprot_genecentric()}},
\code{\link[=rba_uniprot_proteomes]{rba_uniprot_proteomes()}},
\code{\link[=rba_uniprot_proteomes_search]{rba_uniprot_proteomes_search()}}
}
\concept{"UniProt - Proteomes"}
