% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_mutagenesis_search}
\alias{rba_uniprot_mutagenesis_search}
\title{Search Mutagenesis in UniProt}
\usage{
rba_uniprot_mutagenesis_search(
  accession = NULL,
  taxid = NULL,
  db_id = NULL,
  ...
)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{taxid}{Numeric: (optional) NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID}.
You can supply up to 20 taxon IDs.}

\item{db_id}{Character: (optional) The ID in a cross-reference database.
You can supply up to 20 values.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list in which each element represents a matching UniProt entry,
  named by accession when available. Mutagenesis annotations are stored in
  the entry's \code{features} element.
}
\description{
UniProt describes how sequence mutations affect the biological properties
  of a protein, cell, or organism. Use this function to search for
  \href{https://www.uniprot.org/help/mutagen}{
  mutagenesis annotations} using at least one of \code{accession},
  \code{taxid}, or \code{db_id}.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/mutagenesis"
}

\examples{
\donttest{
#search all mutations in COVID19 proteins
rba_uniprot_mutagenesis_search(taxid = 2697049)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Mutagenesis":
\code{\link[=rba_uniprot_mutagenesis]{rba_uniprot_mutagenesis()}}
}
\concept{"UniProt - Mutagenesis"}
