% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteins}
\alias{rba_uniprot_proteins}
\title{Get UniProt entry by accession}
\usage{
rba_uniprot_proteins(accession, interaction = FALSE, isoforms = FALSE, ...)
}
\arguments{
\item{accession}{Character:
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}.}

\item{interaction}{Logical: (default = \code{FALSE}) Retrieve
\href{https://www.uniprot.org/help/interaction_section}{interaction}
partners instead of the entry itself?}

\item{isoforms}{Logical: (default = \code{FALSE}) Retrieve
\href{https://www.uniprot.org/help/alternative_products}{isoforms} of your
supplied UniProt entry instead of the canonical entry itself?}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the requested UniProtKB entry. Isoform and
  interaction results are lists named by accession.
}
\description{
Retrieve a UniProtKB entry by accession. Alternatively, retrieve its
  isoforms or interaction partners by setting \code{isoforms = TRUE} or
  \code{interaction = TRUE}. These two modes are mutually exclusive.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/proteins/\{accession\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/proteins/interaction/\{accession\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/proteins/\{accession\}/isoforms"
}

\examples{
\donttest{
rba_uniprot_proteins(accession = "P01730")
}
\donttest{
rba_uniprot_proteins(accession = "P01730", interaction = TRUE)
}
\donttest{
rba_uniprot_proteins(accession = "Q29983", isoforms = TRUE)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Proteins":
\code{\link[=rba_uniprot_proteins_crossref]{rba_uniprot_proteins_crossref()}},
\code{\link[=rba_uniprot_proteins_search]{rba_uniprot_proteins_search()}}
}
\concept{"UniProt - Proteins"}
