% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomics.R
\name{rba_uniprot_proteomics_non_ptm_search}
\alias{rba_uniprot_proteomics_non_ptm_search}
\title{Search Proteomics data in UniProt}
\usage{
rba_uniprot_proteomics_non_ptm_search(
  accession = NULL,
  taxid = NULL,
  data_source = NULL,
  upid = NULL,
  peptide = NULL,
  unique = NULL,
  ...
)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{taxid}{Numeric: (optional) NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID}. You can supply up to
20 taxon IDs.}

\item{data_source}{Character: (optional) Proteomics data source. You can
supply up to two values. Use \code{rba_uniprot_proteomics_species()} to
retrieve the sources currently available for each species and category.}

\item{upid}{Character: (optional)
\href{https://www.uniprot.org/help/proteome_id}{UniProt Proteome
identifier (UPID)}. You can supply up to 100 UPIDs.}

\item{peptide}{Character: (optional) Peptide sequence(s). You can supply up
to 20 sequences.}

\item{unique}{Logical: (optional) Filter by peptide uniqueness. If
\code{TRUE}, return peptides mapping to one protein; if \code{FALSE},
return non-unique peptides; if \code{NULL}, do not apply this filter.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s arguments
manual for more information on available options.}
}
\value{
A list in which each element corresponds to a UniProt protein and
  proteomics data are stored under the \code{features} element.
}
\description{
Search for non-post-translational-modification proteomics features mapped to
  UniProt proteins. Refine the search by data source, peptide, or other
  supported criteria.
}
\details{
At least one of \code{accession}, \code{taxid}, \code{data_source},
  \code{upid}, or \code{peptide} is required. \code{unique} only refines
  those criteria.

UniProt categorizes proteomics data sources into three main data categories:
  PTM (Post-Translational Modification), non-PTM, and HPP (Human Proteome
  Project); each with corresponding API endpoints, and thus, rbioapi
  functions.
}
\section{Corresponding API Resources}{
 "GET
  https://www.ebi.ac.uk/proteins/api/proteomics/nonPtm"
}

\examples{
\donttest{
  rba_uniprot_proteomics_non_ptm_search(peptide = "NDQVYQPLRDRDDAQYSHLGGNWAR")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Proteomics":
\code{\link[=rba_uniprot_proteomics_hpp]{rba_uniprot_proteomics_hpp()}},
\code{\link[=rba_uniprot_proteomics_hpp_search]{rba_uniprot_proteomics_hpp_search()}},
\code{\link[=rba_uniprot_proteomics_non_ptm]{rba_uniprot_proteomics_non_ptm()}},
\code{\link[=rba_uniprot_proteomics_ptm]{rba_uniprot_proteomics_ptm()}},
\code{\link[=rba_uniprot_proteomics_ptm_search]{rba_uniprot_proteomics_ptm_search()}},
\code{\link[=rba_uniprot_proteomics_species]{rba_uniprot_proteomics_species()}}
}
\concept{"UniProt - Proteomics"}
