% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{rba_uniprot_proteomics_search}
\alias{rba_uniprot_proteomics_search}
\title{Search Proteomics Peptides in UniProt (Deprecated)}
\usage{
rba_uniprot_proteomics_search(
  accession = NULL,
  data_source = NULL,
  taxid = NULL,
  upid = NULL,
  peptide = NULL,
  unique = NULL,
  ...
)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{data_source}{Character: (optional) Proteomics data source. You can
supply up to two values. Use \code{rba_uniprot_proteomics_species()} to
retrieve the sources currently available for each species and category.}

\item{taxid}{Numeric: (optional) NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID}.
You can supply up to 20 taxon IDs.}

\item{upid}{Character: (optional)
\href{https://www.uniprot.org/help/proteome_id}{UniProt Proteome
identifier (UPID)}. You can supply up to 100 UPIDs.}

\item{peptide}{Character: (optional) Peptide sequence(s). You can supply up
to 20 sequences.}

\item{unique}{Logical: (optional) Filter by peptide uniqueness. If
\code{TRUE}, return peptides mapping to one protein; if \code{FALSE},
return non-unique peptides; if \code{NULL}, do not apply this filter.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list Where each element correspond to a UniProt protein and
  proteomics peptides are organized under the "features" sub-list.
}
\description{
This function is Deprecated. Please use
  \code{\link{rba_uniprot_proteomics_non_ptm_search}} instead.\cr
  UniProt maps proteomics peptides from different sources to the proteins'
  sequences. Using this function, you can search for  proteomics
  peptides that has been map to UniProt proteins. You may also refine your
  search with modifiers such as data_source, peptide etc. See
  "Arguments section" for more information.
}
\details{
Note that this is a search function. Thus, you are not required to fill
  every argument; You may use whatever combinations of arguments you see
  fit for your query.
  \cr see also: \href{https://www.uniprot.org/help/proteomics}{Mass
  spectrometry-based proteomics data in UniProtKB}
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/proteomics/nonPtm"
}

\examples{
\dontrun{
#Deprecated
rba_uniprot_proteomics_search(peptide = "MEDYTKIEK")
}
\dontrun{
#Deprecated
### this will generate a very large response!
  rba_uniprot_proteomics_search(taxid = 9606,
  data_source = "PeptideAtlas",
  progress = TRUE, timeout = 999999, unique = TRUE)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "Deprecated functions":
\code{\link[=rba_uniprot_proteomics]{rba_uniprot_proteomics()}},
\code{\link[=rba_uniprot_ptm]{rba_uniprot_ptm()}},
\code{\link[=rba_uniprot_ptm_search]{rba_uniprot_ptm_search()}}
}
\concept{"Deprecated functions"}
