% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy}
\alias{rba_uniprot_taxonomy}
\title{Get UniProt Taxonomy Nodes}
\usage{
rba_uniprot_taxonomy(
  ids,
  hierarchy = NULL,
  node_only = TRUE,
  page_size = 200,
  page_number = 1,
  ...
)
}
\arguments{
\item{ids}{Numeric: One or more
\href{https://www.uniprot.org/help/taxonomic_identifier}{NCBI taxonomic
identifiers}.}

\item{hierarchy}{Character: (optional) Retrieve nodes related to one supplied
node. One of "children", "parent", or "siblings".}

\item{node_only}{Logical: (default = \code{TRUE}) If \code{TRUE}, return
node information without links to parent, sibling, and child nodes.}

\item{page_size}{Numeric: (default = \code{200}) Number of child or sibling
nodes per page. The maximum is 200.}

\item{page_number}{Numeric: (default = \code{1}) Page of child or sibling
nodes to retrieve.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing taxonomy information for the requested nodes or
  their related nodes.
}
\description{
Retrieve taxonomic-node information using
\href{https://www.uniprot.org/help/taxonomic_identifier}{NCBI taxonomic
  identifiers}. You can also retrieve nodes related to one supplied node in
  \href{https://www.uniprot.org/help/taxonomy}{UniProt Taxonomy database}.
  Child and sibling results are paginated.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/ids/\{ids\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/ids/\{ids\}/node"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/node"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/children"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/children/node"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/parent"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/parent/node"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/siblings"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/id/\{id\}/siblings/node"
}

\examples{
\donttest{
rba_uniprot_taxonomy(ids = c(9606, 10090))
}
\donttest{
rba_uniprot_taxonomy(ids = 9989, hierarchy = "children")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Taxonomy":
\code{\link[=rba_uniprot_taxonomy_lca]{rba_uniprot_taxonomy_lca()}},
\code{\link[=rba_uniprot_taxonomy_lineage]{rba_uniprot_taxonomy_lineage()}},
\code{\link[=rba_uniprot_taxonomy_name]{rba_uniprot_taxonomy_name()}},
\code{\link[=rba_uniprot_taxonomy_path]{rba_uniprot_taxonomy_path()}},
\code{\link[=rba_uniprot_taxonomy_relationship]{rba_uniprot_taxonomy_relationship()}}
}
\concept{"UniProt - Taxonomy"}
