% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy_ancestor}
\alias{rba_uniprot_taxonomy_ancestor}
\title{This service returns the lowest common ancestor (LCA) of two taxonomy nodes.}
\usage{
rba_uniprot_taxonomy_ancestor(
  id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)

rba_uniprot_taxonomy_ancestor(
  id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{



}
\description{
This service returns the lowest common ancestor (LCA) of two taxonomy nodes.

This service returns the taxonomic lineage for a given taxonomy node. It
lists the nodes as they appear in the taxonomic tree, with the more
specific listed first.
}
