% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy_lca}
\alias{rba_uniprot_taxonomy_lca}
\title{Get Lowest Common Ancestor (LCA) of Two Taxonomy Nodes}
\usage{
rba_uniprot_taxonomy_lca(ids, ...)
}
\arguments{
\item{ids}{Numeric:
\href{https://www.uniprot.org/help/taxonomic_identifier}{NCBI taxonomic
identifiers}. Supply at least two IDs.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the lowest common ancestor's taxonomy information.
}
\description{
Retrieve the lowest common ancestor (LCA) of two or more nodes in the
  \href{https://www.uniprot.org/help/taxonomy}{UniProt Taxonomy database}.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/taxonomy/ancestor/\{ids\}"
}

\examples{
\donttest{
rba_uniprot_taxonomy_lca(c(9606,10090,9823,7712))
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Taxonomy":
\code{\link[=rba_uniprot_taxonomy]{rba_uniprot_taxonomy()}},
\code{\link[=rba_uniprot_taxonomy_lineage]{rba_uniprot_taxonomy_lineage()}},
\code{\link[=rba_uniprot_taxonomy_name]{rba_uniprot_taxonomy_name()}},
\code{\link[=rba_uniprot_taxonomy_path]{rba_uniprot_taxonomy_path()}},
\code{\link[=rba_uniprot_taxonomy_relationship]{rba_uniprot_taxonomy_relationship()}}
}
\concept{"UniProt - Taxonomy"}
