% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy_name}
\alias{rba_uniprot_taxonomy_name}
\title{Search UniProt Taxonomic Names}
\usage{
rba_uniprot_taxonomy_name(
  name,
  field = "scientific",
  search_type = "equal_to",
  node_only = TRUE,
  page_size = 200,
  page_number = 1,
  ...
)
}
\arguments{
\item{name}{Character: Taxonomic name to search.}

\item{field}{Character: (default = \code{"scientific"}) Name field to
search. One of "scientific", "common", or "mnemonic".}

\item{search_type}{Character: (default = \code{"equal_to"}) Relationship
between the query and taxonomic name. One of "equal_to", "start_with",
"end_with", or "contain".}

\item{node_only}{Logical: (default = \code{TRUE}) If \code{TRUE}, return
node information without links to parent, sibling, and child nodes.}

\item{page_size}{Numeric: (default = \code{200}) Number of results per page.
The maximum is 200.}

\item{page_number}{Numeric: (default = \code{1}) Page to retrieve.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing matching taxonomic nodes and page information.
}
\description{
Search and retrieve taxonomic nodes by name from the
  \href{https://www.uniprot.org/help/taxonomy}{UniProt Taxonomy database}.
  Search results are paginated.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/taxonomy/name/\{name\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/taxonomy/name/\{name\}/node"
}

\examples{
\donttest{
rba_uniprot_taxonomy_name(name = "homo", field = "scientific",
    search_type = "start_with")
}
\donttest{
rba_uniprot_taxonomy_name(name = "adenovirus", field = "scientific",
    search_type = "contain", page_size = 200, page_number = 2)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Taxonomy":
\code{\link[=rba_uniprot_taxonomy]{rba_uniprot_taxonomy()}},
\code{\link[=rba_uniprot_taxonomy_lca]{rba_uniprot_taxonomy_lca()}},
\code{\link[=rba_uniprot_taxonomy_lineage]{rba_uniprot_taxonomy_lineage()}},
\code{\link[=rba_uniprot_taxonomy_path]{rba_uniprot_taxonomy_path()}},
\code{\link[=rba_uniprot_taxonomy_relationship]{rba_uniprot_taxonomy_relationship()}}
}
\concept{"UniProt - Taxonomy"}
