% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc_bestguess}
\alias{rba_uniprot_uniparc_bestguess}
\title{Get UniParc Longest Sequence for Entries}
\usage{
rba_uniprot_uniparc_bestguess(
  upi = NULL,
  accession = NULL,
  db_id = NULL,
  gene = NULL,
  taxid = NULL,
  ...
)
}
\arguments{
\item{upi}{Character: (optional) Unique UniParc identifier(s). You can supply
up to 100 IDs.}

\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{db_id}{Character: (optional) Protein ID in a cross-reference database.
You can supply up to 100 IDs.}

\item{gene}{Character: (optional)
\href{https://www.uniprot.org/help/gene_name}{UniProt gene
name(s)}. You can supply up to 20 gene names.}

\item{taxid}{Numeric: (optional) NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID} used to refine the
search.
You can supply up to 20 taxon IDs.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
The best matching UniParc entry.
}
\description{
This function returns the UniParc Entry with a cross-reference to the
  longest active UniProtKB sequence (preferably from Swiss-Prot and if not
  then TrEMBL). If it finds more than one longest active UniProtKB sequence
  it returns 400 (Bad Request) error response with the list of cross
  references found.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/uniparc/bestguess"
}

\examples{
\donttest{
rba_uniprot_uniparc_bestguess("UPI00000000C9")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - UniParc":
\code{\link[=rba_uniprot_uniparc]{rba_uniprot_uniparc()}},
\code{\link[=rba_uniprot_uniparc_search]{rba_uniprot_uniparc_search()}},
\code{\link[=rba_uniprot_uniparc_sequence]{rba_uniprot_uniparc_sequence()}}
}
\concept{"UniProt - UniParc"}
