% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_variation_dbsnp}
\alias{rba_uniprot_variation_dbsnp}
\title{Get natural variants in UniProt by NIH-NCBI SNP database identifier}
\usage{
rba_uniprot_variation_dbsnp(
  db_id,
  save_peff = FALSE,
  source_type = NA,
  consequence_type = NA,
  wild_type = NA,
  alternative_sequence = NA,
  location = NA,
  ...
)
}
\arguments{
\item{save_peff}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get natural variants in UniProt by NIH-NCBI SNP database identifier
}
