% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_variation_hgvs}
\alias{rba_uniprot_variation_hgvs}
\title{Get natural variants in UniProt by HGVS expression}
\usage{
rba_uniprot_variation_hgvs(
  hgvs,
  save_peff = FALSE,
  source_type = NA,
  consequence_type = NA,
  wild_type = NA,
  alternative_sequence = NA,
  location = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get natural variants in UniProt by HGVS expression
}
