% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{v_msg}
\alias{v_msg}
\title{Alternative messaging system}
\usage{
v_msg(fmt, ..., sprintf = TRUE, cond = "verbose", sep = "", collapse = NULL)
}
\arguments{
\item{collapse}{}
}
\value{

}
\description{
Alternative messaging system
}
