% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_api_call}
\alias{rba_ba_api_call}
\title{Internal function to make http request}
\usage{
rba_ba_api_call(
  input_call,
  skip_error = FALSE,
  no_interet_retry_max = 1,
  no_internet_wait_time = 10,
  verbose = TRUE,
  diagnostics = FALSE
)
}
\arguments{
\item{input_call}{}
}
\value{

}
\description{
Internal function to make http request
}
