% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_file}
\alias{rba_ba_file}
\title{Create/check provided file address}
\usage{
rba_ba_file(file_ext, file_name = NA, dir_name = NA, save_to = TRUE)
}
\arguments{
\item{save_to}{}
}
\value{

}
\description{
changed name from rba_ba_file_path
}
