% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_http_status}
\alias{rba_ba_http_status}
\title{Translate HTTP status code to human readable explanation}
\usage{
rba_ba_http_status(http_status, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
changed name from rba_ba_translate
}
