% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_ba_reactome_input}
\alias{rba_ba_reactome_input}
\title{internal function to feed input to Reactome analysis}
\usage{
rba_ba_reactome_input(
  input,
  type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
internal function to feed input to Reactome analysis
}
