% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_ba_reactome_input}
\alias{rba_ba_reactome_input}
\title{internal function to feed input to Reactome analysis}
\usage{
rba_ba_reactome_input(input, type = NA, handle = TRUE)
}
\arguments{
\item{handle}{}
}
\value{

}
\description{
internal function to feed input to Reactome analysis
}
