% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_skeletion}
\alias{rba_ba_skeletion}
\title{General skeleton for all functions in the package}
\usage{
rba_ba_skeletion(
  call_function,
  batch_mode = FALSE,
  response_parser = NULL,
  parser_type = NA,
  skip_error = FALSE,
  user_agent = FALSE,
  progress_bar = FALSE,
  verbose = TRUE,
  diagnostics = FALSE,
  no_interet_retry_max = getOption("max_retry"),
  no_internet_wait_time = getOption("wait_time")
)
}
\arguments{
\item{progress_bar}{}
}
\value{

}
\description{
General skeleton for all functions in the package
}
