% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_ba_skeleton}
\alias{rba_ba_skeleton}
\title{General skeleton for all functions in the package}
\usage{
rba_ba_skeleton(
  input_call,
  response_parser = NULL,
  skip_error = FALSE,
  no_interet_retry_max = 1,
  no_internet_wait_time = 10
)
}
\arguments{
\item{input_call}{}
}
\value{

}
\description{
General skeleton for all functions in the package
}
