% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rba_connection_test}
\alias{rba_connection_test}
\title{Test if the Supported Services Are Responding}
\usage{
rba_connection_test(print_output = TRUE, diagnostics = FALSE)
}
\arguments{
\item{print_output}{Logical: (default = \code{TRUE}) Send the tests' output
to the console?}

\item{diagnostics}{Logical: (default = \code{FALSE}) Show diagnostics and
detailed messages with internal information.}
}
\value{
Connection test for the supported servers will be displayed
  in console and the results will be invisibly returned as a list.
}
\description{
Run this function to test the internet connectivity of your device and the
  current status of the supported Services.
}
\details{
This function attempts to send a simple query to the supported services.
  If the service successfully responded, you will be informed with a success
  message; If not, the content of the error will be reported to you.
  \cr Please run this function if you encounter any errors while using
  rbioapi. Also, if you need to contact support, kindly call this function
  with 'diagnostic = TRUE' and include the output messages in your support
  request.
}
\examples{
\donttest{
rba_connection_test()
}

}
\seealso{
Other "Helper functions":
\code{\link[=rba_metadata]{rba_metadata()}},
\code{\link[=rba_options]{rba_options()}},
\code{\link[=rba_pages]{rba_pages()}}
}
\concept{"Helper functions"}
\keyword{Helper}
