% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rba_connection_test}
\alias{rba_connection_test}
\title{Test if The supported servers are Responding}
\usage{
rba_connection_test(diagnostics = FALSE)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Test if The supported servers are Responding
}
