% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr}
\alias{rba_enrichr}
\title{Gene Set Enrichment Using Enrichr With One Function Call}
\usage{
rba_enrichr(
  gene_list,
  description = NA,
  gene_set_library = "all",
  regex_library_name = TRUE,
  multi_libs_progress_bar = TRUE,
  ...
)
}
\arguments{
\item{multi_libs_progress_bar}{}
}
\value{

}
\description{
Gene Set Enrichment Using Enrichr With One Function Call
}
