% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr}
\alias{rba_enrichr}
\title{A One-step Wrapper for Gene-list Enrichment Using Enrichr}
\usage{
rba_enrichr(
  gene_list,
  description = NULL,
  gene_set_library = "all",
  regex_library_name = TRUE,
  organism = "human",
  progress_bar = FALSE,
  ...
)
}
\arguments{
\item{gene_list}{A vector with Entrez gene symbols.}

\item{description}{(optional) A name or description to be associated with your
uploaded gene-set to Enrichr servers.}

\item{gene_set_library}{One of the:
\enumerate{
\item "all" to select all of the available Enrichr gene-set libraries.
\item A gene-set library name existed in the results
retrieved via \code{\link{rba_enrichr_libs}}
\item If regex_library_name = TRUE, A partially-matching name a regex
pattern that correspond to one or more of Enrichr library names.
}}

\item{regex_library_name}{logical: if TRUE (default) the supplied
gene_set_library will be regarded as a regex or partially matching name. if
FALSE, gene_set_library will be considered exact match.}

\item{organism}{(default = "human") Which model organism version of Enrichr
to use? Available options are: "human", (H. sapiens & M. musculus),
"fly" (D. melanogaster), "yeast" (S. cerevisiae), "worm" (C. elegans)
and "fish" (D. rerio).}

\item{progress_bar}{logical: In case of selecting multiple Enrichr
libraries, should a progress bar be displayed?}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing data frames of the enrichment results of your
  supplied gene-list against the selected Enrichr libraries.
}
\description{
This function is an easy-to-use wrapper for the multiple function calls
  necessary to enrich a given gene-list using Enrichr. see details section
  for more information.
}
\details{
This function will call other rba_enrichr_*** functions with the following
  order:
  \enumerate{
  \item (If neccessary) Call \code{\link{rba_enrichr_libs}} to obtain a list
    of available libraries in Enrichr.
  \item Call \code{\link{rba_enrichr_add_list}} to upload your gene-list
    and obtain a 'user list ID'.
  \item Call \code{\link{rba_enrichr_enrich}} to enrich the gene-list
    against one or multiple Enrichr libraries
  }
}
\section{Corresponding API Resources}{

 "GET https://maayanlab.cloud/Enrichr/datasetStatistics"
 \cr "POST https://maayanlab.cloud/Enrichr/addList"
 \cr "GET https://maayanlab.cloud/Enrichr/enrich"
}

\examples{
\donttest{
rba_enrichr(gene_list = c("TP53", "TNF", "EGFR"))
}
\donttest{
rba_enrichr(gene_list = c("TP53", "TNF", "EGFR"),
    gene_set_library = "GO_Molecular_Function_2017",
    regex_library_name = FALSE)
}
\donttest{
rba_enrichr(gene_list = c("TP53", "TNF", "EGFR"),
    gene_set_library = "go",
    regex_library_name = TRUE)
}

}
\references{
\itemize{
  \item Chen, E.Y., Tan, C.M., Kou, Y. et al. Enrichr: interactive and
  collaborative HTML5 gene list enrichment analysis tool. Bioinformatics
  14, 128 (2013). https://doi.org/10.1186/1471-2105-14-128
  \item Maxim V. Kuleshov, Matthew R. Jones, Andrew D. Rouillard, Nicolas
  F. Fernandez, Qiaonan Duan, Zichen Wang, Simon Koplev, Sherry L. Jenkins,
  Kathleen M. Jagodnik, Alexander Lachmann, Michael G. McDermott,
  Caroline D. Monteiro, Gregory W. Gundersen, Avi Ma’ayan, Enrichr: a
  comprehensive gene set enrichment analysis web server 2016 update,
  Nucleic Acids Research, Volume 44, Issue W1, 8 July 2016, Pages W90–W97,
  https://doi.org/10.1093/nar/gkw377
  \item Xie, Z., Bailey, A., Kuleshov, M. V., Clarke, D. J. B.,
  Evangelista, J. E., Jenkins, S. L., Lachmann, A., Wojciechowicz, M. L.,
  Kropiwnicki, E., Jagodnik, K. M., Jeon, M., & Ma’ayan, A. (2021). Gene
  set knowledge discovery with Enrichr. Current Protocols, 1, e90.
  doi: https://doi.org/10.1002/cpz1.90
  \item \href{https://maayanlab.cloud/Enrichr/help#api}{Enrichr API
  Documentation}
  \item \href{https://maayanlab.cloud/Enrichr/help#terms}{Citations note
  on Enrichr website}
  }
}
\seealso{
Other "Enrichr": 
\code{\link{rba_enrichr_add_list}()},
\code{\link{rba_enrichr_enrich}()},
\code{\link{rba_enrichr_gene_map}()},
\code{\link{rba_enrichr_libs}()},
\code{\link{rba_enrichr_view_list}()}
}
\concept{"Enrichr"}
