% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_add_list}
\alias{rba_enrichr_add_list}
\title{Upload your gene set to Enrichr}
\usage{
rba_enrichr_add_list(gene_list, description = NA, ...)
}
\arguments{
\item{gene_list}{}
}
\value{

}
\description{
Upload your gene set to Enrichr
}
