% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_enrich}
\alias{rba_enrichr_enrich}
\title{Get Enrichr enrichment results}
\usage{
rba_enrichr_enrich(
  user_list_id,
  gene_set_library = "all",
  regex_library_name = TRUE,
  multi_libs_progress_bar = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Get Enrichr enrichment results
}
