% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_gene_map}
\alias{rba_enrichr_gene_map}
\title{Find Enrichr Terms That Contain a Given Gene}
\usage{
rba_enrichr_gene_map(gene, categorize = FALSE, organism = "human", ...)
}
\arguments{
\item{gene}{character: An Entrez gene symbol.}

\item{categorize}{logical: Should the category information be included?}

\item{organism}{(default = "human") Which model organism version of Enrichr
to use? Available options are: "human", (H. sapiens & M. musculus),
"fly" (D. melanogaster), "yeast" (S. cerevisiae), "worm" (C. elegans)
and "fish" (D. rerio).}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A nested list containing the search results of your supplied gene.
}
\description{
This function will search the gene and retrieve a list of Enrichr
  Terms that contains that gene.
}
\section{Corresponding API Resources}{

 "GET https://maayanlab.cloud/Enrichr/genemap"
}

\examples{
\donttest{
rba_enrichr_gene_map(gene = "p53")
}
\donttest{
rba_enrichr_gene_map(gene = "p53", categorize = TRUE)
}

}
\references{
\itemize{
  \item Chen, E.Y., Tan, C.M., Kou, Y. et al. Enrichr: interactive and
  collaborative HTML5 gene list enrichment analysis tool. Bioinformatics
  14, 128 (2013). https://doi.org/10.1186/1471-2105-14-128
  \item Maxim V. Kuleshov, Matthew R. Jones, Andrew D. Rouillard, Nicolas
  F. Fernandez, Qiaonan Duan, Zichen Wang, Simon Koplev, Sherry L. Jenkins,
  Kathleen M. Jagodnik, Alexander Lachmann, Michael G. McDermott,
  Caroline D. Monteiro, Gregory W. Gundersen, Avi Ma’ayan, Enrichr: a
  comprehensive gene set enrichment analysis web server 2016 update,
  Nucleic Acids Research, Volume 44, Issue W1, 8 July 2016, Pages W90–W97,
  https://doi.org/10.1093/nar/gkw377
  \item Xie, Z., Bailey, A., Kuleshov, M. V., Clarke, D. J. B.,
  Evangelista, J. E., Jenkins, S. L., Lachmann, A., Wojciechowicz, M. L.,
  Kropiwnicki, E., Jagodnik, K. M., Jeon, M., & Ma’ayan, A. (2021). Gene
  set knowledge discovery with Enrichr. Current Protocols, 1, e90.
  doi: 10.1002/cpz1.90
  \item \href{https://maayanlab.cloud/Enrichr/help#api}{Enrichr API
  Documentation}
  \item \href{https://maayanlab.cloud/Enrichr/help#terms}{Citations note
  on Enrichr website}
  }
}
\seealso{
Other "Enrichr":
\code{\link[=rba_enrichr]{rba_enrichr()}},
\code{\link[=rba_enrichr_add_background]{rba_enrichr_add_background()}},
\code{\link[=rba_enrichr_add_list]{rba_enrichr_add_list()}},
\code{\link[=rba_enrichr_enrich]{rba_enrichr_enrich()}},
\code{\link[=rba_enrichr_libs]{rba_enrichr_libs()}},
\code{\link[=rba_enrichr_view_list]{rba_enrichr_view_list()}}
}
\concept{"Enrichr"}
