% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_gene_map}
\alias{rba_enrichr_gene_map}
\title{Find terms that contain a given gene}
\usage{
rba_enrichr_gene_map(
  gene,
  catagorize = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Find terms that contain a given gene
}
