% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_gene_map}
\alias{rba_enrichr_gene_map}
\title{Find Enrichr Terms That Contain a Given Gene}
\usage{
rba_enrichr_gene_map(gene, catagorize = FALSE, ...)
}
\arguments{
\item{gene}{character: An Entrez gene symbol.}

\item{catagorize}{logical: Should the catagory informations be included?}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
a list containing the search results of your provided gene.
}
\description{
This function will search your provided gene and retrieve a list of Enrichr
  Terms that contains that gene.
}
\section{Corresponding API Resources}{

 "GET http://maayanlab.cloud/Enrichr/genemap"
}

\examples{
rba_enrichr_gene_map(gene = "p53")
rba_enrichr_gene_map(gene = "p53", catagorize = TRUE)

}
\references{
\itemize{
  \item Chen EY, Tan CM, Kou Y, Duan Q, Wang Z, Meirelles GV, Clark NR,
  Ma'ayan A. Enrichr: interactive and collaborative HTML5 gene list
  enrichment analysis tool. BMC Bioinformatics. 2013;128(14).
  \item Kuleshov MV, Jones MR, Rouillard AD, Fernandez NF, Duan Q, Wang Z,
  Koplev S, Jenkins SL, Jagodnik KM, Lachmann A, McDermott MG, Monteiro CD,
  Gundersen GW, Ma'ayan A. Enrichr: a comprehensive gene set enrichment
  analysis web server 2016 update. Nucleic Acids Research. 2016; gkw377.
  \item \href{https://maayanlab.cloud/Enrichr/help#api}{Enrichr API
  Documentation}
  }
}
\seealso{
Other "Enrichr API": 
\code{\link{rba_enrichr_add_list}()},
\code{\link{rba_enrichr_enrich}()},
\code{\link{rba_enrichr_libs}()},
\code{\link{rba_enrichr_view_list}()},
\code{\link{rba_enrichr_wrp}()}
}
\concept{"Enrichr API"}
