% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_gene_sets}
\alias{rba_enrichr_gene_sets}
\title{Retrieve Gene Sets From an Enrichr Library}
\usage{
rba_enrichr_gene_sets(gene_set_library, term = NULL, organism = "human", ...)
}
\arguments{
\item{gene_set_library}{Character: A valid gene-set library name which exists
in the results retrieved via \code{\link{rba_enrichr_libs}}.}

\item{term}{Character: (optional) An exact Enrichr term name. If left NULL, all gene
sets in the supplied library will be returned.}

\item{organism}{Character: (default = \code{"human"}) Which model organism version of Enrichr
to use? Available options are: "human", (H. sapiens & M. musculus),
"fly" (D. melanogaster), "yeast" (S. cerevisiae), "worm" (C. elegans)
and "fish" (D. rerio).}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A named list of character vectors. Each element name is an Enrichr
  term and the corresponding character vector contains the genes in that
  gene set. If the \code{save_file} option was set to \code{TRUE} or a valid
  path, the raw GMT file will also be saved to the requested path.
}
\description{
This function retrieves all gene sets in a given Enrichr library. If a
  term is supplied, only the gene set associated with that term will be
  returned.
}
\details{
If the available Enrichr library names are not already stored in the
  "rba_enrichr_libs" global option, this function will retrieve and store
  them before validating the supplied library name.
}
\section{Corresponding API Resources}{

 "GET https://maayanlab.cloud/Enrichr/geneSetLibrary"
}

\section{Saving Gene Sets as a GMT File}{

 To save the raw server's response as a GMT file, supply the
 \code{save_file} rbioapi option through \code{...}. Set it to \code{TRUE}
 to automatically generate a proper file path, or supply a character string
 with a valid file or directory path. See \code{\link{rba_options}}'s
 arguments manual for more information on available options. The organism
 will be included in the automatically-generated file name. If a term is
 supplied, its name will also be included.
}

\examples{
\donttest{
rba_enrichr_gene_sets(
    gene_set_library = "Reactome_Pathways_2024",
    term = "Signaling by NOTCH")
}
\donttest{
rba_enrichr_gene_sets(gene_set_library = "Reactome_Pathways_2024")
}
\dontrun{
rba_enrichr_gene_sets(
    gene_set_library = "Reactome_Pathways_2024",
    save_file = "Reactome_Pathways_2024.gmt")
}

}
\references{
\itemize{
  \item Chen, E.Y., Tan, C.M., Kou, Y. et al. Enrichr: interactive and
  collaborative HTML5 gene list enrichment analysis tool. Bioinformatics
  14, 128 (2013). https://doi.org/10.1186/1471-2105-14-128
  \item Maxim V. Kuleshov, Matthew R. Jones, Andrew D. Rouillard, Nicolas
  F. Fernandez, Qiaonan Duan, Zichen Wang, Simon Koplev, Sherry L. Jenkins,
  Kathleen M. Jagodnik, Alexander Lachmann, Michael G. McDermott,
  Caroline D. Monteiro, Gregory W. Gundersen, Avi Ma’ayan, Enrichr: a
  comprehensive gene set enrichment analysis web server 2016 update,
  Nucleic Acids Research, Volume 44, Issue W1, 8 July 2016, Pages W90–W97,
  https://doi.org/10.1093/nar/gkw377
  \item Xie, Z., Bailey, A., Kuleshov, M. V., Clarke, D. J. B.,
  Evangelista, J. E., Jenkins, S. L., Lachmann, A., Wojciechowicz, M. L.,
  Kropiwnicki, E., Jagodnik, K. M., Jeon, M., & Ma’ayan, A. (2021). Gene
  set knowledge discovery with Enrichr. Current Protocols, 1, e90.
  doi: 10.1002/cpz1.90
  \item \href{https://maayanlab.cloud/Enrichr/help#api}{Enrichr API
  Documentation}
  \item \href{https://maayanlab.cloud/Enrichr/help#terms}{Citations note
  on Enrichr website}
  }
}
\seealso{
\code{\link{rba_enrichr}}

Other "Enrichr":
\code{\link[=rba_enrichr]{rba_enrichr()}},
\code{\link[=rba_enrichr_add_background]{rba_enrichr_add_background()}},
\code{\link[=rba_enrichr_add_list]{rba_enrichr_add_list()}},
\code{\link[=rba_enrichr_enrich]{rba_enrichr_enrich()}},
\code{\link[=rba_enrichr_gene_map]{rba_enrichr_gene_map()}},
\code{\link[=rba_enrichr_libs]{rba_enrichr_libs()}},
\code{\link[=rba_enrichr_view_list]{rba_enrichr_view_list()}}
}
\concept{"Enrichr"}
