% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_info}
\alias{rba_enrichr_info}
\title{Retrieve a List of available libraries from Enrichr}
\usage{
rba_enrichr_info(
  store_in_options = TRUE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
This function will retrieve a list of available libraries and
their statistics from Enrichr. You should call this function once per session
with argument 'store_in_options = TRUE' before querying data from Enrichr.
Nevertheless, rbioapi will do this for you in the background in the first
time you call any function pertinent to Enrichr.
}
