% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_view_list}
\alias{rba_enrichr_view_list}
\title{View your uploaded gene list}
\usage{
rba_enrichr_view_list(
  user_list_id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
View your uploaded gene list
}
