% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{rba_enrichr_view_list}
\alias{rba_enrichr_view_list}
\title{View your uploaded gene list}
\usage{
rba_enrichr_view_list(user_list_id, ...)
}
\arguments{
\item{user_list_id}{}
}
\value{

}
\description{
View your uploaded gene list
}
