% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_alignment_region}
\alias{rba_ensembl_alignment_region}
\title{Retrieves genomic alignments as separate blocks based on a region and species}
\usage{
rba_ensembl_alignment_region(
  region,
  species,
  aligned = TRUE,
  compact = TRUE,
  compara = "vertebrates",
  display_species_set = NA,
  mask = NA,
  method = "EPO",
  species_set = NA,
  species_set_group = "mammals",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves genomic alignments as separate blocks based on a region and species
}
