% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_archive}
\alias{rba_ensembl_archive}
\title{Retrieve the latest version for a set of identifiers}
\usage{
rba_ensembl_archive(ids, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Retrieve the latest version for a set of identifiers
}
