% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_eqtl.R
\name{rba_ensembl_eqtl_tissue}
\alias{rba_ensembl_eqtl_tissue}
\title{Returns all tissues currently available in the DB}
\usage{
rba_ensembl_eqtl_tissue(species, ...)
}
\arguments{
\item{species}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Returns all tissues currently available in the DB
}
