% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_eqtl.R
\name{rba_ensembl_eqtl_variant}
\alias{rba_ensembl_eqtl_variant}
\title{Returns the p-values for a SNP (e.g. rs123)}
\usage{
rba_ensembl_eqtl_variant(
  variant_name,
  species,
  stable_id = NA,
  statistic = NA,
  tissue = NA,
  ...
)
}
\arguments{
\item{tissue}{}
}
\value{

}
\description{
Returns the p-values for a SNP (e.g. rs123)
}
