% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_family}
\alias{rba_ensembl_family}
\title{Retrieves a family information using the family stable identifier}
\usage{
rba_ensembl_family(
  familiy_id,
  aligned = TRUE,
  compara = "vertebrates",
  member_source = "all",
  sequence = "protein",
  ...
)
}
\arguments{
\item{sequence}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Retrieves a family information using the family stable identifier
}
