% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_comp_gen.R
\name{rba_ensembl_homology_symbol}
\alias{rba_ensembl_homology_symbol}
\title{Retrieves homology information (orthologs) by symbol}
\usage{
rba_ensembl_homology_symbol(
  gene_symbol,
  species,
  aligned = TRUE,
  cigar_line = TRUE,
  compara = "vertebrates",
  external_db = NA,
  format = "full",
  sequence = "protein",
  target_species = NA,
  target_taxon = NA,
  type = "all",
  ...
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Retrieves homology information (orthologs) by symbol
}
