% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_infomation.R
\name{rba_ensembl_info_aassembly}
\alias{rba_ensembl_info_aassembly}
\title{List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.}
\usage{
rba_ensembl_info_aassembly(
  species,
  region_name = NA,
  bands = FALSE,
  synonyms = FALSE,
  ...
)
}
\arguments{
\item{region_name}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.
}
