% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_aassembly}
\alias{rba_ensembl_info_aassembly}
\title{List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.}
\usage{
rba_ensembl_info_aassembly(
  species,
  bands = FALSE,
  synonyms = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  store_in_options = TRUE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.
}
