% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_aassembly}
\alias{rba_ensembl_info_aassembly}
\title{List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.}
\usage{
rba_ensembl_info_aassembly(species, bands = FALSE, synonyms = FALSE, ...)

rba_ensembl_info_aassembly(species, bands = FALSE, synonyms = FALSE, ...)
}
\arguments{
\item{synonyms}{}
}
\value{



}
\description{
List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.

List the currently available assemblies for a species, along with toplevel
sequences, chromosomes and cytogenetic bands.
}
