% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_biotypes}
\alias{rba_ensembl_info_biotypes}
\title{List the functional classifications of gene models that Ensembl associates
 with a particular species. Useful for restricting the type of
 genes/transcripts retrieved by other endpoints.}
\usage{
rba_ensembl_info_biotypes(
  species,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
List the functional classifications of gene models that Ensembl associates
 with a particular species. Useful for restricting the type of
 genes/transcripts retrieved by other endpoints.
}
