% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_biotypes_groups}
\alias{rba_ensembl_info_biotypes_groups}
\title{Without argument the list of available biotype groups is returned. With
:group argument provided, list the properties of biotypes within that group.
Object type (gene or transcript) can be provided for filtering.}
\usage{
rba_ensembl_info_biotypes_groups(group = NA, object_type = NA, ...)
}
\arguments{
\item{object_type}{}
}
\value{

}
\description{
Without argument the list of available biotype groups is returned. With
:group argument provided, list the properties of biotypes within that group.
Object type (gene or transcript) can be provided for filtering.
}
