% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_compara}
\alias{rba_ensembl_info_compara}
\title{Lists all available comparative genomics databases and their data release.}
\usage{
rba_ensembl_info_compara(
  compara = NA,
  methods = FALSE,
  species_sets = FALSE,
  methods_class = NA,
  ...
)
}
\arguments{
\item{methods_class}{}
}
\value{

}
\description{
Lists all available comparative genomics databases and their data release.
}
