% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_compara_methods}
\alias{rba_ensembl_info_compara_methods}
\title{List all compara analyses available (an analysis defines the type of
comparative data).}
\usage{
rba_ensembl_info_compara_methods(
  method,
  compara = "vertebrates",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)

rba_ensembl_info_compara_methods(
  method,
  compara = "vertebrates",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)

rba_ensembl_info_compara_methods(
  method,
  compara = "vertebrates",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)

rba_ensembl_info_compara_methods(
  method,
  compara = "vertebrates",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{







}
\description{
List all compara analyses available (an analysis defines the type of
comparative data).

List all collections of species analysed with the specified compara method.

List all compara analyses available (an analysis defines the type of
comparative data).

List all collections of species analysed with the specified compara method.
}
