% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_eg_version}
\alias{rba_ensembl_info_eg_version}
\title{Returns the Ensembl Genomes version of the databases backing this service}
\usage{
rba_ensembl_info_eg_version(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Returns the Ensembl Genomes version of the databases backing this service
}
