% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes}
\alias{rba_ensembl_info_genomes}
\title{Find information about a given genome}
\usage{
rba_ensembl_info_genomes(
  genome_name,
  expand = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Find information about a given genome
}
