% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_genomes_name}
\alias{rba_ensembl_info_genomes_name}
\title{Find information about a given genome}
\usage{
rba_ensembl_info_genomes_name(genome_name, expand = FALSE, ...)
}
\arguments{
\item{expand}{}
}
\value{

}
\description{
Find information about a given genome
}
