% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_rest}
\alias{rba_ensembl_info_rest}
\title{Shows the current version of the Ensembl REST API.}
\usage{
rba_ensembl_info_rest(...)

rba_ensembl_info_rest(...)
}
\arguments{
\item{...}{}
}
\value{



}
\description{
Shows the current version of the Ensembl REST API.

Shows the current version of the Ensembl REST API.
}
