% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_species}
\alias{rba_ensembl_info_species}
\title{Lists all available species, their aliases, available adaptor groups and data release.}
\usage{
rba_ensembl_info_species(
  division = "EnsemblVertebrates",
  hide_strain_info = FALSE,
  strain_collection = NA,
  expand = FALSE,
  ...
)

rba_ensembl_info_species(
  division = "EnsemblVertebrates",
  hide_strain_info = FALSE,
  strain_collection = NA,
  expand = FALSE,
  ...
)
}
\arguments{
\item{expand}{}
}
\value{



}
\description{
Lists all available species, their aliases, available adaptor groups and data release.

Lists all available species, their aliases, available adaptor groups and data release.
}
