% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_species}
\alias{rba_ensembl_info_species}
\title{Lists all available species, their aliases, available adaptor groups and data release.}
\usage{
rba_ensembl_info_species(
  division = "EnsemblVertebrates",
  hide_strain_info = FALSE,
  strain_collection = NA,
  expand = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Lists all available species, their aliases, available adaptor groups and data release.
}
