% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_info.R
\name{rba_ensembl_info_variation_consequence_types}
\alias{rba_ensembl_info_variation_consequence_types}
\title{Lists all variant consequence types.}
\usage{
rba_ensembl_info_variation_consequence_types(rank = FALSE, ...)

rba_ensembl_info_variation_consequence_types(rank = FALSE, ...)
}
\arguments{
\item{rank}{}
}
\value{



}
\description{
Lists all variant consequence types.

Lists all variant consequence types.
}
