% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_infomation.R
\name{rba_ensembl_info_version}
\alias{rba_ensembl_info_version}
\title{Shows the data releases available on this REST server.}
\usage{
rba_ensembl_info_version(
  data = FALSE,
  REST = FALSE,
  software = FALSE,
  eg_version = FALSE,
  ...
)
}
\arguments{
\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Shows the data releases available on this REST server.
}
