% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_lookup.R
\name{rba_ensembl_lookup_id}
\alias{rba_ensembl_lookup_id}
\title{Find the species and database for several identifiers.
IDs that are not found are returned with no data}
\usage{
rba_ensembl_lookup_id(
  ids,
  db_type = NA,
  expand = FALSE,
  format = "full",
  object_type = NA,
  species = NA,
  ...
)
}
\arguments{
\item{species}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Find the species and database for several identifiers.
IDs that are not found are returned with no data
}
