% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_lookup_id}
\alias{rba_ensembl_lookup_id}
\title{Find the species and database for several identifiers.
IDs that are not found are returned with no data}
\usage{
rba_ensembl_lookup_id(
  ids,
  db_type = NA,
  expand = FALSE,
  format = "full",
  object_type = NA,
  species = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Find the species and database for several identifiers.
IDs that are not found are returned with no data
}
