% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_lookup_symbol}
\alias{rba_ensembl_lookup_symbol}
\title{Find the species and database for a set of symbols in a linked external
database. Unknown symbols are omitted from the response}
\usage{
rba_ensembl_lookup_symbol(
  symbols,
  species,
  expand = FALSE,
  format = "full",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Find the species and database for a set of symbols in a linked external
database. Unknown symbols are omitted from the response
}
