% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_mapping.R
\name{rba_ensembl_map_assembly}
\alias{rba_ensembl_map_assembly}
\title{Convert the co-ordinates of one assembly to another}
\usage{
rba_ensembl_map_assembly(
  asm_one,
  asm_two,
  region,
  species,
  coord_system = "chromosome",
  target_coord_system = "chromosome",
  ...
)
}
\arguments{
\item{target_coord_system}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Convert the co-ordinates of one assembly to another
}
