% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_map_cdna}
\alias{rba_ensembl_map_cdna}
\title{Convert from cDNA coordinates to genomic coordinates. Output reflects
forward orientation coordinates as returned from the Ensembl API}
\usage{
rba_ensembl_map_cdna(
  ensembl_id,
  region,
  species = NA,
  include_original_region = FALSE,
  ...
)
}
\arguments{
\item{include_original_region}{}
}
\value{

}
\description{
Convert from cDNA coordinates to genomic coordinates. Output reflects
forward orientation coordinates as returned from the Ensembl API
}
