% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_map_cds}
\alias{rba_ensembl_map_cds}
\title{Convert from CDS coordinates to genomic coordinates. Output reflects
forward orientation coordinates as returned from the Ensembl API.}
\usage{
rba_ensembl_map_cds(
  ensembl_id,
  region,
  species = NA,
  include_original_region = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Convert from CDS coordinates to genomic coordinates. Output reflects
forward orientation coordinates as returned from the Ensembl API.
}
