% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_mapping.R
\name{rba_ensembl_map_translation}
\alias{rba_ensembl_map_translation}
\title{Convert from protein (translation) coordinates to genomic coordinates.
Output reflects forward orientation coordinates as returned from the Ensembl API}
\usage{
rba_ensembl_map_translation(ensembl_id, region, species = NA, ...)
}
\arguments{
\item{species}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Convert from protein (translation) coordinates to genomic coordinates.
Output reflects forward orientation coordinates as returned from the Ensembl API
}
