% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_regulation.R
\name{rba_ensembl_microarray_binding_matrix}
\alias{rba_ensembl_microarray_binding_matrix}
\title{Return the specified binding matrix}
\usage{
rba_ensembl_microarray_binding_matrix(
  binding_matrix_id,
  species,
  unit = "frequencies",
  ...
)
}
\arguments{
\item{unit}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Return the specified binding matrix
}
