% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_microarray_probe}
\alias{rba_ensembl_microarray_probe}
\title{Returns information about a specific probe from a microarray}
\usage{
rba_ensembl_microarray_probe(
  microarray,
  probe,
  species,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Returns information about a specific probe from a microarray
}
