% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_microarray_species}
\alias{rba_ensembl_microarray_species}
\title{Returns information about all microarrays available for the given species}
\usage{
rba_ensembl_microarray_species(species, ...)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Returns information about all microarrays available for the given species
}
