% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_ontology_name}
\alias{rba_ensembl_ontology_name}
\title{Search for a list of ontological terms by their name}
\usage{
rba_ensembl_ontology_name(
  name,
  ontology = NA,
  relation = NA,
  simple = FALSE,
  ...
)
}
\arguments{
\item{simple}{}
}
\value{

}
\description{
Search for a list of ontological terms by their name
}
