% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_overlap_translation}
\alias{rba_ensembl_overlap_translation}
\title{Retrieve features related to a specific Translation as described by
its stable ID (e.g. domains, variants)}
\usage{
rba_ensembl_overlap_translation(
  ensembl_id,
  db_type = NA,
  feature = "protein_feature",
  so_term = NA,
  species = NA,
  type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieve features related to a specific Translation as described by
its stable ID (e.g. domains, variants)
}
