% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_phenotype_accession}
\alias{rba_ensembl_phenotype_accession}
\title{Return phenotype annotations for genomic features given a phenotype
ontology accession}
\usage{
rba_ensembl_phenotype_accession(
  accession,
  species,
  include_children = FALSE,
  include_pubmed_id = FALSE,
  include_review_status = FALSE,
  source = "undef",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Return phenotype annotations for genomic features given a phenotype
ontology accession
}
