% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_pheno_annot.R
\name{rba_ensembl_phenotype_gene}
\alias{rba_ensembl_phenotype_gene}
\title{Return phenotype annotations for a given gene}
\usage{
rba_ensembl_phenotype_gene(
  gene,
  species,
  include_associated = FALSE,
  include_overlap = FALSE,
  include_pubmed_id = FALSE,
  include_review_status = FALSE,
  include_submitter = FALSE,
  ...
)
}
\arguments{
\item{include_submitter}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Return phenotype annotations for a given gene
}
