% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_sequence_region}
\alias{rba_ensembl_sequence_region}
\title{Request multiple types of sequence by a list of regions}
\usage{
rba_ensembl_sequence_region(
  regions,
  species,
  coord_system = NA,
  coord_system_version = NA,
  expand_3prime = NA,
  expand_5prime = NA,
  format = NA,
  mask = NA,
  mask_feature = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Request multiple types of sequence by a list of regions
}
