% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_taxonomy.R
\name{rba_ensembl_taxonomy_classification}
\alias{rba_ensembl_taxonomy_classification}
\title{Return the taxonomic classification of a taxon node}
\usage{
rba_ensembl_taxonomy_classification(taxon_id, ...)
}
\arguments{
\item{taxon_id}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Return the taxonomic classification of a taxon node
}
