% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_taxonomy_classification}
\alias{rba_ensembl_taxonomy_classification}
\title{Return the taxonomic classification of a taxon node}
\usage{
rba_ensembl_taxonomy_classification(taxon_id, ...)
}
\arguments{
\item{taxon_id}{}
}
\value{

}
\description{
Return the taxonomic classification of a taxon node
}
