% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_transcript_haplotypes}
\alias{rba_ensembl_transcript_haplotypes}
\title{Computes observed transcript haplotype sequences based on phased genotype data}
\usage{
rba_ensembl_transcript_haplotypes(
  transcript_id,
  species,
  aligned_sequences = FALSE,
  samples = FALSE,
  sequence = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Computes observed transcript haplotype sequences based on phased genotype data
}
