% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_tran_hapl.R
\name{rba_ensembl_transcript_haplotypes}
\alias{rba_ensembl_transcript_haplotypes}
\title{Computes observed transcript haplotype sequences based on phased genotype data}
\usage{
rba_ensembl_transcript_haplotypes(
  transcript_id,
  species,
  aligned_sequences = FALSE,
  samples = FALSE,
  sequence = FALSE,
  ...
)
}
\arguments{
\item{sequence}{}
}
\value{

}
\description{
Computes observed transcript haplotype sequences based on phased genotype data
}
