% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_variation_pubmed}
\alias{rba_ensembl_variation_pubmed}
\title{Fetch variants by publication using PubMed Central reference number (PMCID)
or PubMed reference number (PMID)}
\usage{
rba_ensembl_variation_pubmed(pmid = NA, pmcid = NA, species, ...)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Fetch variants by publication using PubMed Central reference number (PMCID)
or PubMed reference number (PMID)
}
