% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_xrefs_id}
\alias{rba_ensembl_xrefs_id}
\title{Perform lookups of Ensembl Identifiers and retrieve their external
references in other databases}
\usage{
rba_ensembl_xrefs_id(
  ensembl_id,
  species = NA,
  all_levels = FALSE,
  db_type = NA,
  external_db = NA,
  object_type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Perform lookups of Ensembl Identifiers and retrieve their external
references in other databases
}
