% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_other.R
\name{rba_ensembl_xrefs_name}
\alias{rba_ensembl_xrefs_name}
\title{Performs a lookup based upon the primary accession or display label of an
external reference and returning the information we hold about the entry}
\usage{
rba_ensembl_xrefs_name(
  name,
  species,
  db_type = "core",
  external_db = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Performs a lookup based upon the primary accession or display label of an
external reference and returning the information we hold about the entry
}
