% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar.R
\name{rba_jaspar_matrix}
\alias{rba_jaspar_matrix}
\title{Get a Position Frequency Matrices (PFM) with annotations}
\usage{
rba_jaspar_matrix(matrix_id, file_format = NA, save_to = NA, ...)
}
\arguments{
\item{matrix_id}{Character: A matrix profile
Identifier. It has "base_id.version" naming schema.}

\item{file_format}{Character: Instead of returning a R object, you
can directly download the profile matrix in file with this format.
Supported formats are: "yaml", "jaspar", "transfac", "meme" and "pfm"}

\item{save_to}{NA or Character:\itemize{
\item NA: (only if file_format was provided) Save the file to an
  automatically-generated path.
\item Character string: A valid file path to save the file to.}}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A list that contains the PFM along with its details and
  annotations. If file_format was provided, an un-parsed character string
  with the file's content.
}
\description{
Using this function you can retrieve a Position Frequency Matrices (PFM)
 associated with a matrix profile Identifier along with its details and
 annotations. If a base ID (i.e. without version suffix) was provided,
 the latest version will be returned.
}
\section{Corresponding API Resources}{

 "GET "http://jaspar.genereg.net/api/v1/matrix/{matrix_id}/"
}

\examples{
\donttest{
rba_jaspar_matrix("MA0600.2")
}
\dontrun{
rba_jaspar_matrix(matrix_id = "MA0600.2",
                  file_format = "meme",
                  save_to = "my_matrix.meme")
}

}
\references{
\itemize{
  \item Fornes O, Castro-Mondragon JA, Khan A, et al. JASPAR 2020: update
    of the open-access database of transcription factor binding profiles.
    Nucleic Acids Res. 2019; doi: 10.1093/nar/gkz1001
  \item Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR
    data from any programming language. Bioinformatics, 2017, doi:
    10.1093/bioinformatics/btx804
  \item Khan, A. et al. JASPAR 2018: update of the open-access database
    of transcription factor binding profiles and its web framework.
    Nucleic Acids Res. 2018; 46:D260–D266, doi: 10.1093/nar/gkx1126
  \item
  \href{http://jaspar.genereg.net/api/v1/docs/}{JASPAR API Documentation}
  }
}
\seealso{
Other "JASPAR": 
\code{\link{rba_jaspar_collections_matrices}()},
\code{\link{rba_jaspar_collections}()},
\code{\link{rba_jaspar_matrix_search}()},
\code{\link{rba_jaspar_matrix_versions}()},
\code{\link{rba_jaspar_releases}()},
\code{\link{rba_jaspar_sites}()},
\code{\link{rba_jaspar_species_matrices}()},
\code{\link{rba_jaspar_species}()},
\code{\link{rba_jaspar_taxons_matrices}()},
\code{\link{rba_jaspar_taxons}()},
\code{\link{rba_jaspar_tffm_search}()},
\code{\link{rba_jaspar_tffm}()}
}
\concept{"JASPAR"}
