% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar.R
\name{rba_jaspar_matrix_versions}
\alias{rba_jaspar_matrix_versions}
\title{List matrix profile versions associated with a base ID}
\usage{
rba_jaspar_matrix_versions(base_id, order = NULL, ...)
}
\arguments{
\item{base_id}{Character: A base (stable) identifier. A matrix profile
identifier follows a "base_id.version" naming scheme.}

\item{order}{Character: A character string or a vector of character strings
of field names that will be used to order the results.
\cr Providing multiple field names is supported. You can also use the prefix
"-" before a field name to indicate reverse ordering.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame with information on the matrix profile versions.
}
\description{
Since JASPAR release 2010, matrix profiles have been versioned. A matrix
  profile identifier follows a "base_id.version" naming scheme. Using this
  function, you can retrieve a list of matrix
  profiles associated with a base (stable) ID.
}
\section{Corresponding API Resources}{

 "GET https://jaspar.elixir.no/api/v1/matrix/\{base_id\}/versions/"
}

\examples{
\donttest{
rba_jaspar_matrix_versions("MA0600")
}

}
\references{
\itemize{
  \item Baydar Ovek D, et al. JASPAR 2026: expansion of transcription
  factor binding profiles and integration of deep learning models.
  Nucleic Acids Res. 2026;54(D1):D184-D193;
  doi: 10.1093/nar/gkaf1209
  \item Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data
  from any programming language. Bioinformatics, 2017,
  doi: 10.1093/bioinformatics/btx804
  \item
  \href{https://jaspar.elixir.no/api/v1/docs/}{JASPAR API Documentation}
  \item \href{https://jaspar.elixir.no/faq/}{Citations note
  on JASPAR website}
  }
}
\seealso{
Other "JASPAR":
\code{\link[=rba_jaspar_collections]{rba_jaspar_collections()}},
\code{\link[=rba_jaspar_collections_matrices]{rba_jaspar_collections_matrices()}},
\code{\link[=rba_jaspar_matrix]{rba_jaspar_matrix()}},
\code{\link[=rba_jaspar_matrix_search]{rba_jaspar_matrix_search()}},
\code{\link[=rba_jaspar_releases]{rba_jaspar_releases()}},
\code{\link[=rba_jaspar_sites]{rba_jaspar_sites()}},
\code{\link[=rba_jaspar_species]{rba_jaspar_species()}},
\code{\link[=rba_jaspar_species_matrices]{rba_jaspar_species_matrices()}},
\code{\link[=rba_jaspar_taxons]{rba_jaspar_taxons()}},
\code{\link[=rba_jaspar_taxons_matrices]{rba_jaspar_taxons_matrices()}},
\code{\link[=rba_jaspar_tffm]{rba_jaspar_tffm()}},
\code{\link[=rba_jaspar_tffm_search]{rba_jaspar_tffm_search()}}
}
\concept{"JASPAR"}
