% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar.R
\name{rba_jaspar_releases}
\alias{rba_jaspar_releases}
\title{Get information about JASPAR database releases}
\usage{
rba_jaspar_releases(release_number = NULL, ...)
}
\arguments{
\item{release_number}{Numeric: (optional) Which JASPAR database release number should
be retrieved? If left NULL (the default), a list of all
JASPAR database releases will be returned.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list that contains all JASPAR database releases' information or
  details of a particular release.
}
\description{
If a release number is supplied, this function returns the details of that
  release. Otherwise, when called without the \code{release_number} argument,
  it returns a list of all JASPAR database releases.
}
\section{Corresponding API Resources}{

 "GET https://jaspar.elixir.no/api/v1/releases/"
 \cr "GET https://jaspar.elixir.no/api/v1/releases/\{release_number\}/"
}

\examples{
\donttest{
rba_jaspar_releases()
rba_jaspar_releases(11)
}

}
\references{
\itemize{
  \item Baydar Ovek D, et al. JASPAR 2026: expansion of transcription
  factor binding profiles and integration of deep learning models.
  Nucleic Acids Res. 2026;54(D1):D184-D193;
  doi: 10.1093/nar/gkaf1209
  \item Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data
  from any programming language. Bioinformatics, 2017,
  doi: 10.1093/bioinformatics/btx804
  \item
  \href{https://jaspar.elixir.no/api/v1/docs/}{JASPAR API Documentation}
  \item \href{https://jaspar.elixir.no/faq/}{Citations note
  on JASPAR website}
  }
}
\seealso{
Other "JASPAR":
\code{\link[=rba_jaspar_collections]{rba_jaspar_collections()}},
\code{\link[=rba_jaspar_collections_matrices]{rba_jaspar_collections_matrices()}},
\code{\link[=rba_jaspar_matrix]{rba_jaspar_matrix()}},
\code{\link[=rba_jaspar_matrix_search]{rba_jaspar_matrix_search()}},
\code{\link[=rba_jaspar_matrix_versions]{rba_jaspar_matrix_versions()}},
\code{\link[=rba_jaspar_sites]{rba_jaspar_sites()}},
\code{\link[=rba_jaspar_species]{rba_jaspar_species()}},
\code{\link[=rba_jaspar_species_matrices]{rba_jaspar_species_matrices()}},
\code{\link[=rba_jaspar_taxons]{rba_jaspar_taxons()}},
\code{\link[=rba_jaspar_taxons_matrices]{rba_jaspar_taxons_matrices()}},
\code{\link[=rba_jaspar_tffm]{rba_jaspar_tffm()}},
\code{\link[=rba_jaspar_tffm_search]{rba_jaspar_tffm_search()}}
}
\concept{"JASPAR"}
