% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar.R
\name{rba_jaspar_sites}
\alias{rba_jaspar_sites}
\title{Get binding sites of a matrix profile}
\usage{
rba_jaspar_sites(matrix_id, ...)
}
\arguments{
\item{matrix_id}{Character: A matrix profile
identifier. It follows a "base_id.version" naming scheme.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing a data frame with binding-site information.
}
\description{
Use this function to retrieve a list of transcription factor binding sites
  associated with a matrix profile.
}
\section{Corresponding API Resources}{

 "GET https://jaspar.elixir.no/api/v1/sites/\{matrix_id\}/"
}

\examples{
\donttest{
rba_jaspar_sites("MA0600.1")
}

}
\references{
\itemize{
  \item Baydar Ovek D, et al. JASPAR 2026: expansion of transcription
  factor binding profiles and integration of deep learning models.
  Nucleic Acids Res. 2026;54(D1):D184-D193;
  doi: 10.1093/nar/gkaf1209
  \item Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data
  from any programming language. Bioinformatics, 2017,
  doi: 10.1093/bioinformatics/btx804
  \item
  \href{https://jaspar.elixir.no/api/v1/docs/}{JASPAR API Documentation}
  \item \href{https://jaspar.elixir.no/faq/}{Citations note
  on JASPAR website}
  }
}
\seealso{
Other "JASPAR":
\code{\link[=rba_jaspar_collections]{rba_jaspar_collections()}},
\code{\link[=rba_jaspar_collections_matrices]{rba_jaspar_collections_matrices()}},
\code{\link[=rba_jaspar_matrix]{rba_jaspar_matrix()}},
\code{\link[=rba_jaspar_matrix_search]{rba_jaspar_matrix_search()}},
\code{\link[=rba_jaspar_matrix_versions]{rba_jaspar_matrix_versions()}},
\code{\link[=rba_jaspar_releases]{rba_jaspar_releases()}},
\code{\link[=rba_jaspar_species]{rba_jaspar_species()}},
\code{\link[=rba_jaspar_species_matrices]{rba_jaspar_species_matrices()}},
\code{\link[=rba_jaspar_taxons]{rba_jaspar_taxons()}},
\code{\link[=rba_jaspar_taxons_matrices]{rba_jaspar_taxons_matrices()}},
\code{\link[=rba_jaspar_tffm]{rba_jaspar_tffm()}},
\code{\link[=rba_jaspar_tffm_search]{rba_jaspar_tffm_search()}}
}
\concept{"JASPAR"}
