% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspar.R
\name{rba_jaspar_tffm}
\alias{rba_jaspar_tffm}
\title{Get information about a TF flexible model (TFFM)}
\usage{
rba_jaspar_tffm(tffm_id, ...)
}
\arguments{
\item{tffm_id}{Character: A TF flexible model (TFFM) identifier.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list that contains the TFFM's information and annotations.
}
\description{
Using this function, you can retrieve details and annotations for the
  transcription factor flexible model (TFFM) associated with a TFFM ID.
  If a base ID (i.e., without a version suffix) is supplied, the latest
  version will be returned.
}
\section{Corresponding API Resources}{

 "GET https://jaspar.elixir.no/api/v1/tffm/\{tffm_id\}/"
}

\examples{
\donttest{
rba_jaspar_tffm("TFFM0056.3")
}

}
\references{
\itemize{
  \item Baydar Ovek D, et al. JASPAR 2026: expansion of transcription
  factor binding profiles and integration of deep learning models.
  Nucleic Acids Res. 2026;54(D1):D184-D193;
  doi: 10.1093/nar/gkaf1209
  \item Khan, A. and Mathelier, A. JASPAR RESTful API: accessing JASPAR data
  from any programming language. Bioinformatics, 2017,
  doi: 10.1093/bioinformatics/btx804
  \item
  \href{https://jaspar.elixir.no/api/v1/docs/}{JASPAR API Documentation}
  \item \href{https://jaspar.elixir.no/faq/}{Citations note
  on JASPAR website}
  }
}
\seealso{
Other "JASPAR":
\code{\link[=rba_jaspar_collections]{rba_jaspar_collections()}},
\code{\link[=rba_jaspar_collections_matrices]{rba_jaspar_collections_matrices()}},
\code{\link[=rba_jaspar_matrix]{rba_jaspar_matrix()}},
\code{\link[=rba_jaspar_matrix_search]{rba_jaspar_matrix_search()}},
\code{\link[=rba_jaspar_matrix_versions]{rba_jaspar_matrix_versions()}},
\code{\link[=rba_jaspar_releases]{rba_jaspar_releases()}},
\code{\link[=rba_jaspar_sites]{rba_jaspar_sites()}},
\code{\link[=rba_jaspar_species]{rba_jaspar_species()}},
\code{\link[=rba_jaspar_species_matrices]{rba_jaspar_species_matrices()}},
\code{\link[=rba_jaspar_taxons]{rba_jaspar_taxons()}},
\code{\link[=rba_jaspar_taxons_matrices]{rba_jaspar_taxons_matrices()}},
\code{\link[=rba_jaspar_tffm_search]{rba_jaspar_tffm_search()}}
}
\concept{"JASPAR"}
