% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{rba_mieaa_cats}
\alias{rba_mieaa_cats}
\title{Get Supported Enrichment Categories for a Species and miRNA Type}
\usage{
rba_mieaa_cats(mirna_type, species, mode = "all", ...)
}
\arguments{
\item{mirna_type}{Character: Type of the miRNA identifiers; either "mature"
or "precursor".}

\item{species}{Character or Numeric: Scientific name, abbreviation, or NCBI
taxon ID of one of the following species: \enumerate{
\item "Homo sapiens", "hsa" or 9606
\item "Mus musculus", "mmu" or 10090
\item "Rattus norvegicus", "rno" or 10116
\item "Arabidopsis thaliana", "ath" or 3702
\item "Bos taurus", "bta" or 9913
\item "Caenorhabditis elegans", "cel" or 6239
\item "Drosophila melanogaster", "dme" or 7227
\item "Danio rerio", "dre" or 7955
\item "Gallus gallus", "gga" or 9031
\item "Sus scrofa", "ssc" or  9823}}

\item{mode}{Character: (default = \code{"all"}) Category subset to retrieve.
One of: "all" to include default and expert categories, "default" to
include only default categories, or "expert" to include only expert
categories.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A named character vector containing the supported categories for
  the supplied species, miRNA type, and mode. If the selected subset has no
  categories, returns \code{character(0)}.
}
\description{
Each combination of species and miRNA type supports a predefined set of
  enrichment categories. This function retrieves the categories available
  for a given combination.
}
\section{Corresponding API Resources}{

 "GET https://ccb-compute2.cs.uni-saarland.de/mieaa/api/v1/enrichment_categories/\{species\}/\{mirna_type\}/\{mode\}"
}

\examples{
\donttest{
rba_mieaa_cats("mature", "Homo sapiens")
}

}
\references{
\itemize{
  \item Ernesto Aparicio-Puerta, Pascal Hirsch, Georges P. Schmartz,
  Fabian Kern, Tobias Fehlmann, Andreas Keller, miEAA 2023: updates,
  new functional microRNA sets and improved enrichment visualizations,
  Nucleic Acids Research, Volume 51, Issue W1, 5 July 2023,
  Pages W319–W325, https://doi.org/10.1093/nar/gkad392
  \item
  \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/tutorial/api/}{miEAA
  browsable API tutorial}
  \item \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/}{Citation note
  on miEAA website}
  }
}
\seealso{
Other "miEAA":
\code{\link[=rba_mieaa_convert_type]{rba_mieaa_convert_type()}},
\code{\link[=rba_mieaa_convert_version]{rba_mieaa_convert_version()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_mieaa_enrich_results]{rba_mieaa_enrich_results()}},
\code{\link[=rba_mieaa_enrich_status]{rba_mieaa_enrich_status()}},
\code{\link[=rba_mieaa_enrich_submit]{rba_mieaa_enrich_submit()}}
}
\concept{"miEAA"}
