% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{rba_mieaa_convert_version}
\alias{rba_mieaa_convert_version}
\title{Convert miRNA Identifiers Between Different miRBase Versions}
\usage{
rba_mieaa_convert_version(
  mirna,
  mirna_type,
  input_version,
  output_version,
  simple_output = FALSE,
  ...
)
}
\arguments{
\item{mirna}{Character vector: miRNA identifiers to convert.}

\item{mirna_type}{Character: Type of the supplied miRNA identifiers; either
"mature" or "precursor".}

\item{input_version}{Numeric: miRBase version of the supplied identifiers.}

\item{output_version}{Numeric: miRBase version to which the identifiers
should be converted.}

\item{simple_output}{Logical: (default = \code{FALSE}) If \code{FALSE},
return a two-column data frame containing the input and output identifier
mappings. If \code{TRUE}, return only the converted identifiers without
their association with the supplied identifiers.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Depending on \code{simple_output}, a data frame or character vector
  containing the mappings returned by miEAA. Unrecognized or unmapped
  supplied identifiers can be omitted from the output.
}
\description{
miEAA uses miRBase v22 identifiers. This function converts a set of mature
  or precursor miRNA identifiers between two supported miRBase versions.
}
\section{Corresponding API Resources}{

 "POST https://ccb-compute2.cs.uni-saarland.de/mieaa/api/v1/mirbase_converter/"
}

\examples{
\donttest{
Sys.sleep(1) # to prevent 429 error during R CMD check
rba_mieaa_convert_version(mirna = c("hsa-miR-20b-5p", "hsa-miR-144-5p"),
    mirna_type = "mature", input_version = 22, output_version =  16)
}

}
\references{
\itemize{
  \item Ernesto Aparicio-Puerta, Pascal Hirsch, Georges P. Schmartz,
  Fabian Kern, Tobias Fehlmann, Andreas Keller, miEAA 2023: updates,
  new functional microRNA sets and improved enrichment visualizations,
  Nucleic Acids Research, Volume 51, Issue W1, 5 July 2023,
  Pages W319–W325, https://doi.org/10.1093/nar/gkad392
  \item
  \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/tutorial/api/}{miEAA
  browsable API tutorial}
  \item \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/}{Citation note
  on miEAA website}
  }
}
\seealso{
Other "miEAA":
\code{\link[=rba_mieaa_cats]{rba_mieaa_cats()}},
\code{\link[=rba_mieaa_convert_type]{rba_mieaa_convert_type()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_mieaa_enrich_results]{rba_mieaa_enrich_results()}},
\code{\link[=rba_mieaa_enrich_status]{rba_mieaa_enrich_status()}},
\code{\link[=rba_mieaa_enrich_submit]{rba_mieaa_enrich_submit()}}
}
\concept{"miEAA"}
