% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{rba_mieaa_enrich_results}
\alias{rba_mieaa_enrich_results}
\title{Retrieve the Results of a Finished miEAA Enrichment Analysis}
\usage{
rba_mieaa_enrich_results(job_id, sort_by = "p_adjusted", sort_asc = TRUE, ...)
}
\arguments{
\item{job_id}{Character: Job ID of a submitted enrichment analysis.}

\item{sort_by}{Character: (default = \code{"p_adjusted"}) Result column to
sort by. One of: "category", "subcategory", "enrichment", "p_value",
"p_adjusted", "q_value", or "observed".}

\item{sort_asc}{Logical: (default = \code{TRUE}) If \code{TRUE}, sort the
results in ascending order. If \code{FALSE}, sort them in descending order.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame with your enrichment analysis results.
}
\description{
After your submitted enrichment analysis request has finished (check
  using \code{\link{rba_mieaa_enrich_status}}), you can retrieve the results
  using this function.
}
\details{
Note that using \code{\link{rba_mieaa_enrich}} is a more convenient way to
  automatically perform this and other required function calls to
  perform enrichment analysis on your input miRNA-set using miEAA.
}
\section{Corresponding API Resources}{

 "GET https://ccb-compute2.cs.uni-saarland.de/mieaa/api/v1/enrichment_analysis/results/\{job_id\}/"
}

\examples{
\dontrun{
rba_mieaa_enrich_results("f52d1aef-6d3d-4d51-9020-82e68fe99012")
}

}
\references{
\itemize{
  \item Ernesto Aparicio-Puerta, Pascal Hirsch, Georges P. Schmartz,
  Fabian Kern, Tobias Fehlmann, Andreas Keller, miEAA 2023: updates,
  new functional microRNA sets and improved enrichment visualizations,
  Nucleic Acids Research, Volume 51, Issue W1, 5 July 2023,
  Pages W319–W325, https://doi.org/10.1093/nar/gkad392
  \item
  \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/tutorial/api/}{miEAA
  browsable API tutorial}
  \item \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/}{Citation note
  on miEAA website}
  }
}
\seealso{
Other "miEAA":
\code{\link[=rba_mieaa_cats]{rba_mieaa_cats()}},
\code{\link[=rba_mieaa_convert_type]{rba_mieaa_convert_type()}},
\code{\link[=rba_mieaa_convert_version]{rba_mieaa_convert_version()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_mieaa_enrich_status]{rba_mieaa_enrich_status()}},
\code{\link[=rba_mieaa_enrich_submit]{rba_mieaa_enrich_submit()}}
}
\concept{"miEAA"}
