% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieaa.R
\name{rba_mieaa_enrich_status}
\alias{rba_mieaa_enrich_status}
\title{Check the Status of a Submitted miEAA Enrichment Analysis}
\usage{
rba_mieaa_enrich_status(job_id, ...)
}
\arguments{
\item{job_id}{Character: Job ID of a submitted enrichment analysis.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing a \code{status} element with either the numeric
  completion percentage or \code{"FAILED"} for the supplied job ID. A
  completed job also includes its results URL.
}
\description{
After you have submitted your enrichment analysis (using
   \code{\link{rba_mieaa_enrich_submit}}) and retrieved a job-id,
  you can use this function to check the status of the job. The status is
  either a numeric completion percentage or \code{"FAILED"}. A status value
  equal to 100 means that the requested analysis has finished and you may
  retrieve the results using \code{\link{rba_mieaa_enrich_results}}.
}
\details{
Note that using \code{\link{rba_mieaa_enrich}} is a more convenient way to
  automatically perform this and other required function calls to
  perform enrichment analysis on your input miRNA-set using miEAA.
}
\section{Corresponding API Resources}{

 "GET https://ccb-compute2.cs.uni-saarland.de/mieaa/api/v1/job_status/\{job_id\}/"
}

\examples{
\dontrun{
Sys.sleep(1) # to prevent 429 error during R CMD check
rba_mieaa_enrich_status("f52d1aef-6d3d-4d51-9020-82e68fe99012")
}

}
\references{
\itemize{
  \item Ernesto Aparicio-Puerta, Pascal Hirsch, Georges P. Schmartz,
  Fabian Kern, Tobias Fehlmann, Andreas Keller, miEAA 2023: updates,
  new functional microRNA sets and improved enrichment visualizations,
  Nucleic Acids Research, Volume 51, Issue W1, 5 July 2023,
  Pages W319–W325, https://doi.org/10.1093/nar/gkad392
  \item
  \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/tutorial/api/}{miEAA
  browsable API tutorial}
  \item \href{https://ccb-compute2.cs.uni-saarland.de/mieaa/}{Citation note
  on miEAA website}
  }
}
\seealso{
Other "miEAA":
\code{\link[=rba_mieaa_cats]{rba_mieaa_cats()}},
\code{\link[=rba_mieaa_convert_type]{rba_mieaa_convert_type()}},
\code{\link[=rba_mieaa_convert_version]{rba_mieaa_convert_version()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_mieaa_enrich_results]{rba_mieaa_enrich_results()}},
\code{\link[=rba_mieaa_enrich_submit]{rba_mieaa_enrich_submit()}}
}
\concept{"miEAA"}
