% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rba_options}
\alias{rba_options}
\title{Set rbioapi Global Options}
\usage{
rba_options(
  client_timeout = NA,
  diagnostics = NA,
  dir_name = NA,
  max_retries = NA,
  progress_bar = NA,
  save_resp_file = NA,
  skip_error = NA,
  verbose = NA,
  wait_time = NA
)
}
\arguments{
\item{client_timeout}{numeric: The maximum time in seconds that you are
willing to wait for a server response before giving up and stopping the
function execution.}

\item{diagnostics}{logical: Generate diagnostics and detailed messages with
internal information.}

\item{dir_name}{character: If the package needed to generate a file path to
save the server's response, a directory with this name will be created in
your working directory to save your files.}

\item{max_retries}{numeric: How many times should rbioapi retry in case of
5xx server responses, errors or no internet connectivity?}

\item{progress_bar}{logical: Should a progress bar be displayed?}

\item{save_resp_file}{either:\itemize{
\item TRUE: in this case, the raw server response will be automatically
saved to a proper file path. use "dir_name" argument to alter the file's
parent directory.
\item FALSE: (default) Do not automatticaly save server response.
\item Character: (Only when altering the option via "..." argument) A valid
file path to save the server response to the calling function.}}

\item{skip_error}{logical: If TRUE, the code execution will not be stopped
in case of errors (anything but HTTP status 200 from the server); Instead
the error message will be returned as the function's output. Set this
to 'TRUE' if you are calling rbioapi functions in batch batch mode or
sourcing a script.}

\item{verbose}{logical: Generate informative messages.}

\item{wait_time}{numeric: Time in seconds to wait before each retry in case
of internet connection lost or server problems.}
}
\value{
If called without any argument, Data frame with available options
  and their pertinent information. If Called with an argument, will Return
  NULL but Alters that option globally.
}
\description{
A safe way to alter rbioapi's global options.\cr
  Note that you are not limited to changing the options globally, you can
  include the option names and values in the '...' argument of any rbioapi
  function to alter the option(s) only in that function call;
  e.x. example_function(x, y, diagnostics = TRUE).\cr\cr
  Alternatively, you can call this function with no arguments [ e.g.
  rba_options() ] to retrieve a table of available rbioapi options and their
  current values.\cr
}
\details{
Because this function validates your provided changes, Kindly
  \strong{\emph{only change rbioapi options using this function}} and avoid
  directly editing them.
}
\examples{
rba_options()
rba_options(verbose = FALSE)
rba_options(save_resp_file = TRUE)
rba_options(diagnostics = TRUE, progress_bar = TRUE)
}
\seealso{
Other "Helper functions": 
\code{\link{rba_citation}()},
\code{\link{rba_connection_test}()}
}
\concept{"Helper functions"}
\keyword{Helper}
