% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rba_options}
\alias{rba_options}
\title{Set rbioAPI global options}
\usage{
rba_options(
  verbose = NA,
  diagnostics = NA,
  progress_bar = NA,
  default_dir_name = NA,
  client_timeout = NA,
  wait_time = NA,
  max_retries = NA
)
}
\arguments{
\item{max_retries}{}
}
\value{

}
\description{
Set rbioAPI global options
}
