% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rba_options}
\alias{rba_options}
\title{Set rbioapi global options}
\usage{
rba_options(
  client_timeout = NA,
  diagnostics = NA,
  dir_name = NA,
  max_retries = NA,
  progress_bar = NA,
  save_resp_file = NA,
  skip_error = NA,
  verbose = NA,
  wait_time = NA
)
}
\arguments{
\item{wait_time}{}
}
\value{

}
\description{
Set rbioapi global options
}
