% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_family}
\alias{rba_panther_family}
\title{Get PANTHER Families and Sub-Families}
\usage{
rba_panther_family(id, what, target_organisms = NULL, ...)
}
\arguments{
\item{id}{Character: Panther family id.}

\item{what}{Character: What to retrieve? One of: \itemize{
\item "ortholog": Orthologs ('LDO' for least diverged and 'O' for more
  diverged).
\item "msa": Multiple Sequence Alignment Information,
\item "tree": Tree topology and nodes attributes.
}}

\item{target_organisms}{Numeric: (optional) NCBI taxon ID(s) to filter the results.
run \code{\link{rba_panther_info}} with argument 'what = "organisms"' to
get a list of PANTHER's supported organisms.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
For trees a list and otherwise a data frame with the requested
  family's information.
}
\description{
Using this function, you can retrieve Orthologs, MSA or Tree topology
  information of a given PANTHER family.
}
\section{Corresponding API Resources}{

 "POST https://www.pantherdb.org/services/oai/pantherdb/familyortholog"
 \cr "POST https://www.pantherdb.org/services/oai/pantherdb/familymsa"
 \cr "POST https://www.pantherdb.org/services/oai/pantherdb/treeinfo"
}

\examples{
\donttest{
rba_panther_family("PTHR10000", what = "ortholog")
}

}
\references{
\itemize{
  \item Thomas PD, Ebert D, Muruganujan A, Mushayahama T, Albou L-P,
  Mi H. (2022) PANTHER: Making genome-scale phylogenetics accessible to all.
  Protein Science, 31(1), 8–22.
  https://doi.org/10.1002/pro.4218
  \item \href{https://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  \item
  \href{https://www.pantherdb.org/publications.jsp#HowToCitePANTHER}{Citations
  note on PANTHER website}
  }
}
\seealso{
Other "PANTHER":
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_panther_genome]{rba_panther_genome()}},
\code{\link[=rba_panther_homolog]{rba_panther_homolog()}},
\code{\link[=rba_panther_info]{rba_panther_info()}},
\code{\link[=rba_panther_mapping]{rba_panther_mapping()}},
\code{\link[=rba_panther_ortholog]{rba_panther_ortholog()}},
\code{\link[=rba_panther_tree_grafter]{rba_panther_tree_grafter()}}
}
\concept{"PANTHER"}
