% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_genome}
\alias{rba_panther_genome}
\title{Retrieve Genes from a PANTHER Genome}
\usage{
rba_panther_genome(organism, page, ...)
}
\arguments{
\item{organism}{Numeric: NCBI taxon ID. Run
\code{\link{rba_panther_info}} with argument 'what = "organisms"' to get
a list of PANTHER's supported organisms.}

\item{page}{Numeric: The results page to retrieve. Pages contain up to
1,000 genes and are numbered starting from 1.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s arguments
manual for more information on available options.}
}
\value{
A list with the following elements: \describe{
  \item{gene}{A data frame with one row per returned gene. Fields are kept
  as returned by PANTHER, with annotation information in nested columns.}
  \item{page}{The retrieved page.}
  \item{pages_count}{The total number of available pages.}
  \item{number_of_genes_in_genome}{The total number of genes in the genome.}
  \item{product}{PANTHER product source and version information.}
  \item{search_type}{The search type reported by PANTHER.}
  }
}
\description{
Retrieve one page of genes and their associated information for a genome
  supported by PANTHER. Each page contains up to 1,000 genes.
}
\section{Corresponding API Resources}{

 "POST https://www.pantherdb.org/services/oai/pantherdb/downloadgenome"
}

\examples{
\donttest{
rba_panther_genome(organism = 243273, page = 1)
}

}
\references{
\itemize{
  \item Thomas PD, Ebert D, Muruganujan A, Mushayahama T, Albou L-P,
  Mi H. (2022) PANTHER: Making genome-scale phylogenetics accessible to all.
  Protein Science, 31(1), 8–22.
  https://doi.org/10.1002/pro.4218
  \item \href{https://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  \item
  \href{https://www.pantherdb.org/publications.jsp#HowToCitePANTHER}{Citations
  note on PANTHER website}
  }
}
\seealso{
Other "PANTHER":
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_panther_family]{rba_panther_family()}},
\code{\link[=rba_panther_homolog]{rba_panther_homolog()}},
\code{\link[=rba_panther_info]{rba_panther_info()}},
\code{\link[=rba_panther_mapping]{rba_panther_mapping()}},
\code{\link[=rba_panther_ortholog]{rba_panther_ortholog()}},
\code{\link[=rba_panther_tree_grafter]{rba_panther_tree_grafter()}}
}
\concept{"PANTHER"}
