% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_info}
\alias{rba_panther_info}
\title{Get PANTHER database Information}
\usage{
rba_panther_info(what, organism_chr_loc = FALSE, families_page = 1, ...)
}
\arguments{
\item{what}{Character: what information to retrieve? should be one of: \itemize{
\item "organisms": Retrieve supported organisms in PANTHER.
\item "datasets": Retrieve available annotation datasets.
\item "families": Retrieve available family IDs.
\item "species_tree": Retrieve PANTHER's species tree.
\item "pathways" Retrieve available pathway IDs.}}

\item{organism_chr_loc}{Logical: (default = \code{FALSE}) (only when 'what = "organisms"')
If TRUE, only organisms with chromosome location will be returned.
If FALSE (default), all organisms will be returned.}

\item{families_page}{Numeric: (default = \code{1}) (only when 'what = "families"')
Family information is very long, so results are returned in pages of up
to 1,000 families. Use a positive whole number to define the page to
retrieve.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
For families, a list containing family information, the requested
  page, and the total number of pages. For the species tree, a list; otherwise
  a data frame with pertinent information.
}
\description{
Using this function you can retrieve a list of available organisms,
  annotation datasets, families, and pathways which are supported in
  PANTHER.
}
\section{Corresponding API Resources}{

 "GET https://www.pantherdb.org/services/oai/pantherdb/supportedgenomes"
 \cr "GET https://www.pantherdb.org/services/oai/pantherdb/supportedannotdatasets"
 \cr "GET https://www.pantherdb.org/services/oai/pantherdb/supportedpantherfamilies"
 \cr "GET https://www.pantherdb.org/services/oai/pantherdb/supportedpantherpathways"
 \cr "GET https://www.pantherdb.org/services/oai/pantherdb/speciestree"
}

\examples{
\donttest{
rba_panther_info(what = "organisms")
}
\donttest{
rba_panther_info(what = "families", families_page = 4)
}

}
\references{
\itemize{
  \item Thomas PD, Ebert D, Muruganujan A, Mushayahama T, Albou L-P,
  Mi H. (2022) PANTHER: Making genome-scale phylogenetics accessible to all.
  Protein Science, 31(1), 8–22.
  https://doi.org/10.1002/pro.4218
  \item \href{https://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  \item
  \href{https://www.pantherdb.org/publications.jsp#HowToCitePANTHER}{Citations
  note on PANTHER website}
  }
}
\seealso{
Other "PANTHER":
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_panther_family]{rba_panther_family()}},
\code{\link[=rba_panther_genome]{rba_panther_genome()}},
\code{\link[=rba_panther_homolog]{rba_panther_homolog()}},
\code{\link[=rba_panther_mapping]{rba_panther_mapping()}},
\code{\link[=rba_panther_ortholog]{rba_panther_ortholog()}},
\code{\link[=rba_panther_tree_grafter]{rba_panther_tree_grafter()}}
}
\concept{"PANTHER"}
