% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_mapping}
\alias{rba_panther_mapping}
\title{Map A Gene-set to PANTHER Database}
\usage{
rba_panther_mapping(genes, organism, ...)
}
\arguments{
\item{genes}{Character or Numeric: A vector of gene identifiers with maximum
length of 5,000. Can be any of: Ensembl gene ID, Ensembl protein ID,
Ensembl transcript ID, Entrez gene ID, gene symbol, NCBI GI, HGNC ID,
International protein index ID, NCBI UniGene ID, UniProt accession and/or
UniProt ID.}

\item{organism}{Numeric: NCBI taxon ID. run \code{\link{rba_panther_info}}
with argument 'what = "organisms"' to get a list of PANTHER's
supported organisms.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing your unmapped inputs and mapped genes with
  pertinent information.
}
\description{
Using this function, you can search your genes in PANTHER database and
  retrieve attributes and annotations associated to your genes.
}
\section{Corresponding API Resources}{

 "POST https://www.pantherdb.org/services/oai/pantherdb/geneinfo"
}

\examples{
\donttest{
rba_panther_mapping(genes = c("Cd40", 7124, "ENSG00000203747", "P33681"),
    organism = 9606)
}

}
\references{
\itemize{
  \item Thomas PD, Ebert D, Muruganujan A, Mushayahama T, Albou L-P,
  Mi H. (2022) PANTHER: Making genome-scale phylogenetics accessible to all.
  Protein Science, 31(1), 8–22.
  https://doi.org/10.1002/pro.4218
  \item \href{https://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  \item
  \href{https://www.pantherdb.org/publications.jsp#HowToCitePANTHER}{Citations
  note on PANTHER website}
  }
}
\seealso{
Other "PANTHER":
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_panther_family]{rba_panther_family()}},
\code{\link[=rba_panther_genome]{rba_panther_genome()}},
\code{\link[=rba_panther_homolog]{rba_panther_homolog()}},
\code{\link[=rba_panther_info]{rba_panther_info()}},
\code{\link[=rba_panther_ortholog]{rba_panther_ortholog()}},
\code{\link[=rba_panther_tree_grafter]{rba_panther_tree_grafter()}}
}
\concept{"PANTHER"}
