% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_mapping}
\alias{rba_panther_mapping}
\title{Map A Gene-set to PANTHER Database}
\usage{
rba_panther_mapping(genes, organism, ...)
}
\arguments{
\item{genes}{Character vector of genes identifiers with maximum length of
1000. Can be any of: Ensemble gene ID, Ensemble protein ID, Ensemble
transcript ID, Entrez gene ID, gene symbol, NCBI GI, HGNC ID,
International protein index ID, NCBI UniGene ID, UniProt accession
and/or UniProt ID.}

\item{organism}{(numeric) NCBI taxon ID. run \code{\link{rba_panther_info}}
with argument 'what = "organisms"' to get a list of PANTHER's
supported organisms.}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A list containing your unmapped inputs and mapped genes with
  pertinent information.
}
\description{
Using this function, you can search your genes in PANTHER database and
  retrieve attributes and annotations associated to your genes.
}
\section{Corresponding API Resources}{

 "GET http://www.pantherdb.org/services/oai/pantherdb/geneinfo"
}

\examples{
rba_panther_mapping
}
\references{
\itemize{
  \item Mi, H., Muruganujan, A., Ebert, D., Huang, X., & Thomas, P. D.
  (2019). PANTHER version 14: more genomes, a new PANTHER GO-slim and
  improvements in enrichment analysis tools. Nucleic acids research, 47(D1),
  D419-D426.
  \item Mi, H., Muruganujan, A., Huang, X., Ebert, D., Mills, C., Guo, X.,
  & Thomas, P. D. (2019). Protocol Update for large-scale genome and gene
  function analysis with the PANTHER classification system (v. 14.0).
  Nature protocols, 14(3), 703-721.
  \item \href{http://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  }
}
\seealso{
Other "PANTHER API": 
\code{\link{rba_panther_enrich}()},
\code{\link{rba_panther_family}()},
\code{\link{rba_panther_homolog}()},
\code{\link{rba_panther_info}()},
\code{\link{rba_panther_ortholog}()},
\code{\link{rba_panther_tree_grafter}()}
}
\concept{"PANTHER API"}
