% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panther.R
\name{rba_panther_tree_grafter}
\alias{rba_panther_tree_grafter}
\title{PANTHER Tree Grafter
Use this function to retrieve a PANTHER family's tree topology information
  with a node corresponding to your sequence grafted in the best location
  in that tree.}
\usage{
rba_panther_tree_grafter(protein_seq, target_organisms = NA, ...)
}
\arguments{
\item{protein_seq}{A character string with the protein's sequence. Maximum
allowed sequence length is 50kb.}

\item{target_organisms}{(numeric) NCBI taxon ID(s) to filter the results.
run \code{\link{rba_panther_info}} with argument 'what = "organisms"' to
get a list of PANTHER's supported organisms.}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A list containing PANTHER tree topology information.
}
\description{
For more information, see:
  \href{www.pantherdb.org/tips/tips_graft_sequence.jsp}{PANTHER Graft
  Sequence Help} and
  \href{https://doi.org/10.1093/bioinformatics/bty625}{TreeGrafter:
  phylogenetic tree-based annotation of proteins with Gene Ontology terms
  and other annotations}
}
\section{Corresponding API Resources}{

 "GET http://www.pantherdb.org/services/oai/pantherdb/graftsequence"
}

\examples{
rba_panther_tree_grafter("MKVLWAALLVTFLAGCQAKVEQAVETE")
}
\references{
\itemize{
  \item Mi, H., Muruganujan, A., Ebert, D., Huang, X., & Thomas, P. D.
  (2019). PANTHER version 14: more genomes, a new PANTHER GO-slim and
  improvements in enrichment analysis tools. Nucleic acids research, 47(D1),
  D419-D426.
  \item Mi, H., Muruganujan, A., Huang, X., Ebert, D., Mills, C., Guo, X.,
  & Thomas, P. D. (2019). Protocol Update for large-scale genome and gene
  function analysis with the PANTHER classification system (v. 14.0).
  Nature protocols, 14(3), 703-721.
  \item \href{http://www.pantherdb.org/services/details.jsp}{PANTHER
  Services Details}
  }
}
\seealso{
Other "PANTHER API": 
\code{\link{rba_panther_enrich}()},
\code{\link{rba_panther_family}()},
\code{\link{rba_panther_homolog}()},
\code{\link{rba_panther_info}()},
\code{\link{rba_panther_mapping}()},
\code{\link{rba_panther_ortholog}()}
}
\concept{"PANTHER API"}
