% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis}
\alias{rba_reactome_analysis}
\title{identifiers Queries for multiple identifiers}
\usage{
rba_reactome_analysis(
  input,
  input_type = NA,
  projection = FALSE,
  interactors = FALSE,
  species = NA,
  sort_by = "ENTITIES_PVALUE",
  order = "ASC",
  resource = "TOTAL",
  p_value = NA,
  include_disease = TRUE,
  min = NA,
  max = NA,
  ...
)
}
\arguments{
\item{max}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
identifiers Queries for multiple identifiers
}
