% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_download}
\alias{rba_reactome_analysis_download}
\title{download Methods to download different views of a result}
\usage{
rba_reactome_analysis_download(
  token,
  request,
  save_to = NA,
  resource = "TOTAL",
  ...
)
}
\arguments{
\item{resource}{}
}
\value{

}
\description{
download Methods to download different views of a result
}
