% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_import}
\alias{rba_reactome_analysis_import}
\title{The accepted format is the same as provided by the method
/#/download/{token}/result.json. Note: The submitted file can be gzipped.}
\usage{
rba_reactome_analysis_import(
  input,
  input_type = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
The accepted format is the same as provided by the method
/#/download/{token}/result.json. Note: The submitted file can be gzipped.
}
