% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_import}
\alias{rba_reactome_analysis_import}
\title{Import Saved Analyhsis JSON to Reactome}
\usage{
rba_reactome_analysis_import(input, input_format = NA, ...)
}
\arguments{
\item{input}{A local file path or URL that points to your -optionally
gzipped- JSON file.}

\item{input_format}{(Optional) This function will automatically identify
your provided input's format. But in case of unexpected issues or if you
want to be explicit, set this argument to one of:\itemize{
\item "file": If you provided a local file path pointing to the JSON file.
\item "url": If you provided a URL pointing to the JSON file.}}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A list containing the new token and other information of your
  imported results.
}
\description{
If you have a JSON file of analysis results (only obtained via
  \code{\link{rba_reactome_analysis_download}} with the result argument
  set to "results", or "results_gz"), you can import the results back to
  Reactome and retrieve a token.\cr
  This is useful when you want to use other Reactome services which require
  a token but you do not have a token or your token has been
  expired (i.e. more than 7 days passed from your analysis).
}
\section{Corresponding API Resources}{

"GET https://reactome.org/AnalysisService/import/"\cr
"GET https://reactome.org/AnalysisService/import/form"\cr
"GET https://reactome.org/AnalysisService/import/url"
}

\examples{
\dontrun{rba_reactome_analysis_import("c:/rbioapi/res.json")}
\dontrun{rba_reactome_analysis_import("https://qaz.com/res.json.gz")}
}
\references{
\itemize{
  \item Fabregat A, Sidiropoulos K, Viteri G, Forner O, Marin-Garcia P,
  Arnau V, D'Eustachio P, Stein L, Hermjakob H. Reactome pathway analysis:
  a high-performance in-memory approach. BMC bioinformatics. 2017 Mar;18(1)
  142. doi: 10.1186/s12859-017-1559-2. PubMed PMID: 28249561.
  PubMed Central PMCID: PMC5333408.
  \item \href{https://reactome.org/AnalysisService/}{Reactome Analysis
  Services API Documentation}
  }
}
\seealso{
Other "Reactome Analysis Service, ": 
\code{\link{rba_reactome_analysis_download}()},
\code{\link{rba_reactome_analysis_mapping}()},
\code{\link{rba_reactome_analysis_pdf}()},
\code{\link{rba_reactome_analysis_species}()},
\code{\link{rba_reactome_analysis_token}()},
\code{\link{rba_reactome_analysis}()}
}
\concept{"Reactome Analysis Service, "}
