% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_mapping}
\alias{rba_reactome_analysis_mapping}
\title{Maps Molecule Identifiers}
\usage{
rba_reactome_analysis_mapping(
  input,
  input_format = NULL,
  projection = TRUE,
  interactors = FALSE,
  ...
)
}
\arguments{
\item{input}{Character or Numeric vector: A vector, local file path or URL
that points to your identifiers list.}

\item{input_format}{Character: (optional) This function will automatically identify
your supplied input's format. To be explicit, set this argument to one
of:\itemize{
\item "vector": If you supplied a simple vector (numeric or character) as
input.
\item "file": If you supplied a local file path pointing to a
correctly-formatted text file.
\item "url": If you supplied an HTTP or HTTPS URL pointing to a
correctly-formatted text file.}
An explicit value takes precedence. Otherwise, HTTP and HTTPS addresses
are identified first, followed by existing local files, and then other
non-empty character or numeric inputs as identifier vectors.}

\item{projection}{Logical: (default = \code{TRUE}) Should non-human identifiers
be projected to their human equivalents? (using Reactome orthology data)}

\item{interactors}{Logical: (default = \code{FALSE}) Should IntAct interaction data
be included?}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
List containing your identifiers and the IDS and resources they
  are mapped to.
}
\description{
Use this function to map molecule identifiers of different species to
  Reactome Identifiers.
}
\section{Corresponding API Resources}{

"POST https://reactome.org/AnalysisService/mapping/form"
\cr "POST https://reactome.org/AnalysisService/mapping/form/projection"
\cr "POST https://reactome.org/AnalysisService/mapping/url"
\cr "POST https://reactome.org/AnalysisService/mapping/url/projection"
}

\examples{
\dontrun{
rba_reactome_analysis_mapping(c("Q8SQ34", "cd40"))
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/AnalysisService/}{Reactome Analysis
  Services API Documentation}
  \item \href{https://reactome.org/cite}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Analysis Service":
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_reactome_analysis_download]{rba_reactome_analysis_download()}},
\code{\link[=rba_reactome_analysis_import]{rba_reactome_analysis_import()}},
\code{\link[=rba_reactome_analysis_pdf]{rba_reactome_analysis_pdf()}},
\code{\link[=rba_reactome_analysis_species]{rba_reactome_analysis_species()}},
\code{\link[=rba_reactome_analysis_token]{rba_reactome_analysis_token()}}
}
\concept{"Reactome Analysis Service"}
