% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_mapping}
\alias{rba_reactome_analysis_mapping}
\title{Maps the identifiers over the different species}
\usage{
rba_reactome_analysis_mapping(
  input,
  input_type = NA,
  projection = FALSE,
  interactors = FALSE,
  ...
)
}
\arguments{
\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Maps the identifiers over the different species
}
