% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_pdf}
\alias{rba_reactome_analysis_pdf}
\title{This method provides a report for a given pathway analysis result in a PDF
document. This document contains data about the analysis itself followed by
the pathways overview and the most significant pathways overlaid with the
analysis result. Users can download and store this information in a
convenient format to be checked in the future when the ‘token’ is not
longer available.}
\usage{
rba_reactome_analysis_pdf(
  token,
  species,
  save_to = NA,
  number = 25,
  resource = "TOTAL",
  diagram_profile = "Modern",
  analysis_profile = "Standard",
  fireworks_profile = "Barium Lithium",
  ...
)
}
\arguments{
\item{fireworks_profile}{}
}
\value{

}
\description{
This method provides a report for a given pathway analysis result in a PDF
document. This document contains data about the analysis itself followed by
the pathways overview and the most significant pathways overlaid with the
analysis result. Users can download and store this information in a
convenient format to be checked in the future when the ‘token’ is not
longer available.
}
