% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_species}
\alias{rba_reactome_analysis_species}
\title{Compare Human Pathways with with Other Species}
\usage{
rba_reactome_analysis_species(
  species_dbid,
  sort_by = "ENTITIES_PVALUE",
  order = "ASC",
  resource = "TOTAL",
  p_value = 1,
  min = NULL,
  max = NULL,
  ...
)
}
\arguments{
\item{species_dbid}{Numeric: Reactome DbId (e.g  Mus musculus is 48892) of
the species you want to compare with Homo sapiens. See
 \code{\link{rba_reactome_species}} or
 \href{https://reactome.org/content/schema/objects/Species/}{Reactome
 Data Schema: Entries: Species}.}

\item{sort_by}{Character: (default = \code{"ENTITIES_PVALUE"}) Sort the
result based on what column? Available choices are: "NAME",
"TOTAL_ENTITIES", "TOTAL_INTERACTORS", "TOTAL_REACTIONS", "FOUND_ENTITIES",
"FOUND_INTERACTORS", "FOUND_REACTIONS", "ENTITIES_RATIO",
"ENTITIES_PVALUE", "ENTITIES_FDR" or "REACTIONS_RATIO".}

\item{order}{Character: (default = \code{"ASC"}) Sort Order. Can be either
"ASC" or "DESC".}

\item{resource}{Character: (default = \code{"TOTAL"}) Filter results based on
the resource. Available choices are: "TOTAL", "UNIPROT", "ENSEMBL",
"CHEBI", "IUPHAR", "MIRBASE", "NCBI_PROTEIN", "EMBL", "COMPOUND" or
"PUBCHEM_COMPOUND".}

\item{p_value}{Numeric: (default = \code{1}) Set a P value threshold. Only
results with P value equal to or less than your supplied threshold will be
returned (1 means no P value filtering).}

\item{min}{Numeric: (optional) Minimum number of entities that a pathways
should have to be included in the results.}

\item{max}{Numeric: (optional) Maximum number of entities that a pathways
should have to be included in the results.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
List with the results of the comparison.
}
\description{
Use This function to Compare human's manually-curated pathways
  and computationally inferred pathways (orthologous) in other species.
}
\details{
Reactome incorporate manually curated human reactions and PANTHER's
  protein homology data to Computationally infer events in other eukaryotic
  species.
\cr Reactome uses an orthology-based approach to project curated human
  events to supported non-human species. See
  \href{https://reactome.org/documentation/inferred-events/}{
  Reactome Computationally Inferred Events} for more information.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/AnalysisService/species/homoSapiens/\{species\}"
}

\examples{
\donttest{
rba_reactome_analysis_species(species_dbid = 48892)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/AnalysisService/}{Reactome Analysis
  Services API Documentation}
  \item \href{https://reactome.org/cite}{Citations note on Reactome website}
  }
}
\seealso{
\code{\link{rba_reactome_orthology}}

Other "Reactome Analysis Service":
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_reactome_analysis_download]{rba_reactome_analysis_download()}},
\code{\link[=rba_reactome_analysis_import]{rba_reactome_analysis_import()}},
\code{\link[=rba_reactome_analysis_mapping]{rba_reactome_analysis_mapping()}},
\code{\link[=rba_reactome_analysis_pdf]{rba_reactome_analysis_pdf()}},
\code{\link[=rba_reactome_analysis_token]{rba_reactome_analysis_token()}}
}
\concept{"Reactome Analysis Service"}
