% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_species}
\alias{rba_reactome_analysis_species}
\title{Use page and pageSize to reduce the amount of data retrieved. Use sortBy
and order to sort the result by your preferred option. The resource field
will filter the results to show only those corresponding to the preferred
molecule type (TOTAL includes all the different molecules type)}
\usage{
rba_reactome_analysis_species(
  species_dbid,
  sort_by = "ENTITIES_PVALUE",
  order = "ASC",
  resource = "TOTAL",
  p_value = NA,
  min = NA,
  max = NA,
  ...
)
}
\arguments{
\item{max}{}
}
\value{

}
\description{
Use page and pageSize to reduce the amount of data retrieved. Use sortBy
and order to sort the result by your preferred option. The resource field
will filter the results to show only those corresponding to the preferred
molecule type (TOTAL includes all the different molecules type)
}
