% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_analysis.R
\name{rba_reactome_analysis_token}
\alias{rba_reactome_analysis_token}
\title{Use page and pageSize to reduce the amount of data retrieved. Use
sortBy and order to sort the result by your preferred option. The
resource field will filter the results to show only those corresponding
to the preferred molecule type (TOTAL includes all the different molecules
type)}
\usage{
rba_reactome_analysis_token(
  token,
  species,
  sort_by = "ENTITIES_PVALUE",
  order = "ASC",
  resource = "TOTAL",
  p_value = NA,
  include_disease = TRUE,
  min = NA,
  max = NA,
  ...
)
}
\arguments{
\item{max}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Use page and pageSize to reduce the amount of data retrieved. Use
sortBy and order to sort the result by your preferred option. The
resource field will filter the results to show only those corresponding
to the preferred molecule type (TOTAL includes all the different molecules
type)
}
