% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_complex_list}
\alias{rba_reactome_complex_list}
\title{Retrieves the list of complexes that contain a given (identifier, resource).
The method deconstructs the complexes into all its participants to do so.}
\usage{
rba_reactome_complex_list(id, resource, ...)
}
\arguments{
\item{resource}{}
}
\value{

}
\description{
Retrieves the list of complexes that contain a given (identifier, resource).
The method deconstructs the complexes into all its participants to do so.
}
