% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_complex_subunits}
\alias{rba_reactome_complex_subunits}
\title{Retrieves the list of subunits that constitute any given complex. In case
the complex comprises other complexes, this method recursively traverses
the content returning each contained PhysicalEntity. Contained complexes
and entity sets can be excluded setting the ‘excludeStructures’ optional
parameter to ‘true’}
\usage{
rba_reactome_complex_subunits(id, ...)

rba_reactome_complex_subunits(id, ...)
}
\arguments{
\item{id}{}

\item{exclude_structures}{}
}
\value{



}
\description{
Retrieves the list of subunits that constitute any given complex. In case
the complex comprises other complexes, this method recursively traverses
the content returning each contained PhysicalEntity. Contained complexes
and entity sets can be excluded setting the ‘excludeStructures’ optional
parameter to ‘true’

Retrieves a list containing all the reference entities for a given identifier.
}
