% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_diseases}
\alias{rba_reactome_diseases}
\title{It retrieves the list of diseases annotated in Reactome OR
list of disease DOIDs annotated in Reactome}
\usage{
rba_reactome_diseases(
  doid = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
It retrieves the list of diseases annotated in Reactome OR
list of disease DOIDs annotated in Reactome
}
