% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_entity_other_forms}
\alias{rba_reactome_entity_other_forms}
\title{Retrieves a list containing all other forms of the given PhysicalEntity.
These other forms are PhysicalEntities that share the same ReferenceEntity
identifier, e.g. PTEN H93R[R-HSA-2318524] and PTEN C124R[R-HSA-2317439]
are two forms of PTEN.}
\usage{
rba_reactome_entity_other_forms(
  entity_id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Retrieves a list containing all other forms of the given PhysicalEntity.
These other forms are PhysicalEntities that share the same ReferenceEntity
identifier, e.g. PTEN H93R[R-HSA-2318524] and PTEN C124R[R-HSA-2317439]
are two forms of PTEN.
}
