% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_event_ancestors}
\alias{rba_reactome_event_ancestors}
\title{Get Reactome Events Ancestors}
\usage{
rba_reactome_event_ancestors(event_id, ...)
}
\arguments{
\item{event_id}{Character: Reactome event's identifier.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
List which every element is a Data frame listing your supplied
  event along with it's ancestor events. Because any given event can be
  part of more than one pathway hierarchy, the list may contain multiple
  data frames.
}
\description{
Along with Reactome's events hierarchy, This function will retrieve all the
 events beginning from your supplied event up to the "Top level Pathway".
 see "Details section" for more information.
}
\details{
By Reactome's definition, Events are the building blocks of biological
  processes and could be of two main classes: "Pathway" or
  "Reaction-like events". The events are organized in a hierarchical
  structure; and each event could be child or parent to another event; The
  hierarchy will always begin with a "Top level pathway" event. Also note
  that a given event could be part of more that one hierarchies.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/event/\{id\}/ancestors"
}

\examples{
\donttest{
rba_reactome_event_ancestors("R-HSA-5673001")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Queries Related to Events":
\code{\link[=rba_reactome_event_hierarchy]{rba_reactome_event_hierarchy()}}
}
\concept{"Reactome Content Service - Queries Related to Events"}
