% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_event_ancestors}
\alias{rba_reactome_event_ancestors}
\title{The Reactome definition of events includes pathways and reactions. Although
events are organised in a hierarchical structure, a single event can be in
more than one location, i.e. a reaction can take part in different pathways
while, in the same way, a sub-pathway can take part in many pathways.
Therefore, this method retrieves a list of all possible paths from the
requested event to the top level pathway(s).}
\usage{
rba_reactome_event_ancestors(event_id, ...)
}
\arguments{
\item{event_id}{}
}
\value{

}
\description{
The Reactome definition of events includes pathways and reactions. Although
events are organised in a hierarchical structure, a single event can be in
more than one location, i.e. a reaction can take part in different pathways
while, in the same way, a sub-pathway can take part in many pathways.
Therefore, this method retrieves a list of all possible paths from the
requested event to the top level pathway(s).
}
