% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_event_hierarchy}
\alias{rba_reactome_event_hierarchy}
\title{Get Full Event Hierarchy of a Species}
\usage{
rba_reactome_event_hierarchy(species, ...)
}
\arguments{
\item{species}{Character or Numeric: NCBI Taxonomy identifier (Human Taxonomy
ID is 9606.) or species name (e.g. "Homo sapiens"). See
\code{\link{rba_reactome_species}} or
\href{https://reactome.org/content/schema/objects/Species/}{Reactome
Data Schema: Entries: Species}.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
List which is a representation of the species's events hierarchy
  described in the "Details section".
}
\description{
This function will retrieve the full Events hierarchy of your supplied
  species. Directly under each species, each child element is a "top Level
  Pathway". You can traverse the events tree down by following the "children"
  element.
}
\details{
By Reactome's definition, Events are the building blocks of biological
  processes and could be of two main classes: "Pathway" or
  "Reaction-like events". The events are organized in a hierarchical
  structure; and each event could be child or parent to another event; The
  hierarchy will always begin with a "Top level pathway" event. Also note
  that a given event could be part of more that one hierarchies.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/eventsHierarchy/\{species\}"
}

\examples{
\dontrun{
#very large response!
rba_reactome_event_hierarchy("Homo sapiens")
}
\dontrun{
#very large response!
rba_reactome_event_hierarchy(9606)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Queries Related to Events":
\code{\link[=rba_reactome_event_ancestors]{rba_reactome_event_ancestors()}}
}
\concept{"Reactome Content Service - Queries Related to Events"}
