% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_event_hierarchy}
\alias{rba_reactome_event_hierarchy}
\title{Events (pathways and reactions) in Reactome are organised in a hierarchical
structure for every species. By following all ‘hasEvent’ relationships,
this method retrieves the full event hierarchy for any given species.
The result is a list of tree structures, one for each TopLevelPathway.
Every event in these trees is represented by a PathwayBrowserNode.
The latter contains the stable identifier, the name, the species, the url,
the type, and the diagram of the particular event.}
\usage{
rba_reactome_event_hierarchy(species, ...)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Events (pathways and reactions) in Reactome are organised in a hierarchical
structure for every species. By following all ‘hasEvent’ relationships,
this method retrieves the full event hierarchy for any given species.
The result is a list of tree structures, one for each TopLevelPathway.
Every event in these trees is represented by a PathwayBrowserNode.
The latter contains the stable identifier, the name, the species, the url,
the type, and the diagram of the particular event.
}
