% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_diagram}
\alias{rba_reactome_exporter_diagram}
\title{Get a Reactome Event Diagram}
\usage{
rba_reactome_exporter_diagram(
  event_id,
  save_to = NULL,
  create_document = FALSE,
  resource = "TOTAL",
  diagram_profile = "Modern",
  analysis_profile = "Standard",
  token = NULL,
  exp_column = NULL,
  document_level = 1,
  output_format = "png",
  image_quality = 5,
  flag_element = NULL,
  flg_interactors = TRUE,
  sel = NULL,
  title = TRUE,
  margin = 15,
  ehld = TRUE,
  ...
)
}
\arguments{
\item{event_id}{Character: Reactome event's identifier.}

\item{save_to}{NULL or Character: (default = \code{NULL}) \itemize{
\item NULL: Save the file to an automatically-generated path.
\item Character string: A valid file path to save the file to.}}

\item{create_document}{Logical: (default = \code{FALSE}) Create PDF document
instead of image?}

\item{resource}{Character: (default = \code{"TOTAL"}) The analysis resource
for which the results will be overlaid on top of the given pathways
overview.}

\item{diagram_profile}{Character: (default = \code{"Modern"}) Color profile
of diagrams, should be either "Modern" or "Standard".}

\item{analysis_profile}{Character: (default = \code{"Standard"}) Color
profile of analysis, should be one of: "Standard", "Strosobar" or "Copper
Plus".}

\item{token}{Character: (optional) The analysis Token for which the results
will be overlaid on top of the given pathways overview. see:
\code{\link{rba_reactome_analysis}}.}

\item{exp_column}{Numeric: (optional) (only if token is supplied) Specify the
expression column for the overlay.}

\item{document_level}{Numeric: (default = \code{1}) (Only if
"create_document" is TRUE) use 0 to exclude the event's children or 1 to
include them.}

\item{output_format}{Character: (default = \code{"png"}) (Only if
"create_document" is FALSE) Image format of the saved diagram. Can be one
of: png, jpg, jpeg, svg or gif.}

\item{image_quality}{Numeric: (default = \code{5}) (Only if "create_document"
is FALSE), a number ranging from 1 to 10. 1 is the lowest quality and 10 is
the highest.}

\item{flag_element}{Character: (optional) (Only if "create_document" is
FALSE) gene name, protein ID, chemical ID or Reactome ID of a diagram's
element to be flagged.}

\item{flg_interactors}{Logical: (default = \code{TRUE}) (Only if
"create_document" is FALSE) Should the interactor be considered when
flagging a diagram element?}

\item{sel}{Character vector: (optional) (Only if "create_document" is FALSE)
CSV line for highlighting element(s) selection in the diagram.}

\item{title}{Logical: (default = \code{TRUE}) (Only if "create_document" is
FALSE) Should the pathway name be displayed below the image?}

\item{margin}{Numeric: (default = \code{15}) (Only if "create_document" is
FALSE) A number ranging from 0 to 20 to set as the image's margin.}

\item{ehld}{Logical: (default = \code{TRUE}) (Only if "create_document" is
FALSE) Should Enhanced High Level Diagrams be considered?}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
NULL, Based to the inputs, an image or PDF file will be saved to
  disk.
}
\description{
This function could be called in two scenarios: \enumerate{
  \item With create_document = FALSE: To retrieve an image of that event's
  Diagram.
  \item With create_document = TRUE: To retrieve a PDF document with the
  event's diagram image and additional information.}
  see "Details section" for more information
}
\details{
If the function is called with create_document = FALSE:
  \cr The result will be an image with the format supplied in "output_format"
  argument. If the supplied event ID refers to a pathway, the image's
  content will be the that pathways diagram. If the supplied event ID refers
  to a sub-pathway or reaction event, the parent pathway's diagram will be
  exported, with that reaction or sub-pathway's events highlighted.
  \cr Note that to export an image of reaction-like event separately, you
  should use \code{\link{rba_reactome_exporter_reaction}}.
  \cr If the function is called with create_document = TRUE:
  \cr A PDF document will contain an image of the event's diagram and the
  following information of that events: Summation, Literature references,
  Edit history type, location, compartments and diseases.
  note that if you call the function with "document level = 1", information
  of your supplied event's children will also be included.
}
\section{Corresponding API Resources}{

  "GET https://reactome.org/ContentService/exporter/diagram/\{identifier\}
  .\{ext\}"
  \cr "GET https://reactome.org/ContentService/exporter/document/event/
  \{identifier\}.pdf"
}

\examples{
\dontrun{
rba_reactome_exporter_diagram(event_id = "R-HSA-177929",
  create_document = FALSE)
}
\dontrun{
rba_reactome_exporter_diagram(event_id = "R-HSA-6787403",
    create_document = FALSE)
}
\dontrun{
rba_reactome_exporter_diagram(event_id = "R-HSA-177929",
    create_document = TRUE)
}
\dontrun{
rba_reactome_exporter_diagram(event_id = "R-HSA-177929",
    output_format = "svg",
    save_to = "reactome_event_diagram.svg")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
\code{\link{rba_reactome_exporter_reaction}}
\code{\link{rba_reactome_analysis}}

Other "Reactome Content Service - Format Exporter":
\code{\link[=rba_reactome_exporter_event]{rba_reactome_exporter_event()}},
\code{\link[=rba_reactome_exporter_overview]{rba_reactome_exporter_overview()}},
\code{\link[=rba_reactome_exporter_reaction]{rba_reactome_exporter_reaction()}}
}
\concept{"Reactome Content Service - Format Exporter"}
