% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_diagram}
\alias{rba_reactome_exporter_diagram}
\title{This method accepts identifiers for Event class instances.
When a diagrammed pathway is provided, the diagram is exported to the
specified format. When a subpathway is provided, the diagram for
the parent is exported and the events that are part of the subpathways
are selected. When a reaction is provided, the diagram containing the
reaction is exported and the reaction is selected.}
\usage{
rba_reactome_exporter_diagram(
  event_id,
  output_format = "png",
  save_to = NA,
  quality = 5,
  flg = NA,
  flg_interactors = TRUE,
  sel = NA,
  title = TRUE,
  margin = 15,
  ehld = TRUE,
  diagram_profile = "Modern",
  token = NA,
  resource = "TOTAL",
  analysis_profile = "Standard",
  exp_column = NA,
  ...
)
}
\arguments{
\item{exp_column}{}
}
\value{

}
\description{
This method accepts identifiers for Event class instances.
When a diagrammed pathway is provided, the diagram is exported to the
specified format. When a subpathway is provided, the diagram for
the parent is exported and the events that are part of the subpathways
are selected. When a reaction is provided, the diagram containing the
reaction is exported and the reaction is selected.
}
