% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_event}
\alias{rba_reactome_exporter_event}
\title{Exports a given pathway or reaction to SBGN or SBML}
\usage{
rba_reactome_exporter_event(
  event_id,
  output_format,
  save_to = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Exports a given pathway or reaction to SBGN or SBML
}
