% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_overview}
\alias{rba_reactome_exporter_overview}
\title{Get a Reactome Pathway Overview}
\usage{
rba_reactome_exporter_overview(
  species,
  output_format = "png",
  save_to = NULL,
  image_quality = 5,
  flag_element = NULL,
  flg_interactors = TRUE,
  sel = NULL,
  title = TRUE,
  margin = 15,
  diagram_profile = "Copper",
  token = NULL,
  resource = "TOTAL",
  exp_column = NULL,
  coverage = FALSE,
  ...
)
}
\arguments{
\item{species}{Character or Numeric: NCBI Taxonomy identifier (Human Taxonomy
ID is 9606.) or species name (e.g. "Homo sapiens"). See
\code{\link{rba_reactome_species}} or
\href{https://reactome.org/content/schema/objects/Species/}{Reactome
Data Schema: Entries: Species}.}

\item{output_format}{Character: (default = \code{"png"}) Image format, can be
one of: png, jpg, jpeg, svg or gif.}

\item{save_to}{NULL or Character: (default = \code{NULL}) \itemize{
\item NULL: Save the file to an automatically-generated path.
\item Character string: A valid file path to save the file to.}}

\item{image_quality}{Numeric: (default = \code{5}) A number  ranging from 1
to 10. 1 is the lowest quality and 10 is the highest.}

\item{flag_element}{Character: (optional) Gene name, protein ID, chemical ID
or Reactome ID of a diagram's element to be flagged.}

\item{flg_interactors}{Logical: (default = \code{TRUE}) Should the interactor
be considered when flagging a diagram element?}

\item{sel}{Character vector: (optional) CSV line for highlighting element(s)
selection in the diagram.}

\item{title}{Logical: (default = \code{TRUE}) Should the pathway name be
displayed below the image?}

\item{margin}{Numeric: (default = \code{15}) A number ranging from 0 to 20 to
set as the image's margin.}

\item{diagram_profile}{Character: (default = \code{"Copper"}) Color profile
of the overview, should be one of "Copper", "Copper plus", "Barium Lithium"
or "Calcium Salts".}

\item{token}{Character: (optional) The analysis Token for which the results
will be overlaid on top of the given pathways overview. see:
\code{\link{rba_reactome_analysis}}.}

\item{resource}{Character: (default = \code{"TOTAL"}) The analysis resource
for which the results will be overlaid on top of the given pathways
overview.}

\item{exp_column}{Numeric: (optional) (only if token is supplied) Specify the
expression column for the overlay.}

\item{coverage}{Logical: (default = \code{FALSE}) Should the analysis
coverage values be overlaid?}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
NULL, Based to the inputs, an image file will be saved to disk.
}
\description{
This function will save a pathway overview of the supplied species
  as an image file.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/exporter/fireworks/\{species\}.\{ext\}"
}

\examples{
\dontrun{
rba_reactome_exporter_overview(species = 9606,
    output_format = "svg",
    save_to = "human_pathways.svg")
}
\dontrun{
rba_reactome_exporter_overview(species = 9606,
    token = 123456789)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
\code{\link{rba_reactome_analysis}}

Other "Reactome Content Service - Format Exporter":
\code{\link[=rba_reactome_exporter_diagram]{rba_reactome_exporter_diagram()}},
\code{\link[=rba_reactome_exporter_event]{rba_reactome_exporter_event()}},
\code{\link[=rba_reactome_exporter_reaction]{rba_reactome_exporter_reaction()}}
}
\concept{"Reactome Content Service - Format Exporter"}
