% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_overview}
\alias{rba_reactome_exporter_overview}
\title{Exports a given pathway overview of a specie to the specified image format}
\usage{
rba_reactome_exporter_overview(
  species,
  output_format = "png",
  save_to = NA,
  quality = 5,
  flg = NA,
  flg_interactors = TRUE,
  sel = NA,
  title = TRUE,
  margin = 15,
  diagram_profile = "Copper",
  token = NA,
  resource = "TOTAL",
  exp_column = NA,
  coverage = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Exports a given pathway overview of a specie to the specified image format
}
