% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_exporter_reaction}
\alias{rba_reactome_exporter_reaction}
\title{Get a Reactome Reaction Event}
\usage{
rba_reactome_exporter_reaction(
  event_id,
  save_to = NULL,
  output_format = "png",
  resource = "TOTAL",
  diagram_profile = "Modern",
  analysis_profile = "Standard",
  token = NULL,
  exp_column = NULL,
  image_quality = 5,
  flag_element = NULL,
  flg_interactors = TRUE,
  sel = NULL,
  title = TRUE,
  margin = 15,
  ...
)
}
\arguments{
\item{event_id}{Character: Reactome
\href{https://reactome.org/content/schema/ReactionLikeEvent/}{Reaction-like
event}'s identifier.}

\item{save_to}{Character: (optional) \itemize{
\item NULL: Save the file to an automatically-generated path.
\item Character string: A valid file path to save the file to.}}

\item{output_format}{Character: (default = \code{"png"}) Image format, can be
one of: png, jpg, jpeg, svg or gif.}

\item{resource}{Character: (default = \code{"TOTAL"}) The analysis resource
for which the results will be overlaid on top of the given pathways
overview.}

\item{diagram_profile}{Character: (default = \code{"Modern"}) Color profile
of diagrams, should be either "Modern" or "Standard".}

\item{analysis_profile}{Character: (default = \code{"Standard"}) Color
profile of analysis, should be one of: "Standard", "Strosobar" or "Copper
Plus".}

\item{token}{Character: (optional) The analysis Token for which the results
will be overlaid on top of the given pathways overview. see:
\code{\link{rba_reactome_analysis}}.}

\item{exp_column}{Numeric: (optional) (only if token is supplied) Specify the
expression column for the overlay.}

\item{image_quality}{Numeric: (default = \code{5}) A number  ranging from 1
to 10. 1 is the lowest quality and 10 is the highest.}

\item{flag_element}{Character: (optional) Gene name, protein ID, chemical ID
or Reactome ID of a diagram's element to be flagged.}

\item{flg_interactors}{Logical: (default = \code{TRUE}) Should the interactor
be considered when flagging a diagram element?}

\item{sel}{Character: (optional) CSV line for highlighting element(s)
selection in the diagram.}

\item{title}{Logical: (default = \code{TRUE}) Should the pathway name be
displayed below the image?}

\item{margin}{Numeric: (default = \code{15}) A number ranging from 0 to 20 to
set as the image's margin.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
NULL, Based to the inputs, an image file will be saved to disk.
}
\description{
This function will Save a Reactome event of class "ReactionLikeEvent" as
  an image file.
}
\details{
Note that this function will save Reaction-like event separately and out
  of it's parent pathway context. To overlay a Reaction on it's parent
  pathway, use \code{\link{rba_reactome_exporter_diagram}}.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/exporter/reaction/
 \{identifier\}.\{ext\}"
}

\examples{
\dontrun{
rba_reactome_exporter_diagram(event_id = "R-HSA-6787403",
    create_document = FALSE)
}
\dontrun{
rba_reactome_exporter_diagram(event_id = "R-HSA-6787403",
     output_format = "svg",
     save_to = "reactome_reacion_image.svg")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
\code{\link{rba_reactome_exporter_diagram}}
\code{\link{rba_reactome_analysis}}

Other "Reactome Content Service - Format Exporter":
\code{\link[=rba_reactome_exporter_diagram]{rba_reactome_exporter_diagram()}},
\code{\link[=rba_reactome_exporter_event]{rba_reactome_exporter_event()}},
\code{\link[=rba_reactome_exporter_overview]{rba_reactome_exporter_overview()}}
}
\concept{"Reactome Content Service - Format Exporter"}
