% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_interactors_psicquic}
\alias{rba_reactome_interactors_psicquic}
\title{The interface From Reactome to PSICQUIC}
\usage{
rba_reactome_interactors_psicquic(
  proteins = NULL,
  resource = NULL,
  details = TRUE,
  ...
)
}
\arguments{
\item{proteins}{Character or Numeric vector: (optional) Proteins to retrieve
PSICQUIC interactors.}

\item{resource}{Character: (optional) The PSICQUIC resource for your supplied
proteins. Call rba_reactome_interactors_psicquic() without argument to get
the available options.}

\item{details}{Logical: (default = \code{TRUE}) If TRUE a detailed list of
interactors will be returned. If FALSE, only a summary of available
interactors will be returned.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Depending your input, a list containing the detailed or summary of
  PSICQUIC interactions or a data frame of all registered PSICQUIC
  resources.
}
\description{
You can call this function in two scenarios:
  1- To retrieve information of all available PSICQUIC resources, call the
  function without providing any argument; i.e
  rba_reactome_interactors_psicquic().
  2-To retrieve a list of interactors of specific protein(s), fill out the
  function's arguments.
}
\section{Corresponding API Resources}{

 "POST https://reactome.org/ContentService/interactors/psicquic/molecules/
   \{resource\}/details"
 \cr "POST https://reactome.org/ContentService/interactors/psicquic/molecules/
   \{resource\}/summary"
 \cr "GET https://reactome.org/ContentService/interactors/psicquic/resources"
}

\examples{
\donttest{
rba_reactome_interactors_psicquic()
}
\donttest{
rba_reactome_interactors_psicquic(proteins = c("TP53", "MYC"),
    resource = "BioGrid",
    details = FALSE)
}
\donttest{
rba_reactome_interactors_psicquic(proteins = c("TP53", "MYC"),
    resource = "BioGrid",
    details = TRUE)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Molecule Interactors":
\code{\link[=rba_reactome_interactors_static]{rba_reactome_interactors_static()}}
}
\concept{"Reactome Content Service - Molecule Interactors"}
