% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_interactors_psicquic}
\alias{rba_reactome_interactors_psicquic}
\title{Retrieve clustered interaction, sorted by score, of a given accession(s)
by resource.}
\usage{
rba_reactome_interactors_psicquic(resource, proteins, details = "details", ...)
}
\arguments{
\item{proteins}{}
}
\value{

}
\description{
Retrieve clustered interaction, sorted by score, of a given accession(s)
by resource.
}
