% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_interactors_static}
\alias{rba_reactome_interactors_static}
\title{Retrieve a interaction information of a given accession}
\usage{
rba_reactome_interactors_static(proteins, details = "details", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Retrieve a interaction information of a given accession
}
