% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_mapping}
\alias{rba_reactome_mapping}
\title{Map External ID to Reactome Pathways/Reactions}
\usage{
rba_reactome_mapping(id, resource, map_to, species = "Homo sapiens", ...)
}
\arguments{
\item{id}{Character or Numeric: Molecule's external Identifier}

\item{resource}{Character: What is the resource of your supplied ID? see:
\href{https://reactome.org/content/schema/objects/ReferenceDatabase/}{Reactome External
Identifiers}}

\item{map_to}{Character: Either "pathways" or "reactions".}

\item{species}{Character or Numeric: (default = \code{"Homo sapiens"}) NCBI
Taxonomy identifier (Human is 9606), species name (e.g. "Homo sapiens") or
Reactome DbId (e.g Homo sapiens is 48887). See
\code{\link{rba_reactome_species}} or
 \href{https://reactome.org/content/schema/objects/Species/}{Reactome
 Data Schema: Entries: Species}.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Data frame where each row is a pathway/reaction and columns are
  pertinent information.
}
\description{
By providing an external identifier from a given resource, you can retrieve
  a list of pathways/reactions that include your supplied ID.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/mapping/\{resource\}/
 \{identifier\}/pathways"
 \cr "GET https://reactome.org/ContentService/data/mapping/\{resource\}/
 \{identifier\}/reactions"
}

\examples{
\donttest{
rba_reactome_mapping(id = "PTEN", resource =  "UniProt",
    map_to = "reactions", species = 9606)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\concept{"Reactome Content Service - Mapping Related Queries"}
