% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_mapping_pathways}
\alias{rba_reactome_mapping_pathways}
\title{Entities play different roles in reactions, and reactions are events that
conform a pathway. This method retrieves the pathways for which an
identifier plays a role within one or more of their events.}
\usage{
rba_reactome_mapping_pathways(
  id,
  resource,
  species = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Entities play different roles in reactions, and reactions are events that
conform a pathway. This method retrieves the pathways for which an
identifier plays a role within one or more of their events.
}
