% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_mapping_reactions}
\alias{rba_reactome_mapping_reactions}
\title{Entities play different roles in reactions. This method retrieves the
reactions for which an identifier plays a role .}
\usage{
rba_reactome_mapping_reactions(
  id,
  resource,
  species = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Entities play different roles in reactions. This method retrieves the
reactions for which an identifier plays a role .
}
