% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_orthology}
\alias{rba_reactome_orthology}
\title{Get Orthologous (Computationally Inferred) Events}
\usage{
rba_reactome_orthology(event_ids, species_dbid, ...)
}
\arguments{
\item{event_ids}{Character vector: Human Reactome event ID(s) to retrieve
their orthologous events.}

\item{species_dbid}{Numeric: Reactome database ID (DbId) of the target
species. (e.g Mus musculus is 48892). See
\code{\link{rba_reactome_species}} or
 \href{https://reactome.org/content/schema/objects/Species/}{Reactome
 Data Schema: Entries: Species}.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
List containing found Orthologous event(s) in your supplied species
  and their pertinent information.
}
\description{
Reactome incorporate manually curated human reactions and PANTHER's
  protein homology data to Computationally infer events in other eukaryotic
  species.
}
\details{
Reactome uses an orthology-based approach to project curated human events
  to supported non-human species. See
  \href{https://reactome.org/documentation/inferred-events/}{
  Reactome Computationally Inferred Events} for more information.
}
\section{Corresponding API Resources}{

 "POST https://reactome.org/ContentService/data/orthologies/ids/
   species/\{speciesId\}"
}

\examples{
\donttest{
rba_reactome_orthology(event_ids = c("R-HSA-6799198", " R-HSA-72764"),
    species_dbid = 49633)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
\code{\link{rba_reactome_analysis_species}}
}
\concept{"Reactome Content Service - Orthology Related Queries"}
