% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_orthology}
\alias{rba_reactome_orthology}
\title{Reactome uses the set of manually curated human reactions to computationally
infer reactions in twenty evolutionarily divergent eukaryotic species for
which high-quality whole-genome sequence data are available, and hence a
comprehensive and high-quality set of protein predictions exists. Thus,
this method retrieves the orthology for any given event or entity in the
specified species. Here you can find more information about the
computationally inferred events.}
\usage{
rba_reactome_orthology(
  ids,
  species_id,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Reactome uses the set of manually curated human reactions to computationally
infer reactions in twenty evolutionarily divergent eukaryotic species for
which high-quality whole-genome sequence data are available, and hence a
comprehensive and high-quality set of protein predictions exists. Thus,
this method retrieves the orthology for any given event or entity in the
specified species. Here you can find more information about the
computationally inferred events.
}
