% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_participants}
\alias{rba_reactome_participants}
\title{Participants contains a PhysicalEntity (dbId, displayName) and a
collection of ReferenceEntities (dbId, name, identifier, url)}
\usage{
rba_reactome_participants(
  id,
  physical_entities = FALSE,
  reference_entities = FALSE,
  ...
)
}
\arguments{
\item{reference_entities}{}
}
\value{

}
\description{
Participants contains a PhysicalEntity (dbId, displayName) and a
collection of ReferenceEntities (dbId, name, identifier, url)
}
