% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_pathways_events}
\alias{rba_reactome_pathways_events}
\title{Get Events Contained in an Upstream Events}
\usage{
rba_reactome_pathways_events(event_id, attribute_name = NA, ...)
}
\arguments{
\item{event_id}{Reactome event's database ID (DbId) or Stable ID (StId).}

\item{attribute_name}{An attribute of the events to be returned instead of
the whole events. see \href{https://reactome.org/content/schema/Event}{
Reactome Data Schema: Event} for available options.}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
Data frame where each row is a contained event and columns are
  event's attributes. If an "attribute_name" argument was provided, a
  character vector will be returned.
}
\description{
A Reactome Event could be comprised of other events (meaning, a pathway that
  include other pathways itself). Use this function to recursively return
  all the events which reside downstream of your provided event ID (or
  an attribute of that events).
}
\details{
By Reactome's definition, Events are the building blocks of biological
  processes and could be of two main classes: "Pathway" or
  "Reaction-like events". The events are organized in a hierarchical
  structure; and each event could be child or parent to another event; The
  hierarchy will always begin with a "Top level pathway" event. Also note
  that a given event could be part of more that one hierarchies.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/pathway/{id}/
 containedEvents"\cr
 "GET https://reactome.org/ContentService/data/pathway/{id}/
 containedEvents/{attributeName}"
}

\examples{
rba_reactome_pathways_events(event_id = "R-HSA-5673001")
rba_reactome_pathways_events(event_id = "R-HSA-5673001",
  attribute_name = "displayName")
}
\references{
\itemize{
  \item Jassal B, Matthews L, Viteri G, Gong C, Lorente P, Fabregat A,
  Sidiropoulos K, Cook J, Gillespie M, Haw R, Loney F, May B, Milacic M,
  Rothfels K, Sevilla C, Shamovsky V, Shorser S, Varusai T, Weiser J, Wu G,
  Stein L, Hermjakob H, D'Eustachio P. The reactome pathway knowledgebase.
  Nucleic Acids Res. 2020 Jan 8;48(D1):D498-D503. doi: 10.1093/nar/gkz1031.
  PubMed PMID: 31691815.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  }
}
\seealso{
Other "Reactome Content Service, Pathway Related Queries": 
\code{\link{rba_reactome_pathways_low}()},
\code{\link{rba_reactome_pathways_top}()}
}
\concept{"Reactome Content Service, Pathway Related Queries"}
