% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_pathways_low}
\alias{rba_reactome_pathways_low}
\title{This method traverses the event hierarchy and retrieves the list of all
lower level pathways that have a diagram and contain the given
PhysicalEntity or Event.}
\usage{
rba_reactome_pathways_low(
  entity_id,
  with_diagram = FALSE,
  all_forms = FALSE,
  species = NA,
  ...
)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
This method traverses the event hierarchy and retrieves the list of all
lower level pathways that have a diagram and contain the given
PhysicalEntity or Event.
}
