% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_pathways_top}
\alias{rba_reactome_pathways_top}
\title{Get Top Level Pathways in a Species}
\usage{
rba_reactome_pathways_top(species, ...)
}
\arguments{
\item{species}{Character or Numeric: NCBI Taxonomy identifier (Human Taxonomy
ID is 9606.) or species name (e.g. "Homo sapiens"). See
\code{\link{rba_reactome_species}} or
 \href{https://reactome.org/content/schema/objects/Species/}{Reactome
 Data Schema: Entries: Species}.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Data frame where each row is a Top Level Pathway and columns are
  pertinent information.
}
\description{
This function will Return a list of all pathways with the class
  "TopLevelPathway" which are annotated in your supplied species.
}
\details{
Reactome's events hierarchy for any species begins with pathways with
  class "TopLevelPathway" (e.g. "Immune System", "Metabolism of proteins").
  further down in the event's hierarchy tree, each TopLevelPathway has
  has other events itself (e.g. "Adaptive immune system", "Innate immune
  system"). Based on the chosen pathway, the hierarchy tree would typically
  goes further down.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/pathways/top/\{species\}"
}

\examples{
\donttest{
rba_reactome_pathways_top(species = 9606)
}
\donttest{
rba_reactome_pathways_top(species = "Saccharomyces cerevisiae")
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\seealso{
Other "Reactome Content Service - Pathway Related Queries":
\code{\link[=rba_reactome_pathways_events]{rba_reactome_pathways_events()}},
\code{\link[=rba_reactome_pathways_low]{rba_reactome_pathways_low()}}
}
\concept{"Reactome Content Service - Pathway Related Queries"}
