% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_people_id}
\alias{rba_reactome_people_id}
\title{A person by his identifiers}
\usage{
rba_reactome_people_id(
  id,
  authored_pathways = FALSE,
  publications = FALSE,
  attribute_name = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
A person by his identifiers
}
