% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_people_name}
\alias{rba_reactome_people_name}
\title{Retrieves a list of people in Reactome with either their first or last name
partly/exactly matching the given string.}
\usage{
rba_reactome_people_name(person_name, exact = FALSE, ...)
}
\arguments{
\item{exact}{}
}
\value{

}
\description{
Retrieves a list of people in Reactome with either their first or last name
partly/exactly matching the given string.
}
