% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_query}
\alias{rba_reactome_query}
\title{all in one: query Reactome Data: Common data retrieval}
\usage{
rba_reactome_query(
  ids,
  enhanced = FALSE,
  map = FALSE,
  verbose = TRUE,
  attribute_name = NA,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
all in one: query Reactome Data: Common data retrieval
}
