% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_species}
\alias{rba_reactome_species}
\title{Get Reactome Species}
\usage{
rba_reactome_species(only_main = FALSE, ...)
}
\arguments{
\item{only_main}{Logical: (default = \code{FALSE}) If set to TRUE, will only
return species which have either manually-curated or computationally
inferred pathways.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Data frame where each row is a species and columns are pertinent
  information.
}
\description{
Use this function to retrieve a table of Available species in Reactome.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/species/all"
 \cr "GET https://reactome.org/ContentService/data/species/main"
}

\examples{
\donttest{
rba_reactome_species()
}
\donttest{
rba_reactome_species(only_main = TRUE)
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss, J., Viteri, G., Sidiropoulos, K., Nguyen, V., Fabregat, A.,
  & Hermjakob, H. (2020). ReactomeGSA—Efficient Multi-Omics Comparative
  Pathway Analysis. Molecular & Cellular Proteomics, 19(12), 2115–2125.
  doi: 10.1074/mcp.TIR120.002155
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\concept{"Reactome Content Service - Species Related Queries"}
