% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_species}
\alias{rba_reactome_species}
\title{This method retrieves the list of main species in Reactome knowledgebase,
sorted by name, but having ‘Homo sapiens’ as the first one. It should be
mentioned that for Reactome, main species are considered those have either
manually curated or computationally inferred pathways.}
\usage{
rba_reactome_species(
  species_type = "all",
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
This method retrieves the list of main species in Reactome knowledgebase,
sorted by name, but having ‘Homo sapiens’ as the first one. It should be
mentioned that for Reactome, main species are considered those have either
manually curated or computationally inferred pathways.
}
