% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_version}
\alias{rba_reactome_version}
\title{The version number of current database}
\usage{
rba_reactome_version(...)
}
\arguments{
\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
Character string containing the version of Reactome database.
}
\description{
Returns the current version of Reactome database.
}
\section{Corresponding API Resources}{

 "GET https://reactome.org/ContentService/data/database/version"
}

\examples{
\donttest{
rba_reactome_version()
}

}
\references{
\itemize{
  \item Ragueneau, E., Gong, C., Sinquin, P., Sevilla, C., Beavers, D.,
  Grentner, A., ... D’Eustachio, P. (2026). The Reactome Knowledgebase 2026.
  Nucleic Acids Res., 54(D1), D673–D681. doi: 10.1093/nar/gkaf1223
  \item Griss J, Viteri G, Sidiropoulos K, Nguyen V, Fabregat A,
  Hermjakob H. ReactomeGSA - Efficient Multi-Omics Comparative Pathway
  Analysis. Mol Cell Proteomics. 2020 Sep 9. doi: 10.1074/mcp. PubMed
  PMID: 32907876.
  \item \href{https://reactome.org/ContentService/}{Reactome Content
  Services API Documentation}
  \item \href{https://reactome.org/cite/}{Citations note on Reactome website}
  }
}
\concept{"Reactome Content Service - Database Info Queries"}
