% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_content.R
\name{rba_reactome_version}
\alias{rba_reactome_version}
\title{The version number of current database}
\usage{
rba_reactome_version(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
The version number of current database
}
