% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_annotations}
\alias{rba_string_annotations}
\title{Get Functional Annotations}
\usage{
rba_string_annotations(
  ids,
  species = NULL,
  allow_pubmed = FALSE,
  split_df = TRUE,
  ...
)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.
(Recommended, but optional.)}

\item{allow_pubmed}{Logical (default = \code{FALSE}): Include PubMed
annotations. These annotations are excluded by default because many
publications may be assigned to each protein.}

\item{split_df}{Logical: (default = \code{TRUE}) Split results into a list
of data frames by \code{category}; otherwise, return one data frame.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame in which every row is an assigned term and the columns
  contain the term category, description, number of genes, and other
  pertinent information. If \code{split_df = TRUE}, a list of data frames
  split by category is returned.
}
\description{
STRING cross-references proteins with several annotation resources. This
  function retrieves the complete set of annotations assigned to the input
  proteins, together with information about each term.
}
\details{
STRING currently retrieves annotations based on Gene Ontology (GO), UniProt
  Keywords, PubMed publications, Pfam domains, InterPro domains, and SMART
  domains. KEGG annotations are unavailable from this endpoint because of
  KEGG licensing restrictions.
  \cr This function returns annotations without enrichment filtering. To
  perform enrichment and retrieve only enriched terms, use
  \code{\link{rba_string_enrichment}}.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/functional_annotation?
 identifiers=\{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_annotations(ids = "TP53", species = 9606)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{\link{rba_string_map_ids},
  \link{rba_string_enrichment},
  \link{rba_string_enrichment_image}
  }

Other "STRING":
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}
}
\concept{"STRING"}
