% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_enrichment}
\alias{rba_string_enrichment}
\title{Get Functional Enrichment}
\usage{
rba_string_enrichment(
  ids,
  species = NULL,
  background = NULL,
  split_df = TRUE,
  ...
)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.
Note that if only one ID is supplied, STRING expands the network by 10
proteins.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.
(Recommended, but optional.)}

\item{background}{Character vector: A set of STRING protein IDs
to be used as the statistical background (or universe) when computing
term p-values. Only STRING IDs are accepted. See
\code{\link{rba_string_map_ids}} to map your IDs.}

\item{split_df}{Logical: (default = \code{TRUE}) Split results into a list
of data frames by \code{category}; otherwise, return one data frame.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame in which each row is an enriched term with a raw p-value
  below 0.1 and the columns contain the term category, description, gene
  counts, p-value, FDR, and other pertinent information. If
  \code{split_df = TRUE}, a list of data frames split by category is returned.
}
\description{
STRING cross-references proteins with several annotation resources. (See
  'Details' section). Provide an input protein set and, optionally, a
  background protein set to perform an enrichment test and retrieve
  enriched terms with their associated statistics.
  Use \code{\link{rba_string_enrichment_image}} to retrieve the analysis
  results as a plot.
}
\details{
STRING currently returns enrichment results from Gene
  Ontology (GO), KEGG pathways, UniProt Keywords, PubMed publications, Pfam
  domains, InterPro domains, and SMART domains.
  \cr STRING returns only terms with a raw p-value below 0.1. To retrieve
  annotations without filtering by enrichment p-value, use
  \code{\link{rba_string_annotations}}.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/enrichment?identifiers=
 \{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_enrichment(ids = c("TP53", "TNF", "EGFR"), species = 9606)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{\link{rba_string_map_ids},
  \link{rba_string_annotations},
  \link{rba_string_enrichment_image}
  }

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}

Other "Enrichment/Over-representation":
\code{\link[=rba_enrichr]{rba_enrichr()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}}
}
\concept{"Enrichment/Over-representation"}
\concept{"STRING"}
