% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_enrichment}
\alias{rba_string_enrichment}
\title{Getting Functional Enrichment}
\usage{
rba_string_enrichment(
  ids,
  species = NULL,
  background = NULL,
  split_df = TRUE,
  ...
)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.
Note that if only one id is supplied, STRING expands the network by 10
proteins.}

\item{species}{Numeric: NCBI Taxonomy identifier; Human Taxonomy ID is 9606.
(Recommended, but optional if your input is less than 100 IDs.)}

\item{background}{character vector: A set of STRING protein IDs
to be used as the statistical background (or universe) when computing
P-value for the terms. Only STRING IDs are acceptable. (See
\code{\link{rba_string_map_ids}} to map your IDs.)}

\item{split_df}{(logical, default = TRUE), If TRUE, instead of one
data frame, results from different categories will be split into
multiple data frames based on their 'category'.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list of data frames which every row is an enriched terms with p-value
  smaller than 0.1 and the columns are the terms category, description,
  number of genes, p-value, fdr and other pertinent information.
}
\description{
STRING cross-reference the proteins with several databases (see "Details"
  section). By providing your input set o proteins (and optionally
  background or universe protein set), you can use this function to
  perform enrichment test and retrieve a list of enriched terms in each
  database, among with pertinent information for each term.
  Use \code{\link{rba_string_enrichment_image}} to retrieve the analysis
  results as a plot.
}
\details{
STRING currently maps to and retrieve enrichment results based on Gene
  Ontology (GO), KEGG pathways, UniProt Keywords, PubMed publications, Pfam
  domains, InterPro domains, and SMART domains.
  \cr Note that this function will only return the enriched terms pertinent
  to your proteins that have a p-value lesser than 0.1. To retrieve a full
  list of the terms -unfiltered by enrichment p-values-, use
  \code{\link{rba_string_annotations}}.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output_format\}/enrichment?identifiers=
 \{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_enrichment(ids = c("TP53", "TNF", "EGFR"), species = 9606)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{\link{rba_string_map_ids},
  \link{rba_string_annotations},
  \link{rba_string_enrichment_image}
  }

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}

Other "Enrichment/Over-representation":
\code{\link[=rba_enrichr]{rba_enrichr()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}}
}
\concept{"Enrichment/Over-representation"}
\concept{"STRING"}
