% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_enrichment}
\alias{rba_string_enrichment}
\title{Getting functional enrichment}
\usage{
rba_string_enrichment(
  input,
  species = NA,
  background_string_ids = NA,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{background_string_ids}{}
}
\value{

}
\description{
#' @family STRING
}
