% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_enrichment_image}
\alias{rba_string_enrichment_image}
\title{Get STRING Enrichment Plot}
\usage{
rba_string_enrichment_image(
  ids,
  species = NULL,
  category = "Process",
  image_format = "image",
  save_image = TRUE,
  group_by_similarity = NULL,
  color_palette = "mint_blue",
  number_of_term_shown = 10,
  x_axis = "signal",
  ...
)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.
If only one ID is supplied, STRING expands the network by 10
proteins.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.
(Recommended, but required if your input contains more than 10 unique IDs.)}

\item{category}{Character: The term set to use for enrichment analysis.
Valid values are:
"Process" (default), "Function", "Component", "Keyword", "KEGG", "RCTM",
"HPO", "MPO", "DPO", "WPO", "ZPO", "FYPO", "Pfam", "SMART", "InterPro",
"PMID", "NetworkNeighborAL", "COMPARTMENTS", "TISSUES", "DISEASES", or
"WikiPathways". See Details for descriptions.}

\item{image_format}{Character: One of:\itemize{
\item "image": PNG image with normal resolution.
\item "highres_image": High-resolution PNG image.
\item "svg": Scalable Vector Graphics image.}}

\item{save_image}{Logical or Character:\itemize{
\item TRUE: Save the image to an automatically-generated path.
\item FALSE: Do not save the image, just return it as an R object.
\item Character string: A valid file path to save the image to.}}

\item{group_by_similarity}{Numeric: Jaccard-index threshold used to group
related terms visually. Valid values range from 0.1 to 1 in increments of
0.1. The default is \code{NULL}, which disables grouping.}

\item{color_palette}{Character: Color palette used to represent FDR values.
Valid values are
"mint_blue" (default), "lime_emerald", "green_blue", "peach_purple",
"straw_navy", or "yellow_pink".}

\item{number_of_term_shown}{Numeric: (default = 10) Maximum number of terms
to include in the plot.}

\item{x_axis}{Character: Variable displayed on the x-axis and used to rank
the results. Valid values are "signal" (default), "strength", "FDR", and
"gene_count".}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A PNG image array or raw SVG content, depending on
  \code{image_format}.
}
\description{
Retrieve a plot that summarizes STRING functional-enrichment results. Use
  \code{\link{rba_string_enrichment}} to retrieve the results as a data
  frame.
}
\details{
Available \code{category} values are listed below. The default is "Process".
  \itemize{
  \item Process: Biological Process (Gene Ontology)
  \item Function: Molecular Function (Gene Ontology)
  \item Component: Cellular Component (Gene Ontology)
  \item Keyword: Annotated Keywords (UniProt)
  \item KEGG: KEGG Pathways
  \item RCTM: Reactome Pathways
  \item HPO: Human Phenotype (Monarch)
  \item MPO: The Mammalian Phenotype Ontology (Monarch)
  \item DPO: Drosophila Phenotype (Monarch)
  \item WPO: C. elegans Phenotype Ontology (Monarch)
  \item ZPO: Zebrafish Phenotype Ontology (Monarch)
  \item FYPO: Fission Yeast Phenotype Ontology (Monarch)
  \item Pfam: Protein Domains (Pfam)
  \item SMART: Protein Domains (SMART)
  \item InterPro: Protein Domains and Features (InterPro)
  \item PMID: Reference Publications (PubMed)
  \item NetworkNeighborAL: Local Network Cluster (STRING)
  \item COMPARTMENTS: Subcellular Localization (COMPARTMENTS)
  \item TISSUES: Tissue Expression (TISSUES)
  \item DISEASES: Disease-gene Associations (DISEASES)
  \item WikiPathways: WikiPathways}
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/enrichmentfigure?
 identifiers=\{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\dontrun{
  rba_string_enrichment_image(
  ids = c("TP53", "TNF", "EGFR"),
  species = 9606,
  category = "KEGG"
  )
}
\dontrun{
  rba_string_enrichment_image(
  ids = c("TP53", "TNF", "EGFR"),
  species = 9606,
  x_axis = "strength",
  number_of_term_shown = 20
  )
}
\dontrun{
  rba_string_enrichment_image(
  ids = c("TP53", "TNF", "EGFR"),
  species = 9606,
  color_palette = "straw_navy"
  )
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{
  \link{rba_string_map_ids},
  \link{rba_string_enrichment},
  \link{rba_string_annotations},
  \link{rba_string_functional_terms}
  }

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_functional_terms]{rba_string_functional_terms()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}

Other "Enrichment/Over-representation":
\code{\link[=rba_enrichr]{rba_enrichr()}},
\code{\link[=rba_mieaa_enrich]{rba_mieaa_enrich()}},
\code{\link[=rba_panther_enrich]{rba_panther_enrich()}},
\code{\link[=rba_reactome_analysis]{rba_reactome_analysis()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}}
}
\concept{"Enrichment/Over-representation"}
\concept{"STRING"}
