% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_enrichment_ppi}
\alias{rba_string_enrichment_ppi}
\title{Get Protein-Protein Interaction Enrichment}
\usage{
rba_string_enrichment_ppi(
  ids,
  species = NULL,
  required_score = NULL,
  background = NULL,
  ...
)
}
\arguments{
\item{ids}{Character or Numeric: Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.}

\item{species}{Numeric: (optional)
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.
(Recommended, but required if your input contains more than 10 unique IDs.)}

\item{required_score}{Numeric: (optional) (between 0 and 1000): Minimum interaction
score used when calculating PPI enrichment. If omitted, STRING applies a
network-dependent threshold. Common
confidence thresholds are 150 (low), 400 (medium), 700 (high), and 900
(highest).}

\item{background}{Character: (optional) A set of STRING protein IDs
to be used as the background proteome. Only STRING IDs are acceptable.
See \code{\link{rba_string_map_ids}} to map your IDs.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list with protein-protein interaction enrichment results.
}
\description{
STRING compares the interaction pattern of your input proteins with the
  proteome-wide background interaction distribution to determine whether
  the protein set contains more interactions than expected.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/ppi_enrichment?identifiers=
 \{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_enrichment_ppi(ids = c("p53", "BRCA1", "cdk2", "Q99835",
       "CDC42", "CDK1", "KIF23", "PLK1", "RAC2", "RACGAP1"),
    species = 9606)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{
  \link{rba_string_map_ids},
  \link{rba_string_interactions_network},
  \link{rba_string_interaction_partners},
  \link{rba_string_network_image}
  }

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_functional_terms]{rba_string_functional_terms()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}
}
\concept{"STRING"}
