% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_functional_annotation}
\alias{rba_string_functional_annotation}
\title{Retrieving functional annotation}
\usage{
rba_string_functional_annotation(
  input,
  species = NA,
  allow_pubmed = FALSE,
  caller_identity = NA,
  ...
)
}
\arguments{
\item{caller_identity}{}
}
\value{

}
\description{
Retrieving functional annotation
}
