% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_homology}
\alias{rba_string_homology}
\title{Retrieving similarity scores of the protein set}
\usage{
rba_string_homology(input, species = NA, caller_identity = NA, ...)
}
\arguments{
\item{caller_identity}{}
}
\value{

}
\description{
#' @family STRING
}
