% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_homology_inter}
\alias{rba_string_homology_inter}
\title{Get Best Protein Similarity Hits Across Species}
\usage{
rba_string_homology_inter(ids, species = NULL, species_b = NULL, ...)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier} of your input proteins; Human Taxonomy ID is
9606. (Recommended, but required if your input contains more than 10
unique IDs.)}

\item{species_b}{Numeric: One or more
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifiers} used to restrict the search for closest
homologs. The default is \code{NULL}, which searches all STRING species.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame containing each input protein and its closest homolog
  in every other STRING species, or in the species selected by
  \code{species_b}.
}
\description{
Retrieve the highest Smith-Waterman bit-score hit between each input protein
  and proteins in every other STRING species. STRING uses these
  sequence-similarity scores as a proxy for protein homology.
}
\details{
To retrieve pairwise similarity scores among input proteins within one
  species, see \code{\link{rba_string_homology_intra}}.

STRING imports the similarity matrix from the
  \href{https://pubmed.ncbi.nlm.nih.gov/24165881/}{Similarity Matrix of
  Proteins (SIMAP)} project.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/homology_best?
 identifiers=\{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_homology_inter(ids = "p53",
    species = 9606,
    species_b = 7070)
}
\donttest{
rba_string_homology_inter(ids = "ENSP00000269305", species = 9606)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{
  \link{rba_string_map_ids},
  \link{rba_string_homology_intra}
  }

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_functional_terms]{rba_string_functional_terms()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interaction_partners]{rba_string_interaction_partners()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}
}
\concept{"STRING"}
