% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_homology_inter}
\alias{rba_string_homology_inter}
\title{Get Similarity Scores Hits of Proteins in Different Species}
\usage{
rba_string_homology_inter(ids, species = NA, species_b = NA, ...)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to provide
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.}

\item{species}{Numeric: NCBI Taxonomy identifier of your input proteins;
Human Taxonomy ID is 9606. (Recommended, but optional if your input is
less than 100 IDs.)}

\item{species_b}{(optional) Numeric: one or more NCBI Taxonomy identifiers
of species to limit the closets homologous proteins search.}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A data frame with Your input proteins and it's closest homologous
  proteins among all other (or a defined) STRING species.
}
\description{
Using this function, you can retrieve highest Smith-Waterman bit scores
  among your input proteins and proteins in every other STRING species
  (e.g. the closest homologous protein of your input protein in other
  species).
  Bit Scores serve as similarity scores between protein sequence;
  And, according to STRING documentations, as a proxy for protein homology.
}
\details{
Note that this function will return the highest similarity score hits of
  your given protein(s) and their closets homologous proteins in other
  species. to retrieve similarity scores of different proteins within the
  same species refer to \code{\link{rba_string_homology_intra}}.\cr
  Similarity matrix is imported -by STRING- from:
  \href{http://cube.univie.ac.at/resources/simap}{Similarity Matrix of
  Proteins (SIMAP)}
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/[output-format]/homology_best?
 identifiers=[your_identifiers]"
}

\examples{
\donttest{
rba_string_homology_inter(ids = "p53",
    species = 9606,
    species_b = c(6087, 7070))
}
\donttest{
rba_string_homology_inter(ids = "ENSP00000269305", species = 9606)
}

}
\references{
\itemize{
  \item Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J,
  Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, Mering CV.
  STRING v11: protein-protein association networks with increased coverage,
  supporting functional discovery in genome-wide experimental datasets.
  Nucleic Acids Res. 2019 Jan 8;47(D1):D607-D613. doi: 10.1093/nar/gky1131.
  PMID: 30476243; PMCID: PMC6323986.
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  }
}
\seealso{
\code{\link{rba_string_map_ids}, \link{rba_string_homology_intra}}

Other "STRING API": 
\code{\link{rba_string_annotations}()},
\code{\link{rba_string_enrichment_ppi}()},
\code{\link{rba_string_enrichment}()},
\code{\link{rba_string_homology_intra}()},
\code{\link{rba_string_interaction_partners}()},
\code{\link{rba_string_interactions_network}()},
\code{\link{rba_string_map_ids}()},
\code{\link{rba_string_network_image}()},
\code{\link{rba_string_version}()}
}
\concept{"STRING API"}
