% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_homology_intra}
\alias{rba_string_homology_intra}
\title{Get Similarity Scores Hits of Proteins in a Species}
\usage{
rba_string_homology_intra(ids, species = NA, ...)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to provide
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.}

\item{species}{Numeric: NCBI Taxonomy identifier; Human Taxonomy ID is 9606.
(Recommended, but optional if your input is less than 100 IDs.)}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s
arguments documentation for more information on available options.}
}
\value{
A data frame with bit scores between your provided proteins and
  their self-hit. To Reduce the transferred data, STRING returns only one
  half of the similarity matrix; This will not pose a problem because
  similarity matrix is symmetrical.
}
\description{
Using this function, you can retrieve the Smith-Waterman bit scores among
  proteins of the same species.
  Bit Scores serve as similarity scores between protein sequence;
  And, according to STRING documentations, as a proxy for protein homology.
}
\details{
Note that this function will retrieve similarity scores of different
  proteins "within the same species". To Get a similarity scores of a given
  protein and it's closets homologous proteins in other species, refer to
  \code{\link{rba_string_homology_inter}}.
  \cr Similarity matrix is imported -by STRING- from:
  \href{https://cube.univie.ac.at/resources/simap}{Similarity Matrix of
  Proteins (SIMAP)}
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/[output-format]/homology?identifiers=
 [your_identifiers]"
}

\examples{
\donttest{
rba_string_homology_intra(ids = c("CDK1", "CDK2"), species = 9606)
}

}
\references{
\itemize{
  \item Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J,
  Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, Mering CV.
  STRING v11: protein-protein association networks with increased coverage,
  supporting functional discovery in genome-wide experimental datasets.
  Nucleic Acids Res. 2019 Jan 8;47(D1):D607-D613. doi: 10.1093/nar/gky1131.
  PMID: 30476243; PMCID: PMC6323986.
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  }
}
\seealso{
\code{\link{rba_string_map_ids}, \link{rba_string_homology_inter}}

Other "STRING": 
\code{\link{rba_string_annotations}()},
\code{\link{rba_string_enrichment_ppi}()},
\code{\link{rba_string_enrichment}()},
\code{\link{rba_string_homology_inter}()},
\code{\link{rba_string_interaction_partners}()},
\code{\link{rba_string_interactions_network}()},
\code{\link{rba_string_map_ids}()},
\code{\link{rba_string_network_image}()},
\code{\link{rba_string_version}()}
}
\concept{"STRING"}
