% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_info}
\alias{rba_string_info}
\title{Get Statistics from STRING database}
\usage{
rba_string_info(type = "statistics", verbose = TRUE, diagnostics = FALSE)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Get Statistics from STRING database
}
