% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_interaction_partners}
\alias{rba_string_interaction_partners}
\title{Get All STRING Interaction Partners}
\usage{
rba_string_interaction_partners(
  ids,
  species = NULL,
  required_score = NULL,
  network_type = "functional",
  limit = NULL,
  ...
)
}
\arguments{
\item{ids}{Your protein ID(s). It is strongly recommended to supply
STRING IDs. See \code{\link{rba_string_map_ids}} for more information.}

\item{species}{Numeric:
\href{https://www.ncbi.nlm.nih.gov/taxonomy/}{
NCBI Taxonomy identifier}; Human Taxonomy ID is 9606.
(Recommended, but required if your input contains more than 10 unique IDs.)}

\item{required_score}{Numeric (between 0 and 1000): Minimum interaction
score required for an interaction to be included in the returned
interactions. If omitted, STRING applies a network-dependent threshold.
Common confidence thresholds are 150 (low), 400 (medium), 700 (high), and
900 (highest).}

\item{network_type}{Character: One of:\itemize{
\item "functional": (default) Edges indicate both physical and
functional associations.
\item "physical": Edges indicate that two proteins have a physical
interaction or are parts of a complex.}}

\item{limit}{Numeric: Maximum number of interaction partners returned for
each input protein, ordered by confidence.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A data frame in which each row is a network interaction and the
  columns contain interactor information and interaction scores.
}
\description{
This function retrieves STRING interactions involving any of your input
  proteins as one party of the interaction, including interactions with
  proteins outside the input set.
  \cr Given the size of the STRING database, this function can return many
  interactions. Use the filtering arguments to limit the results.
}
\details{
To retrieve only interactions among the input proteins, see
  \code{\link{rba_string_interactions_network}}.
}
\section{Corresponding API Resources}{

 "POST https://string-db.org/api/\{output-format\}/interaction_partners?
 identifiers=\{your_identifiers\}&\{optional_parameters\}"
}

\examples{
\donttest{
rba_string_interaction_partners(ids = c("9606.ENSP00000269305",
    "9606.ENSP00000398698",
    "9606.ENSP00000275493"),
    network_type = "functional")
}
\donttest{
    rba_string_interaction_partners(ids = "9606.ENSP00000269305",
    species = 9606,
    required_score = 700)
}

}
\references{
\itemize{
  \item Damian Szklarczyk, Rebecca Kirsch, Mikaela Koutrouli, Katerina
   Nastou, Farrokh Mehryary, Radja Hachilif, Annika L Gable, Tao Fang,
   Nadezhda T Doncheva, Sampo Pyysalo, Peer Bork, Lars J Jensen, Christian
   von Mering, The STRING database in 2023: protein–protein association
   networks and functional enrichment analyses for any sequenced genome of
   interest, Nucleic Acids Research, Volume 51, Issue D1, 6 January 2023,
   Pages D638–D646, https://doi.org/10.1093/nar/gkac1000
  \item \href{https://string-db.org/help/api/}{STRING API Documentation}
  \item
  \href{https://string-db.org/cgi/about?footer_active_subpage=references}{
  Citations note on STRING website}
  }
}
\seealso{
\code{\link{rba_string_map_ids}, \link{rba_string_interactions_network}}

Other "STRING":
\code{\link[=rba_string_annotations]{rba_string_annotations()}},
\code{\link[=rba_string_enrichment]{rba_string_enrichment()}},
\code{\link[=rba_string_enrichment_image]{rba_string_enrichment_image()}},
\code{\link[=rba_string_enrichment_ppi]{rba_string_enrichment_ppi()}},
\code{\link[=rba_string_homology_inter]{rba_string_homology_inter()}},
\code{\link[=rba_string_homology_intra]{rba_string_homology_intra()}},
\code{\link[=rba_string_interactions_network]{rba_string_interactions_network()}},
\code{\link[=rba_string_map_ids]{rba_string_map_ids()}},
\code{\link[=rba_string_network_image]{rba_string_network_image()}},
\code{\link[=rba_string_version]{rba_string_version()}}
}
\concept{"STRING"}
