% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_interaction_partners}
\alias{rba_string_interaction_partners}
\title{Getting all the STRING interaction partners of the protein set}
\usage{
rba_string_interaction_partners(
  input,
  species = NA,
  required_score = NA,
  limit = NA,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Getting all the STRING interaction partners of the protein set
}
