% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_map_ids}
\alias{rba_string_map_ids}
\title{Map a Set of Identifiers to String Identifiers}
\usage{
rba_string_map_ids(
  input,
  species = NA,
  echo_query = FALSE,
  limit = 1,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{input}{A character or/and numeric vector containing your desired Identifiers to be mapped}

\item{species}{(optional) Numeric, NCBI taxon identifiers (e.g. Human is 9606, see:
STRING organisms).}

\item{echo_query}{(optional) Logical, Insert column with your input identifier. (default = FALSE)}

\item{limit}{(optional) limits the number of matches per query identifier (best matches
come first)}

\item{caller_identity}{(optional) Character, your identifier for STRING servers.}

\item{verbose}{Should the function print details? (default = FALSE)}

\item{progress_bar}{Logical, Display Progress bar? (default = FALSE)}

\item{diagnostics}{}
}
\value{

}
\description{
This function Calls String's API to Convert a set of identifiers
  to String Identifiers. Although You can call STRING with a variety of
  Identifiers, It is recommended by STRING's documentations that you first
  map Your Protein/genes Identifiers to STRING Identifiers and then proceed
  with other STRING API functions. This  function's Documentation is based
  on: \url{https://string-db.org/help/api/#mapping-identifiers}
}
\references{
\url{https://string-db.org/help/api/#mapping-identifiers}
}
\concept{STRING}
