% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_network_image}
\alias{rba_string_network_image}
\title{Getting STRING network image}
\usage{
rba_string_network_image(
  input,
  output_format = "image",
  save_to = NA,
  species = NA,
  add_color_nodes = FALSE,
  add_white_nodes = NA,
  required_score = NA,
  network_flavor = "confidence",
  hide_node_labels = FALSE,
  hide_disconnected_nodes = FALSE,
  block_structure_pics_in_bubbles = FALSE,
  caller_identity = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
#' @family STRING
}
