% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_network_interactions}
\alias{rba_string_network_interactions}
\title{Getting the STRING network interactions}
\usage{
rba_string_network_interactions(
  input,
  species = NA,
  required_score = NA,
  add_nodes = NA,
  caller_identity = NA,
  ...
)
}
\arguments{
\item{caller_identity}{}
}
\value{

}
\description{
Getting the STRING network interactions
}
