% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{rba_string_version}
\alias{rba_string_version}
\title{Getting current STRING version}
\usage{
rba_string_version(verbose = TRUE, progress_bar = FALSE, diagnostics = FALSE)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
#' @family STRING
}
