% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_antigen_search}
\alias{rba_uniprot_antigen_search}
\title{Search antigens in UniProt}
\usage{
rba_uniprot_antigen_search(
  accession = NA,
  antigen_sequence = NA,
  antigen_id = NA,
  ensembl_id = NA,
  match_score = NA,
  ...
)
}
\arguments{
\item{accession}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search antigens in UniProt
}
