% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_antigens_search}
\alias{rba_uniprot_antigens_search}
\title{Search Antigens in UniProt}
\usage{
rba_uniprot_antigens_search(
  accession = NULL,
  antigen_sequence = NULL,
  antigen_id = NULL,
  ensembl_id = NULL,
  match_score = NULL,
  ...
)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{antigen_sequence}{Character: (optional) A single antigenic protein
sequence of at least four residues.}

\item{antigen_id}{Character: (optional) Human Protein Atlas (HPA) antigen ID.
You
can supply up to
20 IDs.}

\item{ensembl_id}{Character: (optional) Ensembl stable transcript ID. You can
supply up to
20 IDs.}

\item{match_score}{Numeric: (optional) A whole number from 0 to 100 giving
the minimum
alignment score between the antigen sequence and target protein sequence.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list in which each element represents a matching UniProt entry,
  named by accession when available. Antigenic annotations are stored in the
  entry's \code{features} element.
}
\description{
UniProt maps antigenic (antibody-binding) features from several sources to
  protein sequences. Search those mappings using one or more criteria. At
  least one of \code{accession}, \code{antigen_sequence}, \code{antigen_id},
  \code{ensembl_id}, or \code{match_score} is required.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/antigen"
}

\examples{
\donttest{
rba_uniprot_antigens_search(antigen_id = "HPA001060")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Antigen":
\code{\link[=rba_uniprot_antigens]{rba_uniprot_antigens()}}
}
\concept{"UniProt - Antigen"}
