% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates}
\alias{rba_uniprot_coordinates}
\title{Get Genomic Coordinates of a Protein}
\usage{
rba_uniprot_coordinates(accession = NULL, db_type = NULL, db_id = NULL, ...)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}.}

\item{db_type}{Character: (optional) Cross-reference database name. One of
"Ensembl", "CCDS", "HGNC", or "RefSeq".}

\item{db_id}{Character: (optional) Protein identifier in the
cross-reference database.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the requested protein's genomic coordinates.
}
\description{
Retrieve \href{https://www.uniprot.org/help/genomic_coordinates}{genomic
  coordinates} for a protein using either its UniProt accession or its ID
  in a cross-reference database (Ensembl, CCDS, HGNC, or RefSeq). You
  should supply either \code{accession} alone or \code{db_type} and
  \code{db_id} together.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/coordinates/\{accession\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/coordinates/\{dbtype\}:\{dbid\}"
}

\examples{
\donttest{
rba_uniprot_coordinates(accession = "P25942")
}
\donttest{
rba_uniprot_coordinates(db_type = "HGNC", db_id = "CD40")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item McGarvey, P. B., Nightingale, A., Luo, J., Huang, H., Martin,
  M. J., Wu, C., & The UniProt Consortium. (2019). UniProt genomic
  mapping for deciphering functional effects of missense variants.
  Human Mutation, 40(6), 694–705.
  https://doi.org/10.1002/humu.23738
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Coordinates":
\code{\link[=rba_uniprot_coordinates_location]{rba_uniprot_coordinates_location()}},
\code{\link[=rba_uniprot_coordinates_location_genome]{rba_uniprot_coordinates_location_genome()}},
\code{\link[=rba_uniprot_coordinates_location_protein]{rba_uniprot_coordinates_location_protein()}},
\code{\link[=rba_uniprot_coordinates_search]{rba_uniprot_coordinates_search()}}
}
\concept{"UniProt - Coordinates"}
