% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates}
\alias{rba_uniprot_coordinates}
\title{Get genomic coordinates for a UniProt accession Search UniProt entries by
genomic database cross reference IDs: Ensembl, CCDS, HGNC or RefSeq}
\usage{
rba_uniprot_coordinates(accession = NA, db_type = NA, db_id = NA, ...)
}
\arguments{
\item{db_id}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get genomic coordinates for a UniProt accession Search UniProt entries by
genomic database cross reference IDs: Ensembl, CCDS, HGNC or RefSeq
}
