% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates_location}
\alias{rba_uniprot_coordinates_location}
\title{Get genome coordinate by protein sequence position/ position range}
\usage{
rba_uniprot_coordinates_location(
  accession,
  p_position = NA,
  p_start = NA,
  p_end = NA,
  ...
)
}
\arguments{
\item{p_end}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get genome coordinate by protein sequence position/ position range
}
