% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates_search}
\alias{rba_uniprot_coordinates_search}
\title{Search Genomic Coordinates of UniProt entries}
\usage{
rba_uniprot_coordinates_search(
  accession = NULL,
  chromosome = NULL,
  ensembl_id = NULL,
  gene = NULL,
  protein = NULL,
  taxid = NULL,
  location = NULL,
  ...
)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{chromosome}{Character or Numeric: (optional) Chromosome name, such as
"X", "Y", 1, or 20. You can supply up to 20 values.}

\item{ensembl_id}{Character: (optional) Ensembl stable gene ID, transcript
ID, or translation ID. You can supply up to 20 IDs.}

\item{gene}{Character: (optional)
\href{https://www.uniprot.org/help/gene_name}{UniProt gene
name(s)}. You can supply up to 20 gene names.}

\item{protein}{Character: (optional)
\href{https://www.uniprot.org/help/protein_names}{UniProt protein name}.}

\item{taxid}{Numeric: (optional) NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID}.
You can supply up to 20 taxon IDs.}

\item{location}{Character: (optional) Genome location range, such as
"58205437-58219305".}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list named by accession. Each element contains one matching
  protein's genomic-coordinate information.
}
\description{
Search \href{https://www.uniprot.org/help/genomic_coordinates}{genomic
  coordinates} associated with UniProt entries by accession, chromosome,
  Ensembl identifier, gene, protein, taxonomy, or genomic range.
}
\details{
At least one search criterion is required.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/coordinates"
}

\examples{
\donttest{
rba_uniprot_coordinates_search(taxid = 9606, chromosome = "y")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item McGarvey, P. B., Nightingale, A., Luo, J., Huang, H., Martin,
  M. J., Wu, C., & The UniProt Consortium. (2019). UniProt genomic
  mapping for deciphering functional effects of missense variants.
  Human Mutation, 40(6), 694–705.
  https://doi.org/10.1002/humu.23738
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Coordinates":
\code{\link[=rba_uniprot_coordinates]{rba_uniprot_coordinates()}},
\code{\link[=rba_uniprot_coordinates_location]{rba_uniprot_coordinates_location()}},
\code{\link[=rba_uniprot_coordinates_location_genome]{rba_uniprot_coordinates_location_genome()}},
\code{\link[=rba_uniprot_coordinates_location_protein]{rba_uniprot_coordinates_location_protein()}}
}
\concept{"UniProt - Coordinates"}
