% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_coordinates.R
\name{rba_uniprot_coordinates_search}
\alias{rba_uniprot_coordinates_search}
\title{Search genomic coordinates for UniProt entries}
\usage{
rba_uniprot_coordinates_search(
  accession = NA,
  chromosome = NA,
  ensembl = NA,
  gene = NA,
  protein = NA,
  taxid = NA,
  location = NA,
  ...
)
}
\arguments{
\item{location}{}
}
\value{

}
\description{
Search genomic coordinates for UniProt entries
}
