% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_epitope_search}
\alias{rba_uniprot_epitope_search}
\title{Search UniProt Epitopes}
\usage{
rba_uniprot_epitope_search(
  accession = NULL,
  epitope_sequence = NULL,
  iedb_id = NULL,
  match_score = NULL,
  ...
)
}
\arguments{
\item{accession}{Character: (optional)
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{epitope_sequence}{Character: (optional) A single epitope protein
sequence.}

\item{iedb_id}{Character or Numeric: (optional)
\href{https://www.iedb.org/}{IEDB} epitope identifier(s).
You can supply up to 20 identifiers.}

\item{match_score}{Numeric: (optional) A whole number from 0 to 100 giving
the minimum
alignment score between the epitope sequence and target protein sequence.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list in which each element represents a matching UniProt entry and
  is named by accession when available.
}
\description{
Use this function to search epitope data associated with UniProt entries,
  using various criteria such as UniProt accession, epitope sequence,
  IEDB ID, and match score. At least one search criterion is required.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/epitope"
}

\examples{
\donttest{
  rba_uniprot_epitope_search(accession = c("Q84ZX5", "P36222"))
}
\donttest{
  rba_uniprot_epitope_search(epitope_sequence = "DKKCIEWEKAQHGA")
}
\donttest{
  rba_uniprot_epitope_search(iedb_id = 20354)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Epitopes":
\code{\link[=rba_uniprot_epitope]{rba_uniprot_epitope()}}
}
\concept{"UniProt - Epitopes"}
