% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_features}
\alias{rba_uniprot_features}
\title{Get UniProt protein sequence features by accession}
\usage{
rba_uniprot_features(
  accession,
  types = NA,
  categories = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get UniProt protein sequence features by accession
}
