% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_features_search}
\alias{rba_uniprot_features_search}
\title{Search protein sequence features in UniProt}
\usage{
rba_uniprot_features_search(
  accession = NA,
  gene = NA,
  exact_gene = NA,
  protein = NA,
  reviewed = NA,
  organism = NA,
  taxid = NA,
  categories = NA,
  types = NA,
  ...
)
}
\arguments{
\item{types}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search protein sequence features in UniProt
}
