% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_features_search}
\alias{rba_uniprot_features_search}
\title{Search protein sequence features in UniProt}
\usage{
rba_uniprot_features_search(
  accession = NA,
  gene = NA,
  exact_gene = NA,
  protein = NA,
  reviewed = NA,
  organism = NA,
  taxid = NA,
  categories = NA,
  types = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Search protein sequence features in UniProt
}
