% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_features_type}
\alias{rba_uniprot_features_type}
\title{Search protein sequence features of a given type in UniProt
Search for term(s) that appear in feature description for your specified
feature type. For example, you can search by type=DOMAIN and
Term=Kinase. Comma separated values}
\usage{
rba_uniprot_features_type(terms, type, categories = NA, ...)
}
\arguments{
\item{categories}{}
}
\value{

}
\description{
Search protein sequence features of a given type in UniProt
Search for term(s) that appear in feature description for your specified
feature type. For example, you can search by type=DOMAIN and
Term=Kinase. Comma separated values
}
