% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_genecentric}
\alias{rba_uniprot_genecentric}
\title{Get gene centric proteins by Uniprot accession}
\usage{
rba_uniprot_genecentric(accession, ...)
}
\arguments{
\item{accession}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get gene centric proteins by Uniprot accession
}
