% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_genecentric}
\alias{rba_uniprot_genecentric}
\title{Get gene centric proteins by Uniprot accession}
\usage{
rba_uniprot_genecentric(accession, ...)
}
\arguments{
\item{accession}{}
}
\value{

}
\description{
Get gene centric proteins by Uniprot accession
}
