% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_genecentric_search}
\alias{rba_uniprot_genecentric_search}
\title{Search gene centric proteins}
\usage{
rba_uniprot_genecentric_search(upid = NA, accession = NA, gene = NA, ...)
}
\arguments{
\item{gene}{}
}
\value{

}
\description{
Search gene centric proteins
}
