% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_mutagenesis}
\alias{rba_uniprot_mutagenesis}
\title{Get Mutagenesis by UniProt Accession}
\usage{
rba_uniprot_mutagenesis(accession, location = NULL, ...)
}
\arguments{
\item{accession}{Character:
\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}.}

\item{location}{Character: (optional) A valid amino acid range (e.g. 10-25)
within the sequence
of the given protein.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list containing the mutagenesis description of your supplied
  UniProt protein's sequence.
}
\description{
UniProt describes how sequence mutations affect the biological properties
  of a protein, cell, or organism. Retrieve the
  \href{https://www.uniprot.org/help/mutagen}{
  mutagenesis annotations} mapped to one UniProt protein.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/mutagenesis/\{accession\}"
}

\examples{
\donttest{
rba_uniprot_mutagenesis(accession = "P0DTC2", location = "300-400")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Mutagenesis":
\code{\link[=rba_uniprot_mutagenesis_search]{rba_uniprot_mutagenesis_search()}}
}
\concept{"UniProt - Mutagenesis"}
