% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteins}
\alias{rba_uniprot_proteins}
\title{Get UniProt entry by accession}
\usage{
rba_uniprot_proteins(
  accession = NA,
  interaction = FALSE,
  isoforms = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get UniProt entry by accession
}
