% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteins_crossref}
\alias{rba_uniprot_proteins_crossref}
\title{Get UniProt Entry by UniProt Cross-Reference Database and ID}
\usage{
rba_uniprot_proteins_crossref(
  db_id,
  db_name,
  reviewed = NULL,
  isoform = NULL,
  ...
)
}
\arguments{
\item{db_id}{Character: Protein identifier in the cross-reference database.}

\item{db_name}{Character:
\href{https://www.uniprot.org/database/}{Cross-reference database name}.}

\item{reviewed}{Logical: (optional) If \code{TRUE}, return only reviewed
Swiss-Prot entries. If \code{FALSE}, return only unreviewed TrEMBL entries.}

\item{isoform}{Numeric: (optional) One of:\itemize{
\item 0: Exclude isoforms.
\item 1: Return isoforms only.}
See \href{https://www.uniprot.org/help/alternative_products}{alternative
products}.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list named by UniProt accession. Each element is a UniProtKB entry
  corresponding to the supplied cross-reference identifier.
}
\description{
UniProt cross-references connect protein entries with identifiers in
  \href{https://www.uniprot.org/database/}{external databases}. Retrieve
  UniProtKB entries associated with an identifier from one of these
  databases.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/proteins/\{dbtype\}:\{dbid\}"
}

\examples{
\donttest{
rba_uniprot_proteins_crossref("cd40", "hgnc")
}
\donttest{
rba_uniprot_proteins_crossref("cd40", "hgnc", reviewed = TRUE)
}
\donttest{
rba_uniprot_proteins_crossref("mica", "hgnc", isoform = 0)
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Proteins":
\code{\link[=rba_uniprot_proteins]{rba_uniprot_proteins()}},
\code{\link[=rba_uniprot_proteins_search]{rba_uniprot_proteins_search()}}
}
\concept{"UniProt - Proteins"}
