% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteins_crossref}
\alias{rba_uniprot_proteins_crossref}
\title{Get UniProt entries by UniProt cross reference and its ID}
\usage{
rba_uniprot_proteins_crossref(db_type, db_id, reviewed = NA, isoform = NA, ...)
}
\arguments{
\item{isoform}{}
}
\value{

}
\description{
Get UniProt entries by UniProt cross reference and its ID
}
