% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteins_search}
\alias{rba_uniprot_proteins_search}
\title{Search UniProt entries}
\usage{
rba_uniprot_proteins_search(
  accession = NA,
  reviewed = NA,
  isoform = NA,
  go_term = NA,
  keyword = NA,
  ec = NA,
  gene = NA,
  exact_gene = NA,
  protein = NA,
  organism = NA,
  taxid = NA,
  pubmed = NA,
  seq_length = NA,
  md5 = NA,
  ...
)
}
\arguments{
\item{md5}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search UniProt entries
}
