% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_proteomes}
\alias{rba_uniprot_proteomes}
\title{Get proteome by proteome/proteins UPID}
\usage{
rba_uniprot_proteomes(upid, get_proteins = FALSE, reviewed = NA, ...)
}
\arguments{
\item{reviewed}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get proteome by proteome/proteins UPID
}
