% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_proteomes_search}
\alias{rba_uniprot_proteomes_search}
\title{Search Proteomes in UniProt}
\usage{
rba_uniprot_proteomes_search(
  name = NULL,
  upid = NULL,
  taxid = NULL,
  keyword = NULL,
  xref = NULL,
  genome_acc = NULL,
  is_ref_proteome = NULL,
  is_redundant = NULL,
  ...
)
}
\arguments{
\item{name}{Character: (optional) A term in the proteome name.}

\item{upid}{Character: (optional)
\href{https://www.uniprot.org/help/proteome_id}{UniProt Proteome
identifier (UPID)}. You can supply up to 100 UPIDs.}

\item{taxid}{Numeric: (optional) NIH-NCBI
\href{https://www.uniprot.org/taxonomy/}{Taxon ID}.
You can supply up to 20 taxon IDs.}

\item{keyword}{Character: (optional) Limit the search to entries containing
the keyword. See \href{https://www.uniprot.org/keywords/}{UniProt Keywords}.}

\item{xref}{Character: (optional) Proteome cross-references such as genome
assembly ID or
Biosample ID. You can supply up to 20 cross-reference IDs.}

\item{genome_acc}{Character: (optional) Genome accession associated with the
proteome's components. You can supply up to 20 accessions.}

\item{is_ref_proteome}{Logical: (optional) If \code{TRUE}, return only
reference proteomes; if \code{FALSE}, return only non-reference proteomes;
if \code{NULL}, do not filter by this criterion. See
\href{https://www.uniprot.org/help/reference_proteome}{'What are reference
proteomes?'} for more information.}

\item{is_redundant}{Logical: (optional) If \code{TRUE}, return only redundant
proteomes; if \code{FALSE}, return only non-redundant proteomes; if
\code{NULL}, do not filter by redundancy. See
\href{https://www.uniprot.org/help/proteome_redundancy}{'Reducing proteome
redundancy'} for more information.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list named by UPID. Each element contains one matching proteome's
  metadata.
}
\description{
UniProt collects and annotates proteomes (protein sets expressed in an
  organism). Search available proteomes by name, identifier, taxonomy,
  keyword, cross-reference, genome accession, or status. See
  \href{https://www.uniprot.org/help/proteome}{What are proteomes?} for
  more information.
}
\details{
At least one search criterion is required.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/proteomes"
}

\examples{
\donttest{
rba_uniprot_proteomes_search(name = "SARS-CoV")
}
\donttest{
rba_uniprot_proteomes_search(genome_acc = "AY274119")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Proteomes":
\code{\link[=rba_uniprot_genecentric]{rba_uniprot_genecentric()}},
\code{\link[=rba_uniprot_genecentric_search]{rba_uniprot_genecentric_search()}},
\code{\link[=rba_uniprot_proteomes]{rba_uniprot_proteomes()}}
}
\concept{"UniProt - Proteomes"}
