% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteomes.R
\name{rba_uniprot_proteomes_search}
\alias{rba_uniprot_proteomes_search}
\title{Search proteomes in UniProt}
\usage{
rba_uniprot_proteomes_search(
  upid = NA,
  name = NA,
  taxid = NA,
  keyword = NA,
  xref = NA,
  genome_acc = NA,
  is_ref_proteome = NA,
  is_redundant = NA,
  ...
)
}
\arguments{
\item{taxid}{}
}
\value{

}
\description{
Search proteomes in UniProt
}
