% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteomics}
\alias{rba_uniprot_proteomics}
\title{Get proteomics peptides mapped to UniProt by accession}
\usage{
rba_uniprot_proteomics(accession, ...)
}
\arguments{
\item{accession}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get proteomics peptides mapped to UniProt by accession
}
