% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteomics}
\alias{rba_uniprot_proteomics}
\title{Get proteomics peptides mapped to UniProt by accession}
\usage{
rba_uniprot_proteomics(
  accession,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get proteomics peptides mapped to UniProt by accession
}
