% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteomics_search}
\alias{rba_uniprot_proteomics_search}
\title{Search proteomics peptides in UniProt}
\usage{
rba_uniprot_proteomics_search(
  accession = NA,
  taxid = NA,
  upid = NA,
  data_source = NA,
  peptide = NA,
  unique = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Search proteomics peptides in UniProt
}
