% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_proteomics_search}
\alias{rba_uniprot_proteomics_search}
\title{Search proteomics peptides in UniProt}
\usage{
rba_uniprot_proteomics_search(
  accession = NA,
  taxid = NA,
  upid = NA,
  data_source = NA,
  peptide = NA,
  unique = NA,
  ...
)
}
\arguments{
\item{unique}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Search proteomics peptides in UniProt
}
