% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{rba_uniprot_ptm_search}
\alias{rba_uniprot_ptm_search}
\title{Search Post-Translational Modification in UniProt (Deprecated)}
\usage{
rba_uniprot_ptm_search(
  accession = NULL,
  ptm = NULL,
  data_source = NULL,
  taxid = NULL,
  upid = NULL,
  peptide = NULL,
  unique = NULL,
  ...
)
}
\arguments{
\item{accession}{\href{https://www.uniprot.org/help/accession_numbers}{
UniProtKB primary or secondary accession}(s). You can supply up to 100
accession numbers.}

\item{ptm}{Post-translational modification name}

\item{data_source}{Proteomics data source. You can choose up to two of:
\itemize{
\item \href{https://www.uniprot.org/database/DB-0186}{"MaxQB"}
\item \href{https://www.uniprot.org/database/DB-0071}{"PeptideAtlas"}
\item \href{https://www.uniprot.org/database/DB-0205}{"EPD"}
\item \href{https://www.uniprot.org/database/DB-0229}{"ProteomicsDB"}
}}

\item{taxid}{NIH-NCBI \href{https://www.uniprot.org/taxonomy/}{Taxon ID}.
You can supply up to 20 taxon IDs.}

\item{upid}{\href{https://www.uniprot.org/help/proteome_id}{UniProt Proteome
identifier (UPID)}. You can supply up to 100 UPIDs.}

\item{peptide}{Peptide sequence(s). You can supply up to 20 sequences.}

\item{unique}{Logical: Should the results be filtered based on the
Peptide's uniqueness (the fact that a peptide maps to only 1 protein). If
TRUE, Only unique peptides will be returned, if FALSE only un-unique
peptides will be returned; If NULL (default) the results will not be
filtered based on this.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list Where each element correspond to a UniProt protein and
  post-translational modification are organized under the "features"
  sub-list.
}
\description{
This function is Deprecated. Please use
  \code{\link{rba_uniprot_proteomics_ptm_search}} instead.\cr
UniProt maps proteomics peptides from different sources to the proteins'
  sequences. Using this function, you can search for  proteomics
  peptides that has been map to UniProt proteins. You may also refine your
  search with modifiers such as data_source, peptide etc. See
  "Arguments section" for more information.
}
\details{
Note that this is a search function. Thus, you are not required to fill
  every argument; You may use whatever combinations of arguments you see
  fit for your query.
  \cr see also:
  \href{https://www.uniprot.org/help/ptm_processing_section}{PTM /
  Processing section in UniProtKB}
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/proteomics-ptm"
}

\examples{
\dontrun{
#Deprecated
rba_uniprot_ptm_search(peptide = "NDQVYQPLRDRDDAQYSHLGGNWAR")
}

}
\references{
\itemize{
  \item The UniProt Consortium, UniProt: the universal protein
  knowledgebase in 2021, Nucleic Acids Research, Volume 49, Issue D1,
  8 January 2021, Pages D480–D489, https://doi.org/10.1093/nar/gkaa1100
  \item Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas
  Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd
  Turner, Maria Martin, The Proteins API: accessing key integrated protein
  and genome information, Nucleic Acids Research, Volume 45, Issue W1,
  3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "Deprecated functions":
\code{\link[=rba_uniprot_proteomics]{rba_uniprot_proteomics()}},
\code{\link[=rba_uniprot_proteomics_search]{rba_uniprot_proteomics_search()}},
\code{\link[=rba_uniprot_ptm]{rba_uniprot_ptm()}}
}
\concept{"Deprecated functions"}
