% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy}
\alias{rba_uniprot_taxonomy}
\title{This service returns a list of children/siblings/parent nodes that belongs
to a taxonomy node with links to its parent, sibling and children nodes.}
\usage{
rba_uniprot_taxonomy(
  ids,
  hierarchy = NA,
  node = FALSE,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
This service returns a list of children/siblings/parent nodes that belongs
to a taxonomy node with links to its parent, sibling and children nodes.
}
