% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy_lineage}
\alias{rba_uniprot_taxonomy_lineage}
\title{Get Taxonomic Lineage}
\usage{
rba_uniprot_taxonomy_lineage(id, ...)
}
\arguments{
\item{id}{(numeric) a
\href{https://www.uniprot.org/help/taxonomic_identifier}{NCBI taxonomic
  identifier}}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
A list with a data frame containing All the nodes that preceded your
  supplied node in the taxonomic tree. with your node as the first row
  and the root node in the last row.
}
\description{
Use this function to retrieve the taxonomic lineage of your supplied
  taxonomy node.
}
\section{Corresponding API Resources}{

 "GET https://ebi.ac.uk/proteins/api/lineage/\{id\}"
}

\examples{
\donttest{
rba_uniprot_taxonomy_lineage(id = 9989)
}

}
\references{
\itemize{
  \item The UniProt Consortium , UniProt: the Universal Protein
  Knowledgebase in 2025, Nucleic Acids Research, 2024;, gkae1010,
  https://doi.org/10.1093/nar/gkae1010
  \item Andrew Nightingale, Ricardo Antunes, Emanuele Alpi, Borisas
  Bursteinas, Leonardo Gonzales, Wudong Liu, Jie Luo, Guoying Qi, Edd
  Turner, Maria Martin, The Proteins API: accessing key integrated protein
  and genome information, Nucleic Acids Research, Volume 45, Issue W1,
  3 July 2017, Pages W539–W544, https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Taxonomy":
\code{\link[=rba_uniprot_taxonomy]{rba_uniprot_taxonomy()}},
\code{\link[=rba_uniprot_taxonomy_lca]{rba_uniprot_taxonomy_lca()}},
\code{\link[=rba_uniprot_taxonomy_name]{rba_uniprot_taxonomy_name()}},
\code{\link[=rba_uniprot_taxonomy_path]{rba_uniprot_taxonomy_path()}},
\code{\link[=rba_uniprot_taxonomy_relationship]{rba_uniprot_taxonomy_relationship()}}
}
\concept{"UniProt - Taxonomy"}
