% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy_path}
\alias{rba_uniprot_taxonomy_path}
\title{get /path This service returns all taxonomic nodes that have a relationship
with the queried taxonomy ID in a specific direction (TOP or BOTTOM) and
depth level.}
\usage{
rba_uniprot_taxonomy_path(
  id,
  direction,
  depth = 5,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
get /path This service returns all taxonomic nodes that have a relationship
with the queried taxonomy ID in a specific direction (TOP or BOTTOM) and
depth level.
}
