% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_taxonomy.R
\name{rba_uniprot_taxonomy_relationship}
\alias{rba_uniprot_taxonomy_relationship}
\title{This service returns the shortest path between two taxonomy nodes showing
their relationship}
\usage{
rba_uniprot_taxonomy_relationship(
  from,
  to,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
This service returns the shortest path between two taxonomy nodes showing
their relationship
}
