% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc}
\alias{rba_uniprot_uniparc}
\title{Get UniParc entry only by UniProt accession}
\usage{
rba_uniprot_uniparc(
  accession = NA,
  rf_dd_type = NA,
  rf_db_id = NA,
  rf_active = NA,
  rf_tax_id = NA,
  ...
)
}
\arguments{
\item{rf_tax_id}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get UniParc entry only by UniProt accession
}
