% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc_bestguess}
\alias{rba_uniprot_uniparc_bestguess}
\title{For a given user input (request parameters), Best Guess returns the
UniParcEntry with a cross-reference to the longest active UniProtKB
sequence (preferably from Swiss-Prot and if not then TrEMBL). It also
returns the sequence and related information. If it finds more than one
longest active UniProtKB sequence it returns 400 (Bad Request) error
response with the list of cross references found.}
\usage{
rba_uniprot_uniparc_bestguess(
  upi = NA,
  accession = NA,
  db_id = NA,
  gene = NA,
  taxid = NA,
  ...
)
}
\arguments{
\item{taxid}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
For a given user input (request parameters), Best Guess returns the
UniParcEntry with a cross-reference to the longest active UniProtKB
sequence (preferably from Swiss-Prot and if not then TrEMBL). It also
returns the sequence and related information. If it finds more than one
longest active UniProtKB sequence it returns 400 (Bad Request) error
response with the list of cross references found.
}
