% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc_get}
\alias{rba_uniprot_uniparc_get}
\title{Get UniParc entries by all UniParc cross reference accessions}
\usage{
rba_uniprot_uniparc_get(
  db_id = NA,
  upid = NA,
  upi = NA,
  rf_dd_type = NA,
  rf_db_id = NA,
  rf_active = NA,
  rf_tax_id = NA,
  ...
)
}
\arguments{
\item{rf_tax_id}{}

\item{...}{rbioapi option(s). Refer to \code{\link{rba_options}}'s arguments documentation for more information on available options.}
}
\value{

}
\description{
Get UniParc entries by all UniParc cross reference accessions
}
