% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc_search}
\alias{rba_uniprot_uniparc_search}
\title{Search UniParc entries}
\usage{
rba_uniprot_uniparc_search(
  upi = NA,
  accession = NA,
  db_type = NA,
  db_id = NA,
  gene = NA,
  protein = NA,
  taxid = NA,
  organism = NA,
  sequence_checksum = NA,
  ipr = NA,
  signature_type = NA,
  signature_id = NA,
  upid = NA,
  seq_length = NA,
  rf_dd_type = NA,
  rf_db_id = NA,
  rf_active = NA,
  rf_tax_id = NA,
  location = NA,
  ...
)
}
\arguments{
\item{location}{}
}
\value{

}
\description{
Search UniParc entries
}
