% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc_sequence}
\alias{rba_uniprot_uniparc_sequence}
\title{Get UniParc Entries by Sequence}
\usage{
rba_uniprot_uniparc_sequence(
  sequence,
  rf_dd_type = NULL,
  rf_db_id = NULL,
  rf_active = NULL,
  rf_tax_id = NULL,
  ...
)
}
\arguments{
\item{sequence}{Character: Exact protein sequence. Partial matches are not
accepted.}

\item{rf_dd_type}{Character: (optional) Filter the UniParc entry's content by
\href{https://www.uniprot.org/database/}{cross-reference} names. You can
supply multiple values.}

\item{rf_db_id}{Character: (optional) Filter the UniParc entry's content by
protein
identifiers in any cross-reference database. You can supply multiple
values.}

\item{rf_active}{Logical: (optional) Filter the UniParc entry's content by
active status in the source database: \code{TRUE} retains active database
references, \code{FALSE} retains inactive references, and \code{NULL}
applies no active-status filter.}

\item{rf_tax_id}{Numeric: (optional) Filter the UniParc entry's content by
NIH-NCBI \href{https://www.uniprot.org/taxonomy/}{Taxon ID}. You can
supply multiple values.}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
The matching UniParc entry.
}
\description{
Retrieve a UniParc entry using an exact protein sequence. Partial matches
  are not accepted. The \code{rf_*} arguments filter cross-references within
  the returned entry.
}
\section{Corresponding API Resources}{

 "POST https://www.ebi.ac.uk/proteins/api/uniparc/sequence"
}

\examples{
\donttest{
rba_uniprot_uniparc_sequence("GMRSCPRGCSQRGRCENGRCVCNPGYTGEDC")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - UniParc":
\code{\link[=rba_uniprot_uniparc]{rba_uniprot_uniparc()}},
\code{\link[=rba_uniprot_uniparc_bestguess]{rba_uniprot_uniparc_bestguess()}},
\code{\link[=rba_uniprot_uniparc_search]{rba_uniprot_uniparc_search()}}
}
\concept{"UniProt - UniParc"}
