% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_uniparc.R
\name{rba_uniprot_uniparc_sequence}
\alias{rba_uniprot_uniparc_sequence}
\title{Get UniParc entries by sequence}
\usage{
rba_uniprot_uniparc_sequence(
  sequence,
  rf_dd_type = NA,
  rf_db_id = NA,
  rf_active = NA,
  rf_tax_id = NA,
  ...
)
}
\arguments{
\item{rf_tax_id}{}
}
\value{

}
\description{
Get UniParc entries by sequence
}
