% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_variation}
\alias{rba_uniprot_variation}
\title{Retrieve UniProt Natural Variants by Identifier}
\usage{
rba_uniprot_variation(
  id,
  id_type,
  source_type = NULL,
  consequence_type = NULL,
  wild_type = NULL,
  alternative_sequence = NULL,
  location = NULL,
  save_peff = FALSE,
  ...
)
}
\arguments{
\item{id}{Character: A single identifier: either a
\href{https://www.uniprot.org/help/accession_numbers}{UniProt primary or
secondary accession}, an \href{https://www.ncbi.nlm.nih.gov/snp/}{NIH-NCBI
dbSNP ID}, or an \href{https://varnomen.hgvs.org/}{HGVS expression}.}

\item{id_type}{Character: The type of supplied ID argument, one of:
\href{https://www.uniprot.org/help/accession_numbers}{"uniprot"},
\href{https://www.ncbi.nlm.nih.gov/snp/}{"dbsnp"} or
\href{https://varnomen.hgvs.org/}{"hgvs"}.}

\item{source_type}{Character: (optional) Variation's source type. You
can choose up to two of:
"uniprot", "large scale study", "mixed", "clinvar", "nci-tcga",
"cosmic curated", "ensembl", "gnomad", "topmed", or "exac".}

\item{consequence_type}{Character: (optional) Variation's consequence
type. You can choose up to
two of: "missense", "stop gained" or "stop lost".}

\item{wild_type}{Character: (optional) Wild-type amino acid. Accepted
values are IUPAC
single-letter amino acid codes and "*" for a stop codon. You can supply up
to 20 values.}

\item{alternative_sequence}{Character: (optional) Alternative amino
acid. Accepted values are
IUPAC single-letter amino acid codes, "*" for a stop codon, and "-" for a
deletion. You can supply up to 20 values.}

\item{location}{Character: (optional) A valid amino acid range (e.g. 10-25)
within the sequence
where the variation occurs.}

\item{save_peff}{Logical or Character: (default = \code{FALSE}) \itemize{
\item FALSE: Return the parsed JSON response.
\item TRUE: Save as PEFF file to an automatically-generated path.
\item Character string: A valid file path to save the PEFF file.}}

\item{...}{rbioapi option(s). See \code{\link{rba_options}}'s
arguments manual for more information on available options.}
}
\value{
If \code{save_peff = FALSE}, a list. For \code{id_type = "uniprot"},
  it represents the requested entry; for \code{"dbsnp"} or \code{"hgvs"},
  each element represents a matching entry and is named by accession when
  available. If PEFF output is requested, the response is written to disk
  and returned as a character string.
}
\description{
Retrieve natural variant annotations by UniProt accession, dbSNP identifier,
  or HGVS expression.
}
\section{Corresponding API Resources}{

 "GET https://www.ebi.ac.uk/proteins/api/variation/dbsnp/\{dbid\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/variation/hgvs/\{hgvs\}"
 \cr "GET https://www.ebi.ac.uk/proteins/api/variation/\{accession\}"
}

\examples{
\donttest{
rba_uniprot_variation(id = "rs121434451", id_type = "dbsnp")
}
\donttest{
rba_uniprot_variation(id = "NC_000008.11:g.22119227C>T", id_type = "hgvs")
}
\donttest{
rba_uniprot_variation(id = "O43593", id_type = "uniprot")
}

}
\references{
\itemize{
  \item The UniProt Consortium. (2025). UniProt: the Universal Protein
  Knowledgebase in 2025. Nucleic Acids Research, 53(D1), D609–D617.
  https://doi.org/10.1093/nar/gkae1010
  \item Nightingale, A., Antunes, R., Alpi, E., Bursteinas, B., Gonzales,
  L., Liu, W., Luo, J., Qi, G., Turner, E., & Martin, M. (2017). The
  Proteins API: Accessing key integrated protein and genome information.
  Nucleic Acids Research, 45(W1), W539–W544.
  https://doi.org/10.1093/nar/gkx237
  \item \href{https://www.ebi.ac.uk/proteins/api/doc/}{Proteins API
  Documentation}
  \item \href{https://www.uniprot.org/help/publications}{Citations note
  on UniProt website}
  }
}
\seealso{
Other "UniProt - Variation":
\code{\link[=rba_uniprot_variation_search]{rba_uniprot_variation_search()}}
}
\concept{"UniProt - Variation"}
