% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_variation_accession}
\alias{rba_uniprot_variation_accession}
\title{Get natural variants by UniProt accession}
\usage{
rba_uniprot_variation_accession(
  accession,
  source_type = NA,
  consequence_type = NA,
  wild_type = NA,
  alternative_sequence = NA,
  location = NA,
  save_peff = FALSE,
  save_to = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get natural variants by UniProt accession
}
