% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_variation_dbsnp}
\alias{rba_uniprot_variation_dbsnp}
\title{Get natural variants in UniProt by NIH-NCBI SNP database identifier}
\usage{
rba_uniprot_variation_dbsnp(
  db_id,
  source_type = NA,
  consequence_type = NA,
  wild_type = NA,
  alternative_sequence = NA,
  location = NA,
  save_peff = FALSE,
  save_to = NA,
  verbose = TRUE,
  progress_bar = FALSE,
  diagnostics = FALSE
)
}
\arguments{
\item{diagnostics}{}
}
\value{

}
\description{
Get natural variants in UniProt by NIH-NCBI SNP database identifier
}
