% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_proteins.R
\name{rba_uniprot_variation_search}
\alias{rba_uniprot_variation_search}
\title{Search natural variants in UniProt
Implementation Notes Among the available response content types, PEFF
format (PSI Extended FASTA Format from the Human Proteome Organisation -
Proteomics Standards Initiative, HUPO-PSI) is provided with only variants
reported in the output.}
\usage{
rba_uniprot_variation_search(
  accession = NA,
  save_peff = FALSE,
  source_type = NA,
  consequence_type = NA,
  wild_type = NA,
  alternative_sequence = NA,
  location = NA,
  disease = NA,
  omim = NA,
  evidence = NA,
  taxid = NA,
  db_type = NA,
  db_id = NA,
  ...
)
}
\arguments{
\item{save_peff}{}
}
\value{

}
\description{
Search natural variants in UniProt
Implementation Notes Among the available response content types, PEFF
format (PSI Extended FASTA Format from the Human Proteome Organisation -
Proteomics Standards Initiative, HUPO-PSI) is provided with only variants
reported in the output.
}
