% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rbioapi}
\alias{rbioapi}
\alias{rbioapi-package}
\title{rbioapi: User-friendly R Interface to Biologic Web-Services' API}
\description{
The goal of rbioapi is to provide a user-friendly and unified
  interface to biological databases, to insulate the user from
  technicalities when it comes to using API services and to create a
  unified, easy-to-learn, and easy-to-implement way to connect to biologic
  databases.\cr
  With rbioapi, You are not required to have any technical knowledge. Just
  fill a function's arguments and the rest is handled for you.\cr
  This an ongoing project. New databases and services will be implemented
  periodically. Feel free to suggest any databases or services you often use.
}
\section{Currently rbioapi fully supports}{

\enumerate{\item Enrichr \item MiEAA \item PANTHER \item Reactome
  \item STRING \item UniProt}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moosa-r/rbioapi}
  \item Report bugs at \url{https://github.com/moosa-r/rbioapi/issues}
}

}
\author{
\strong{Maintainer}: Moosa Rezwani \email{moosa.rezwani@gmail.com} [copyright holder]

}
\keyword{internal}
