% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rbioapi}
\alias{rbioapi}
\title{rbioapi: User-friendly interface to Biological Databases' APIs}
\description{
This package provides and easy, user-friendly and unified access to Biological databases' API.
STRING, Reactome, Uniprot, Pubmed
}
\section{STRING functions}{

map, interactions, image
}

