% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rbioapi}
\alias{rbioapi-package}
\alias{rbioapi}
\title{rbioapi: User-Friendly R Interface to Biologic Web Services' API}
\description{
Currently fully supports Enrichr, JASPAR, miEAA, PANTHER, Reactome, STRING, and UniProt! The goal of rbioapi is to provide a user-friendly and consistent interface to biological databases and services. In a way that insulates the user from the technicalities of using web services API and creates a unified and easy-to-use interface to biological and medical web services. This is an ongoing project; New databases and services will be added periodically. Feel free to suggest any databases or services you often use.
}
\section{Supported Database/Services}{

\itemize{
  \item Enrichr
  \item JASPAR
  \item MiEAA
  \item PANTHER
  \item Reactome
  \item STRING
  \item UniProt
  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rbioapi.moosa-r.com}
  \item \url{https://github.com/moosa-r/rbioapi}
  \item Report bugs at \url{https://github.com/moosa-r/rbioapi/issues}
}

}
\author{
\strong{Maintainer}: Moosa Rezwani \email{moosa.rezwani@gmail.com} (\href{https://orcid.org/0000-0001-6325-4444}{ORCID})

}
\keyword{internal}
