% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rbioapi}
\alias{rbioapi}
\alias{rbioapi-package}
\title{rbioapi: User-Friendly R Interface to Biologic Web Services' API}
\description{
Currently fully supports miEAA, PANTHER, Reactome, STRING,
  and UniProt! The goal of rbioapi is to provide a user-friendly and
  consistent interface to biological databases and services; It is
  designed in a way that insulates the user from technicalities when
  it comes to using API services and creates a unified and
  easy-to-implement tool to connect to biological databases and services.
  With rbioapi, You are not required to have any prior technical
  knowledge. Just fill in a function's arguments and the rest is handled
  for you. This an ongoing project. New databases and services will be
  implemented periodically to gradually make rbioapi more comprehensive.
  Feel free to suggest any databases or services you often use.
}
\section{Currently, rbioapi fully supports and Covers API resources of}{

\itemize{
  \item MiEAA
  \item PANTHER
  \item Reactome
  \item STRING
  \item UniProt}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moosa-r/rbioapi}
  \item Report bugs at \url{https://github.com/moosa-r/rbioapi/issues}
}

}
\author{
\strong{Maintainer}: Moosa Rezwani \email{moosa.rezwani@gmail.com} [copyright holder]

}
\keyword{internal}
